### 2025-11-21 #9492 @editorialbot
[PRE REVIEW]: ROMS-Tools: A Python Package for Preparing and Analyzing ROMS Simulations
**Submitting author:** @NoraLoose (<a href="http://orcid.org/0000-0002-3684-9634">Nora Loose</a>)
**Repository:** https://github.com/CWorthy-ocean/roms-tools
**Branch with paper.md** (empty if default branch):
**Version:** v3.2.0
