### 2025-11-21 100d1d6
Check xesmf installation in CI
