### 2025-11-21 470f756
Separate updating mask and topography
