### 2025-11-21 197be88
Clean up ChildGrid class
