+++ roms_tools/setup/nesting.py
@@ -1,4 +1,5 @@
 import logging
+from collections.abc import Callable
 from dataclasses import dataclass, field
 from pathlib import Path
 from typing import Any
@@ -80,40 +81,6 @@ class ChildGrid(Grid):
     """An xarray Dataset containing boundary mappings, where child grid boundaries are
     mapped onto parent grid indices."""
 
-    def __post_init__(self):
-        self._input_checks()
-
-        # Horizontal grid
-        self._create_horizontal_grid()
-
-        # Check if the Greenwich meridian goes through the domain.
-        self._straddle()
-
-        # Mask
-        self._create_mask(verbose=self.verbose)
-
-        # Coarsen the dataset if needed
-        self._coarsen()
-
-        # Map child boundaries onto parent grid
-        self._map_child_boundaries_onto_parent_grid_indices(verbose=self.verbose)
-
-        # Topography
-        self.update_topography(
-            topography_source=self.topography_source,
-            hmin=self.hmin,
-            verbose=self.verbose,
-        )
-
-        # Vertical coordinate system
-        self.update_vertical_coordinate(
-            N=self.N,
-            theta_s=self.theta_s,
-            theta_b=self.theta_b,
-            hc=self.hc,
-            verbose=self.verbose,
-        )
-
     def _map_child_boundaries_onto_parent_grid_indices(self, verbose: bool = False):
         """Maps child grid boundary points onto absolute indices of the parent grid."""
         with Timed(
@@ -134,66 +101,124 @@ class ChildGrid(Grid):
 
             self.ds_nesting = ds_nesting
 
-    def _modify_child_topography_and_mask(self, verbose: bool = False):
-        """Adjust the topography and mask of the child grid to align with the parent grid.
-
-        Uses a weighted sum based on boundary distance and clips depth values to a
-        minimum.
-        """
-        with Timed("=== Modifying child topography and mask ===", verbose=verbose):
-            # Prepare parent and child grid datasets by adjusting longitudes for dateline crossing
+    def _apply_child_modification(
+        self,
+        modifier: Callable,
+        modifier_name: str,
+        verbose: bool = False,
+    ):
+        """Shared logic for modifying child mask/topography."""
+        with Timed(f"=== Modifying child {modifier_name} ===", verbose=verbose):
+            # Prepare datasets (fix dateline)
             parent_grid_ds, child_grid_ds = self._prepare_grid_datasets()
-            child_grid_ds = modify_child_topography_and_mask(
-                parent_grid_ds, child_grid_ds, self.boundaries, self.hmin
-            )
 
-            # Finalize grid datasets by adjusting longitudes back to [0, 360] range
-            parent_grid_ds, child_grid_ds = self._finalize_grid_datasets(
+            # Apply modification function
+            child_grid_ds = modifier(parent_grid_ds, child_grid_ds)
+
+            # Restore longitudes to 0-360
+            _, child_grid_ds = self._finalize_grid_datasets(
                 parent_grid_ds, child_grid_ds
             )
 
             self.ds = child_grid_ds
 
-    def update_topography(
-        self, topography_source=None, hmin=None, verbose=False
-    ) -> None:
-        """Update the child grid topography via the following steps.
+    def _modify_child_mask(self, verbose: bool = False) -> None:
+        """Adjust child grid mask to align with the parent grid."""
+        self._apply_child_modification(
+            modifier=lambda p, c: modify_child_mask(p, c, self.boundaries),
+            modifier_name="topography and mask",
+            verbose=verbose,
+        )
 
-        - Regrids the topography based on the specified source.
-        - Applies global and local smoothing.
-        - Adjusts the child grid topography and mask to match the parent grid.
[…]
