### 2025-11-21 7a7128a
Test plotting wide grid
