### 2025-11-21 8a9f5fa
Plot wide grids with different cartopy projection (#496)
* Test plotting wide grid

* Plot wide grid with different projection

* Update plots in notebook

* Add to release notes
