### 2025-11-21 b91d2dd
More tweaking of log message
