### 2025-11-24 23b251e
Option to not apply the mask in plots
