### 2025-11-24 1180310
Add to release notes
