### 2025-11-24 2d7119e
Correct typo in release notes
