### 2025-11-24 b341449
Start drafting fixing the NaNs
