### 2025-11-24 81fe7d6
Start drafting fixing the NaNs
