### 2025-11-25 281ee26
Don't mask h for local smoothing
