### 2025-12-01 67857db
Make plotting safe
