### 2025-12-02 ebc128c
Initial refactor
