### 2025-12-02 a1a3170
Initial refactor
