### 2025-12-02 cc97888
Make unit handling in `plot_locations()` robust for grids without `mask_rho` units (#499)
* Make plotting safe

* Add to release notes
