### 2025-12-02 f295bca
Finish refactor
