+++ roms_tools/setup/utils.py
@@ -4,10 +4,10 @@ import time
 import typing
 from collections.abc import Sequence
 from dataclasses import asdict, fields, is_dataclass
-from datetime import datetime
+from datetime import datetime, timedelta
 from enum import StrEnum
 from pathlib import Path
-from typing import Any, Literal
+from typing import Any, Literal, TypeAlias
 
 import cftime
 import numba as nb
