+++ roms_tools/tests/test_setup/test_initial_conditions.py
@@ -11,7 +11,7 @@ import xarray as xr
 from conftest import calculate_data_hash
 from roms_tools import Grid, InitialConditions
 from roms_tools.download import download_test_data
-from roms_tools.setup.datasets import (
+from roms_tools.setup.lat_lon_datasets import (
     CESMBGCDataset,
     UnifiedBGCDataset,
 )
