+++ roms_tools/tests/test_setup/test_surface_forcing.py
@@ -10,7 +10,7 @@ import xarray as xr
 from conftest import calculate_data_hash
 from roms_tools import Grid, SurfaceForcing
 from roms_tools.download import download_test_data
-from roms_tools.setup.datasets import RawDataSource
+from roms_tools.setup.utils import RawDataSource
 
 
 @pytest.fixture

[…]
