### 2025-12-02 8f9e3df
Add to release notes
