### 2025-12-03 28c83ea
Finish up NaN fix
