### 2025-12-03 6d8001c
Exclude boundary forcing tests
