+++ pyproject.toml
@@ -43,17 +43,17 @@ dynamic = ["version"]
 
 [project.optional-dependencies]
 dask = [
-    "dask[diagnostics]",
+    "dask[diagnostics,distributed]",
     "zarr"
 ]
 stream = [
-    "dask[diagnostics]",
+    "dask[diagnostics,distributed]",
     "gcsfs",
     "zarr",
     "copernicusmarine",
 ]
 notebooks = [
-    "dask[diagnostics]",
+    "dask[diagnostics,distributed]",
     "gcsfs",
     "zarr",
     "copernicusmarine",
