### 2025-12-04 042fc55
Force Dask to use multi-threading because HDF5 is not process-safe
