+++ pyproject.toml
@@ -43,17 +43,17 @@ dynamic = ["version"]
 
 [project.optional-dependencies]
 dask = [
-    "dask[diagnostics,distributed]",
+    "dask[diagnostics]",
     "zarr"
 ]
 stream = [
-    "dask[diagnostics,distributed]",
+    "dask[diagnostics]",
     "gcsfs",
     "zarr",
     "copernicusmarine",
 ]
 notebooks = [
-    "dask[diagnostics,distributed]",
+    "dask[diagnostics]",
     "gcsfs",
     "zarr",
     "copernicusmarine",
