### 2025-12-04 554dc35
Configure Dask to use threaded scheduler with 2 workers in CI (#501)
* Exclude boundary forcing tests

* Correct path

* Split up tests

* Dask cluster with 2 workers

* Instal dask distributed

* Force Dask to use multi-threading because HDF5 is not process-safe

* Correct typo

* Split tests up more evenly
