### 2025-12-04 570e753
Rename `Dataset` to `LatLonDataset` and modularize `RiverDataset` (#500)
* Initial refactor

* Initial refactor

* Finish refactor

* Add to release notes

* Correct module name in api
