### 2025-12-04 43ec93b
Avoid fill value and improve messages
