### 2025-12-04 22549b8
Adjust error message
