### 2025-12-05 3b84e8a
Allow child domains with land points outside parent (#491)
* Allow child domains with land points outside parent

* Fill NaNs

* plot_nesting only uses xarray

* Adjust warning message

* Finalize plot_nesting

* Remove unneccessary function

* Correct logging statement

* Avoid modifying topo and mask twice

* Add test for child grid that exceeds parent grid

* Adapt test

* Update documentation

* Save copy of grids

* Separate updating mask and topography

* Improve docstrings and typehints

* Clean up ChildGrid class

* Adjust tests

* Correct typo

* More tweaking of log message

* Docume
[…]
