+++ roms_tools/plot.py
@@ -180,25 +180,8 @@ def plot_nesting(parent_grid_ds, child_grid_ds, parent_straddle, with_dim_names=
         with_dim_names=with_dim_names,
     )
 
-    vmax = 3
-    vmin = 0
-    cmap = plt.colormaps.get_cmap("Blues")
-    cmap.set_bad(color="gray")
-    kwargs = {"vmax": vmax, "vmin": vmin, "cmap": cmap}
-
-    field = parent_grid_ds.mask_rho
-    field = field.where(field)
-
-    _add_field_to_ax(
-        ax,
-        parent_lon_deg,
-        parent_lat_deg,
-        field,
-        add_colorbar=False,
-        kwargs=kwargs,
-    )
-
     _add_gridlines(ax)
+    ax.coastlines()
 
     ax.legend(loc="best")
 
@@ -836,6 +819,7 @@ def plot(
     yincrease: bool | None = None,
     use_coarse_grid: bool = False,
     with_dim_names: bool = False,
+    apply_mask: bool = True,
     ax: Axes | None = None,
     save_path: str | None = None,
     cmap_name: str = "YlOrRd",
@@ -908,6 +892,9 @@ def plot(
 
     with_dim_names : bool, optional
         Add grid dimension names (`xi`, `eta`) to the outer plot edges. Only for 2D plots. Default is False.
+
+    apply_mask: bool, optional
+        Whether to apply the land mask to the field. Default is True.
 
     ax : Axes, optional
         Matplotlib axes object. If None, a new figure is created. Default is None.
@@ -991,7 +978,8 @@ def plot(
     field = field.assign_coords({"lon": lon_deg, "lat": lat_deg})
 
     # Mask the field
-    field = field.where(mask)
+    if apply_mask:
+        field = field.where(mask)
 
     # Assign eta and xi as coordinates
     coords_to_assign = {dim: field[dim] for dim in horizontal_dims.values()}

[…]
