### 2025-12-05 3e89677
Prohibit child grids with boundaries on land
