### 2025-12-05 09b43d3
Functions for checking and validating boundaries dict
