### 2025-12-05 7cd707a
New boundaries defaults
