### 2025-12-05 c5e3106
Allow child boundaries on land
