### 2025-12-08 af23646
Only update child boundary indices if boundary not entirely on land (#503)
* Prohibit child grids with boundaries on land

* Correct tests with preaction of wrapping lons

* Add to release notes

* Allow child boundaries on land

* Remove obsolete test

* error-nesting

* Add test
