### 2025-12-08 e6c03ab
Fix release notes
