+++ roms_tools/tests/test_setup/test_nesting.py
@@ -409,70 +409,6 @@ class TestNesting:
                         == "vbar, v, vp"
                     )
 
-    def test_error_if_child_boundary_entirely_on_land(self):
-        parent_grid = Grid(
-            nx=7,
-            ny=7,
-            size_x=240,
-            size_y=240,
-            center_lon=-4.1,
-            center_lat=52.36,
-            rot=0,
-        )
-
-        child_grid_parameters = {
-            "nx": 5,
-            "ny": 5,
-            "size_x": 100,
-            "size_y": 100,
-            "center_lon": -4.1,
-            "center_lat": 52.36,
-        }
-
-        with pytest.raises(
-            ValueError,
-            match="The entire southern boundary of the child grid lies on land.",
-        ):
-            ChildGrid(
-                **child_grid_parameters,
-                parent_grid=parent_grid,
-                boundaries={
-                    "south": True,
-                    "east": True,
-                    "north": True,
-                    "west": True,
-                },
-            )
-
-        with pytest.raises(
-            ValueError,
-            match="The entire eastern boundary of the child grid lies on land.",
-        ):
-            ChildGrid(
-                **child_grid_parameters,
-                parent_grid=parent_grid,
-                boundaries={
-                    "south": False,
-                    "east": True,
-                    "north": True,
-                    "west": True,
-                },
-            )
-
-        child_grid = ChildGrid(
-            **child_grid_parameters,
-            parent_grid=parent_grid,
-            boundaries={
-                "south": False,
-                "east": False,
-                "north": True,
-                "west": True,
-            },
-        )
-
-        assert isinstance(child_grid.ds, xr.Dataset)
-        assert isinstance(child_grid.ds_nesting, xr.Dataset)
-
     @pytest.mark.parametrize(
         "big_grid_fixture, small_grid_fixture",
         [
