### 2025-12-08 1106feb
Add to release notes
