### 2025-12-08 0b303e1
Introduce ROMSDataset class
