+++ roms_tools/roms_dataset.py
@@ -1,0 +1,574 @@
+import logging
+import re
+from dataclasses import dataclass, field
+from datetime import datetime, timedelta
+from pathlib import Path
+from typing import Any
+
+import numpy as np
+import xarray as xr
+
+from roms_tools import Grid
+from roms_tools.setup.fill import LateralFill
+from roms_tools.setup.utils import (
+    one_dim_fill,
+)
+from roms_tools.utils import load_data
+from roms_tools.vertical_coordinate import (
+    compute_depth_coordinates,
+)
+
+DEFAULT_NR_BUFFER_POINTS = (
+    20  # Default number of buffer points for subdomain selection.
+)
+# Balances performance and accuracy:
+# - Too many points → more expensive computations
+# - Too few points → potential boundary artifacts when lateral refill is performed
+# See discussion: https://github.com/CWorthy-ocean/roms-tools/issues/153
+# This default will be applied consistently across all datasets requiring lateral fill.
+
+
+@dataclass(kw_only=True)
+class ROMSDataset:
+    """Represents ROMS model output.
+
+    Parameters
+    ----------
+    grid : Grid
+        Object representing the grid information.
+    path : Union[str, Path, List[Union[str, Path]]]
+        Filename, or list of filenames with model output.
+    model_reference_date : datetime, optional
+        Reference date of ROMS simulation.
+        If not specified, this is inferred from metadata of the model output
+        If specified and does not coincide with metadata, a warning is raised.
+    adjust_depth_for_sea_surface_height : bool, optional
+        Whether to account for sea surface height variations when computing depth coordinates.
+        Defaults to `False`.
+    use_dask: bool, optional
+        Indicates whether to use dask for processing. If True, data is processed with dask; if False, data is processed eagerly. Defaults to False.
+    """
+
+    grid: Grid
+    """Object representing the grid information."""
+    path: str | Path
+    """Filename, or list of filenames with model output."""
+    use_dask: bool = False
+    """Whether to use dask for processing."""
+    model_reference_date: datetime | None = None
+    """Reference date of ROMS simulation."""
+    adjust_depth_for_sea_surface_height: bool | None = False
+    """Whether to account for sea surface height variations when computing depth
+    coordinates."""
+
+    ds: xr.Dataset = field(init=False, repr=False)
+    """An xarray Dataset containing the ROMS output."""
+
+    def __post_init__(self):
+        ds = self._load_model_output()
+        self._infer_model_reference_date_from_metadata(ds)
+        self._check_vertical_coordinate(ds)
+        ds = self._add_absolute_time(ds)
+        ds = self._add_lat_lon_coords_and_masks(ds)
+        self.ds = ds
+
+        # Dataset for depth coordinates
+        self.ds_depth_coords = xr.Dataset()
+
+    def _get_depth_coordinates(self, depth_type="layer", locations=["rho"]):
+        """Ensure depth coordinates are stored for a given location and depth type.
+
+        Calculates vertical depth coordinates (layer or interface) for specified locations (e.g., rho, u, v points)
+        and updates them in the dataset (`self.ds`).
+
+        Parameters
+        ----------
+        depth_type : str
+            The type of depth coordinate to compute. Valid options:
+            - "layer": Compute layer depth coordinates.
+            - "interface": Compute interface depth coordinates.
+        locations : list[str], optional
+            Locations for which to compute depth coordinates. Default is ["rho", "u", "v"].
+            Valid options include:
+            - "rho": Depth coordinates at rho points.
+            - "u": Depth coordinates at u points.
+            - "v": Depth coordinates at v points.
+
+        Updates
+        -------
+        self.ds_depth_coords : xarray.Dataset
+
+        Raises
+        ------
+        ValueError
[…]
