+++ roms_tools/analysis/roms_output.py
@@ -1,32 +1,24 @@
-import logging
-import re
 import warnings
-from dataclasses import dataclass, field
-from datetime import datetime, timedelta
-from pathlib import Path
+from dataclasses import dataclass
 
 import numpy as np
 import xarray as xr
 from matplotlib.axes import Axes
 
-from roms_tools import Grid
 from roms_tools.analysis.cdr_analysis import compute_cdr_metrics
 from roms_tools.plot import plot, plot_uptake_efficiency
 from roms_tools.regrid import LateralRegridFromROMS, VerticalRegridFromROMS
+from roms_tools.roms_dataset import ROMSDataset
 from roms_tools.utils import (
     generate_coordinate_range,
     infer_nominal_horizontal_resolution,
     interpolate_from_rho_to_u,
     interpolate_from_rho_to_v,
-    load_data,
-)
-from roms_tools.vertical_coordinate import (
-    compute_depth_coordinates,
 )
 
 
 @dataclass(kw_only=True)
-class ROMSOutput:
+class ROMSOutput(ROMSDataset):
     """Represents ROMS model output.
 
     Parameters
@@ -45,32 +37,6 @@ class ROMSOutput:
     use_dask: bool, optional
         Indicates whether to use dask for processing. If True, data is processed with dask; if False, data is processed eagerly. Defaults to False.
     """
-
-    grid: Grid
-    """Object representing the grid information."""
-    path: str | Path
-    """Filename, or list of filenames with model output."""
-    use_dask: bool = False
-    """Whether to use dask for processing."""
-    model_reference_date: datetime | None = None
-    """Reference date of ROMS simulation."""
-    adjust_depth_for_sea_surface_height: bool | None = False
-    """Whether to account for sea surface height variations when computing depth
-    coordinates."""
-
-    ds: xr.Dataset = field(init=False, repr=False)
-    """An xarray Dataset containing the ROMS output."""
-
-    def __post_init__(self):
-        ds = self._load_model_output()
-        self._infer_model_reference_date_from_metadata(ds)
-        self._check_vertical_coordinate(ds)
-        ds = self._add_absolute_time(ds)
-        ds = self._add_lat_lon_coords(ds)
-        self.ds = ds
-
-        # Dataset for depth coordinates
-        self.ds_depth_coords = xr.Dataset()
 
     def cdr_metrics(self) -> None:
         """
@@ -387,264 +353,3 @@ class ROMSOutput:
             ds["lat"].attrs = {"long_name": "Latitude", "units": "Degrees North"}
 
             return ds
-
-    def _get_depth_coordinates(self, depth_type="layer", locations=["rho"]):
-        """Ensure depth coordinates are stored for a given location and depth type.
-
-        Calculates vertical depth coordinates (layer or interface) for specified locations (e.g., rho, u, v points)
-        and updates them in the dataset (`self.ds`).
-
-        Parameters
-        ----------
-        depth_type : str
-            The type of depth coordinate to compute. Valid options:
-            - "layer": Compute layer depth coordinates.
-            - "interface": Compute interface depth coordinates.
-        locations : list[str], optional
-            Locations for which to compute depth coordinates. Default is ["rho", "u", "v"].
-            Valid options include:
-            - "rho": Depth coordinates at rho points.
-            - "u": Depth coordinates at u points.
-            - "v": Depth coordinates at v points.
-
-        Updates
-        -------
-        self.ds_depth_coords : xarray.Dataset
-
-        Raises
-        ------
-        ValueError
-            If `adjust_depth_for_sea_surface_height` is enabled but `zeta` is missing from `self.ds`.
-
-        Notes
-        -----
-        - This method relies on the `compute_depth_coordinates` function to perform calculations.
[…]
