+++ roms_tools/tests/test_roms_dataset.py
@@ -1,0 +1,284 @@
+import logging
+import os
+from datetime import datetime
+from pathlib import Path
+
+import pytest
+import xarray as xr
+
+from roms_tools import Grid
+from roms_tools.download import download_test_data
+from roms_tools.roms_dataset import ROMSDataset
+
+try:
+    import xesmf  # type: ignore
+except ImportError:
+    xesmf = None
+
+
+@pytest.fixture
+def roms_dataset_from_restart_file(use_dask):
+    fname_grid = Path(download_test_data("epac25km_grd.nc"))
+    grid = Grid.from_file(fname_grid)
+
+    # Single file
+    return ROMSDataset(
+        grid=grid,
+        path=Path(download_test_data("eastpac25km_rst.19980106000000.nc")),
+        use_dask=use_dask,
+    )
+
+
+@pytest.fixture
+def roms_dataset_from_restart_file_adjusted_for_zeta(use_dask):
+    fname_grid = Path(download_test_data("epac25km_grd.nc"))
+    grid = Grid.from_file(fname_grid)
+
+    # Single file
+    return ROMSDataset(
+        grid=grid,
+        path=Path(download_test_data("eastpac25km_rst.19980106000000.nc")),
+        adjust_depth_for_sea_surface_height=True,
+        use_dask=use_dask,
+    )
+
+
+@pytest.fixture
+def roms_dataset_from_restart_file_with_straddling_grid(use_dask):
+    # Make fake grid that straddles the dateline and that has consistent sizes with test data below
+    grid = Grid(
+        nx=8, ny=13, center_lon=0, center_lat=60, rot=32, size_x=244, size_y=365
+    )
+
+    return ROMSDataset(
+        grid=grid,
+        path=Path(download_test_data("eastpac25km_rst.19980106000000.nc")),
+        use_dask=use_dask,
+    )
+
+
+@pytest.mark.parametrize(
+    "roms_dataset_fixture",
+    [
+        "roms_dataset_from_restart_file",
+        "roms_dataset_from_restart_file_adjusted_for_zeta",
+        "roms_dataset_from_restart_file_with_straddling_grid",
+    ],
+)
+def test_load_model_output_file(roms_dataset_fixture, request):
+    roms_dataset = request.getfixturevalue(roms_dataset_fixture)
+
+    assert isinstance(roms_dataset.ds, xr.Dataset)
+
+
+@pytest.fixture
+def roms_dataset_from_two_restart_files(use_dask):
+    fname_grid = Path(download_test_data("epac25km_grd.nc"))
+    grid = Grid.from_file(fname_grid)
+
+    # List of files
+    file1 = Path(download_test_data("eastpac25km_rst.19980106000000.nc"))
+    file2 = Path(download_test_data("eastpac25km_rst.19980126000000.nc"))
+    return ROMSDataset(grid=grid, path=[file1, file2], use_dask=use_dask)
+
+
+def test_load_model_output_file_list(roms_dataset_from_two_restart_files):
+    assert isinstance(roms_dataset_from_two_restart_files.ds, xr.Dataset)
+
+
+def test_load_model_output_with_wildcard(use_dask):
+    fname_grid = Path(download_test_data("epac25km_grd.nc"))
+    grid = Grid.from_file(fname_grid)
+
+    # Download at least two files, so these will be found within the pooch directory
+    Path(download_test_data("eastpac25km_rst.19980106000000.nc"))
+    Path(download_test_data("eastpac25km_rst.19980126000000.nc"))
+    directory = Path(
+        os.path.dirname(download_test_data("eastpac25km_rst.19980106000000.nc"))
+    )
+
[…]
