+++ roms_tools/tests/test_analysis/test_roms_output.py
@@ -1,6 +1,3 @@
-import logging
-import os
-from datetime import datetime
 from pathlib import Path
 
 import matplotlib.pyplot as plt
@@ -58,20 +55,6 @@ def roms_output_from_restart_file_with_straddling_grid(use_dask):
     )
 
 
-@pytest.mark.parametrize(
-    "roms_output_fixture",
-    [
-        "roms_output_from_restart_file",
-        "roms_output_from_restart_file_adjusted_for_zeta",
-        "roms_output_from_restart_file_with_straddling_grid",
-    ],
-)
-def test_load_model_output_file(roms_output_fixture, request):
-    roms_output = request.getfixturevalue(roms_output_fixture)
-
-    assert isinstance(roms_output.ds, xr.Dataset)
-
-
 @pytest.fixture
 def roms_output_from_two_restart_files(use_dask):
     fname_grid = Path(download_test_data("epac25km_grd.nc"))
@@ -81,208 +64,6 @@ def roms_output_from_two_restart_files(use_dask):
     file1 = Path(download_test_data("eastpac25km_rst.19980106000000.nc"))
     file2 = Path(download_test_data("eastpac25km_rst.19980126000000.nc"))
     return ROMSOutput(grid=grid, path=[file1, file2], use_dask=use_dask)
-
-
-def test_load_model_output_file_list(roms_output_from_two_restart_files):
-    assert isinstance(roms_output_from_two_restart_files.ds, xr.Dataset)
-
-
-def test_load_model_output_with_wildcard(use_dask):
-    fname_grid = Path(download_test_data("epac25km_grd.nc"))
-    grid = Grid.from_file(fname_grid)
-
-    # Download at least two files, so these will be found within the pooch directory
-    Path(download_test_data("eastpac25km_rst.19980106000000.nc"))
-    Path(download_test_data("eastpac25km_rst.19980126000000.nc"))
-    directory = Path(
-        os.path.dirname(download_test_data("eastpac25km_rst.19980106000000.nc"))
-    )
-
-    output = ROMSOutput(grid=grid, path=directory / "*rst*.nc", use_dask=use_dask)
-    assert isinstance(output.ds, xr.Dataset)
-
-
-def test_invalid_path(use_dask):
-    fname_grid = Path(download_test_data("epac25km_grd.nc"))
-    grid = Grid.from_file(fname_grid)
-
-    # Non-existent file
-    with pytest.raises(FileNotFoundError):
-        ROMSOutput(
-            grid=grid,
-            path=Path("/path/to/nonexistent/file.nc"),
-            use_dask=use_dask,
-        )
-
-
-def test_set_correct_model_reference_date(use_dask):
-    fname_grid = Path(download_test_data("epac25km_grd.nc"))
-    grid = Grid.from_file(fname_grid)
-
-    output = ROMSOutput(
-        grid=grid,
-        path=Path(download_test_data("eastpac25km_rst.19980106000000.nc")),
-        use_dask=use_dask,
-    )
-    assert output.model_reference_date == datetime(1995, 1, 1)
-
-
-def test_model_reference_date_mismatch(use_dask):
-    fname_grid = Path(download_test_data("epac25km_grd.nc"))
-    grid = Grid.from_file(fname_grid)
-
-    # Create a ROMSOutput with a specified model_reference_date
-    model_ref_date = datetime(2020, 1, 1)
-    with pytest.raises(
-        ValueError, match="Mismatch between `self.model_reference_date`"
-    ):
-        ROMSOutput(
-            grid=grid,
-            path=Path(download_test_data("eastpac25km_rst.19980106000000.nc")),
-            model_reference_date=model_ref_date,
-            use_dask=use_dask,
-        )
-
-
-def test_model_reference_date_no_metadata(use_dask, tmp_path, caplog):
[…]
