+++ roms_tools/setup/surface_forcing.py
@@ -8,9 +8,7 @@ import numpy as np
 import xarray as xr
 
 from roms_tools import Grid
-from roms_tools.plot import plot
-from roms_tools.regrid import LateralRegridToROMS
-from roms_tools.setup.lat_lon_datasets import (
+from roms_tools.datasets.lat_lon_datasets import (
     CESMBGCSurfaceForcingDataset,
     ERA5ARCODataset,
     ERA5Correction,
@@ -18,6 +16,8 @@ from roms_tools.setup.lat_lon_datasets import (
     LatLonDataset,
     UnifiedBGCSurfaceDataset,
 )
+from roms_tools.plot import plot
+from roms_tools.regrid import LateralRegridToROMS
 from roms_tools.setup.utils import (
     RawDataSource,
     add_time_info_to_ds,
