+++ roms_tools/datasets/river_datasets.py
@@ -7,7 +7,7 @@ from pathlib import Path
 import numpy as np
 import xarray as xr
 
-from roms_tools.download import download_river_data
+from roms_tools.datasets.download import download_river_data
 from roms_tools.setup.utils import (
     assign_dates_to_climatology,
     check_dataset,
