+++ roms_tools/tests/test_analysis/test_roms_output.py
@@ -6,7 +6,7 @@ import pytest
 import xarray as xr
 
 from roms_tools import Grid, ROMSOutput
-from roms_tools.download import download_test_data
+from roms_tools.datasets.download import download_test_data
 
 try:
     import xesmf  # type: ignore
