+++ roms_tools/tests/test_setup/test_utils.py
@@ -7,8 +7,8 @@ import pytest
 import xarray as xr
 
 from roms_tools import BoundaryForcing, Grid
-from roms_tools.download import download_test_data
-from roms_tools.setup.lat_lon_datasets import ERA5Correction
+from roms_tools.datasets.download import download_test_data
+from roms_tools.datasets.lat_lon_datasets import ERA5Correction
 from roms_tools.setup.utils import (
     get_target_coords,
     interpolate_from_climatology,
