### 2025-12-08 fc67252
Shared utility function
