### 2025-12-08 5260870
Test shared utility function
