### 2025-12-08 f81ecb7
Move wrap_longitude function
