### 2025-12-09 1802637
Adjust longitude range
