### 2025-12-09 2c69d61
Make wrap_longitudes function safe
