### 2025-12-09 f4d3c0c
Test choose_subdomain
