### 2025-12-09 726b285
Remove obsolete comment
