### 2025-12-09 feebe9a
Change smooth factor from 6.0 to 5.0
