### 2025-12-09 b26ef97
Test extrapolation to bottom
