### 2025-12-09 3b2764f
Use shared extrapolation function
