### 2025-12-09 d32243a
Move one_dim_fill function
