+++ roms_tools/setup/boundary_forcing.py
@@ -17,6 +17,7 @@ from roms_tools.datasets.lat_lon_datasets import (
     GLORYSDefaultDataset,
     UnifiedBGCDataset,
 )
+from roms_tools.fill import one_dim_fill
 from roms_tools.plot import line_plot, section_plot
 from roms_tools.regrid import LateralRegridToROMS, VerticalRegridToROMS
 from roms_tools.setup.utils import (
@@ -30,7 +31,6 @@ from roms_tools.setup.utils import (
     get_variable_metadata,
     group_dataset,
     nan_check,
-    one_dim_fill,
     rotate_velocities,
     substitute_nans_by_fillvalue,
     to_dict,
