### 2025-12-09 625770d
Test variable_mapping
