### 2025-12-09 90d2085
Add to release notes
