### 2025-12-09 3ea8aa7
Improved bathymetry smoothing to better match UCLA MATLAB tools (#506)
* Don't mask h for local smoothing

* Remove obsolete comment

* Change smooth factor from 6.0 to 5.0

* Overwrite test data due to changed bathy smoothing

* Add to release notes
