### 2025-12-09 07c3ef7
Update methods
