### 2025-12-09 4b7edf7
Allow non-negative tracer fluxes
