+++ roms_tools/tests/test_setup/test_cdr_release.py
@@ -328,12 +328,6 @@ class TestTracerPerturbation:
         with pytest.raises(ValidationError):
             TracerPerturbation(**self.params, tracer_fluxes={"ALK": ["not", "valid"]})
 
-    def test_tracer_fluxes_non_negative(self):
-        with pytest.raises(ValidationError):
-            TracerPerturbation(**self.params, tracer_fluxes={"ALK": -1})
-        with pytest.raises(ValidationError):
-            TracerPerturbation(**self.params, tracer_fluxes={"ALK": [-1, 15]})
-
     def test_zero_fill_strategy(self):
         alk_value = 100.0
         tp = TracerPerturbation(
