### 2025-12-09 3448c20
Test that negative tracer fluxes are allowed
