+++ roms_tools/tests/test_setup/test_cdr_release.py
@@ -328,6 +328,11 @@ class TestTracerPerturbation:
         with pytest.raises(ValidationError):
             TracerPerturbation(**self.params, tracer_fluxes={"ALK": ["not", "valid"]})
 
+    def test_tracer_fluxes_negative(self):
+        tp = TracerPerturbation(**self.params, tracer_fluxes={"DIC": -1})
+
+        assert tp.tracer_fluxes["DIC"].values <= 0.0
+
     def test_zero_fill_strategy(self):
         alk_value = 100.0
         tp = TracerPerturbation(
