### 2025-12-11 5b18663
Shared validate_start_end_time function
