### 2025-12-11 e8bc1c9
Allow negative tracer flux values (#510)
* Allow non-negative tracer fluxes

* Test that negative tracer fluxes are allowed

* Add DOR example to docs

* Add to release notes
