### 2025-12-11 e1126f4
Move, modify and test check_dataset
