+++ roms_tools/datasets/utils.py
@@ -1,3 +1,4 @@
+import logging
 from datetime import datetime
 
 import xarray as xr
@@ -49,6 +50,57 @@ def convert_to_float64(ds: xr.Dataset) -> xr.Dataset:
         Input dataset with data variables converted to double precision.
     """
     return ds.astype({var: "float64" for var in ds.data_vars})
+
+
+def check_dataset(
+    ds: xr.Dataset,
+    dim_names: dict[str, str] | None = None,
+    var_names: dict[str, str] | None = None,
+    opt_var_names: dict[str, str] | None = None,
+) -> None:
+    """Check if the dataset contains the specified variables and dimensions.
+
+    Parameters
+    ----------
+    ds : xr.Dataset
+        The xarray Dataset to check.
+    dim_names: dict[str, str], optional
+        Dictionary specifying the names of dimensions in the dataset.
+    var_names: dict[str, str], optional
+        Dictionary of variable names that are required in the dataset.
+    opt_var_names : dict[str, str], optional
+        Dictionary of optional variable names.
+        These variables are not strictly required, and the function will not raise an error if they are missing.
+        Default is None, meaning no optional variables are considered.
+
+
+    Raises
+    ------
+    ValueError
+        If the dataset does not contain the specified variables or dimensions.
+    """
+    if dim_names:
+        missing_dims = [dim for dim in dim_names.values() if dim not in ds.dims]
+        if missing_dims:
+            raise ValueError(
+                f"Dataset does not contain all required dimensions. The following dimensions are missing: {missing_dims}"
+            )
+
+    if var_names:
+        missing_vars = [var for var in var_names.values() if var not in ds.data_vars]
+        if missing_vars:
+            raise ValueError(
+                f"Dataset does not contain all required variables. The following variables are missing: {missing_vars}"
+            )
+
+    if opt_var_names:
+        missing_optional_vars = [
+            var for var in opt_var_names.values() if var not in ds.data_vars
+        ]
+        if missing_optional_vars:
+            logging.warning(
+                f"Optional variables missing (but not critical): {missing_optional_vars}"
+            )
 
 
 def validate_start_end_time(
