+++ roms_tools/tests/test_datasets/test_utils.py
@@ -1,3 +1,4 @@
+import logging
 from datetime import datetime, timedelta
 
 import numpy as np
@@ -5,6 +6,7 @@ import pytest
 import xarray as xr
 
 from roms_tools.datasets.utils import (
+    check_dataset,
     convert_to_float64,
     extrapolate_deepest_to_bottom,
     validate_start_end_time,
@@ -101,3 +103,63 @@ def test_end_time_before_start_time():
 
     with pytest.raises(ValueError):
         validate_start_end_time(start, end)
+
+
+# tests for check_dataset
+
+
+def test_valid_dataset():
+    ds = xr.Dataset(
+        data_vars={
+            "temp": (("x", "y"), [[1, 2], [3, 4]]),
+            "salt": (("x", "y"), [[5, 6], [7, 8]]),
+        },
+        coords={
+            "x": [0, 1],
+            "y": [0, 1],
+        },
+    )
+
+    dim_names = {"xdim": "x", "ydim": "y"}
+    var_names = {"temperature": "temp", "salinity": "salt"}
+
+    # Should NOT raise
+    check_dataset(ds, dim_names, var_names)
+
+
+def test_missing_required_dimension():
+    ds = xr.Dataset(data_vars={"temp": (("x",), [1])}, coords={"x": [0]})
+
+    dim_names = {"xdim": "x", "ydim": "y"}
+
+    with pytest.raises(ValueError, match="missing"):
+        check_dataset(ds, dim_names=dim_names)
+
+
+def test_missing_required_variable():
+    ds = xr.Dataset(data_vars={"temp": (("x",), [1])}, coords={"x": [0]})
+
+    var_names = {"temperature": "temp", "salinity": "salt"}
+
+    with pytest.raises(ValueError, match="missing"):
+        check_dataset(ds, var_names=var_names)
+
+
+def test_optional_variables_warning(caplog):
+    ds = xr.Dataset(data_vars={"temp": (("x",), [1])}, coords={"x": [0]})
+
+    opt_var_names = {"some_opt": "opt_var"}
+
+    with caplog.at_level(logging.WARNING):
+        # Should not raise, only warn
+        check_dataset(ds, opt_var_names=opt_var_names)
+
+    assert "Optional variables missing" in caplog.text
+    assert "opt_var" in caplog.text
+
+
+def test_no_checks_pass():
+    ds = xr.Dataset(data_vars={"temp": (("x",), [1])}, coords={"x": [0]})
+
+    # Should not fail because no dim_names/var_names provided
+    check_dataset(ds)
