### 2025-12-11 785dacc
More utility function moving and testing
