+++ roms_tools/utils.py
@@ -10,6 +10,7 @@ from pathlib import Path
 from typing import TypeAlias
 
 import numpy as np
+import pandas as pd
 import xarray as xr
 
 from roms_tools.constants import R_EARTH
@@ -1071,3 +1072,153 @@ def wrap_longitudes(ds: xr.Dataset, straddle: bool) -> xr.Dataset:
         ds[lon_dim] = lon_wrapped
 
     return ds
+
+
+def interpolate_from_climatology(
+    field: xr.DataArray | xr.Dataset,
+    time_dim_name: str,
+    time: xr.DataArray | pd.DatetimeIndex,
+) -> xr.DataArray | xr.Dataset:
+    """Interpolates a climatological field to specified time points.
+
+    This function interpolates the input `field` based on `day_of_year` values
+    extracted from the provided `time` points. If `field` is an `xarray.Dataset`,
+    interpolation is applied to all its data variables individually.
+
+    Parameters
+    ----------
+    field : xarray.DataArray or xarray.Dataset
+        The input field to be interpolated.
+        - If `field` is an `xarray.DataArray`, it must have a time dimension identified by `time_dim_name`.
+        - If `field` is an `xarray.Dataset`, all variables within the dataset are interpolated along `time_dim_name`.
+        The time dimension is assumed to represent `day_of_year` for climatological purposes.
+    time_dim_name : str
+        The name of the time dimension in `field`. This dimension is used for interpolation.
+    time : xarray.DataArray or pandas.DatetimeIndex
+        The target time points for interpolation. These are internally converted to `day_of_year`
+        before performing interpolation.
+
+    Returns
+    -------
+    xarray.DataArray or xarray.Dataset
+        The interpolated field, maintaining the same type (`xarray.DataArray` or `xarray.Dataset`)
+        but aligned to the specified `time` values.
+
+    Notes
+    -----
+    - This function assumes that `field` represents a climatological dataset, where time is expressed as `day_of_year` (1-365).
+    - The `time` input is automatically converted to `day_of_year`, so manual conversion is not required before calling this function.
+    """
+
+    def interpolate_single_field(data_array: xr.DataArray) -> xr.DataArray:
+        if isinstance(time, xr.DataArray):
+            # Extract day of year from xarray.DataArray
+            day_of_year = time.dt.dayofyear
+        else:
+            if np.size(time) == 1:
+                # Convert single datetime64 object to pandas.Timestamp
+                date = pd.Timestamp(time)
+                day_of_year = (
+                    date.dayofyear
+                    + (date.hour / 24)
+                    + (date.minute / 1440)
+                    + (date.second / 86400)
+                )
+            else:
+                # Convert each datetime64 object in the array to pandas.Timestamp and compute fractional day of year
+                day_of_year = np.array(
+                    [
+                        pd.Timestamp(t).dayofyear
+                        + (pd.Timestamp(t).hour / 24)
+                        + (pd.Timestamp(t).minute / 1440)
+                        + (pd.Timestamp(t).second / 86400)
+                        for t in time
+                    ]
+                )
+
+        data_array_interpolated = interpolate_cyclic_time(
+            data_array, time_dim_name, day_of_year
+        )
+
+        if np.size(time) == 1:
+            data_array_interpolated = data_array_interpolated.expand_dims(
+                {time_dim_name: 1}
+            )
+        return data_array_interpolated
+
+    if isinstance(field, xr.DataArray):
+        return interpolate_single_field(field)
+    elif isinstance(field, xr.Dataset):
+        interpolated_data_vars = {
+            var: interpolate_single_field(data_array)
+            for var, data_array in field.data_vars.items()
+        }
[…]
