+++ roms_tools/setup/utils.py
@@ -4,12 +4,11 @@ import time
 import typing
 from collections.abc import Sequence
 from dataclasses import asdict, fields, is_dataclass
-from datetime import datetime, timedelta
+from datetime import datetime
 from enum import StrEnum
 from pathlib import Path
 from typing import Any, Literal, TypeAlias
 
-import cftime
 import numba as nb
 import numpy as np
 import pandas as pd
@@ -154,240 +153,6 @@ def assign_dates_to_climatology(ds: xr.Dataset, time_dim: str) -> xr.Dataset:
     time = xr.DataArray(timedelta_ns, dims=[time_dim])
     ds = ds.assign_coords({"time": time})
     return ds
-
-
-def interpolate_cyclic_time(
-    data_array: xr.DataArray,
-    time_dim_name: str,
-    day_of_year: int | float | np.ndarray | xr.DataArray | Sequence[int | float],
-) -> xr.DataArray:
-    """Interpolates a DataArray cyclically across the start and end of the year.
-
-    This function extends the data cyclically by appending the last time step
-    (shifted back by one year) at the beginning and the first time step
-    (shifted forward by one year) at the end. It then performs linear interpolation
-    to match the specified `day_of_year` values.
-
-    Parameters
-    ----------
-    data_array : xr.DataArray
-        The input data array containing a time-like dimension.
-    time_dim_name : str
-        The name of the time dimension in the dataset.
-    day_of_year : Union[int, float, np.ndarray, xr.DataArray, Sequence[Union[int, float]]]
-        The target day(s) of the year for interpolation. This can be:
-        - A single integer or float representing the day of the year.
-        - A NumPy array or xarray DataArray containing multiple days.
-        - A list or tuple of integers or floats for multiple target days.
-
-    Returns
-    -------
-    xr.DataArray
-        The interpolated DataArray, ensuring cyclic continuity across year boundaries.
-
-    Notes
-    -----
-    - This function is useful for interpolating climatological data, where the time axis
-      represents a repeating annual cycle.
-    - The `day_of_year` values should be within the range [1, 365] or [1, 366] for leap years.
-    """
-    # Concatenate across the beginning and end of the year
-    time_concat = xr.concat(
-        [
-            data_array[time_dim_name][-1] - 365.25,  # Shift last time backward
-            data_array[time_dim_name],
-            data_array[time_dim_name][0] + 365.25,  # Shift first time forward
-        ],
-        dim=time_dim_name,
-    )
-
-    data_array_concat = xr.concat(
-        [
-            data_array.isel(
-                **{time_dim_name: -1}
-            ),  # Append last value at the beginning
-            data_array,
-            data_array.isel(**{time_dim_name: 0}),  # Append first value at the end
-        ],
-        dim=time_dim_name,
-    )
-    data_array_concat[time_dim_name] = time_concat
-
-    # Interpolate to specified times
-    data_array_interpolated = data_array_concat.interp(
-        **{time_dim_name: day_of_year}, method="linear"
-    )
-
-    return data_array_interpolated
-
-
-def interpolate_from_climatology(
-    field: xr.DataArray | xr.Dataset,
-    time_dim_name: str,
-    time: xr.DataArray | pd.DatetimeIndex,
-) -> xr.DataArray | xr.Dataset:
-    """Interpolates a climatological field to specified time points.
-
-    This function interpolates the input `field` based on `day_of_year` values
-    extracted from the provided `time` points. If `field` is an `xarray.Dataset`,
-    interpolation is applied to all its data variables individually.
-
-    Parameters
-    ----------
-    field : xarray.DataArray or xarray.Dataset
-        The input field to be interpolated.
-        - If `field` is an `xarray.DataArray`, it must have a time dimension identified by `time_dim_name`.
[…]
