### 2025-12-11 2d1c962
Explicit parameters
