### 2025-12-11 a59ea8f
Add more methods and params to ROMSDataset
