+++ roms_tools/datasets/roms_dataset.py
@@ -9,7 +9,13 @@ import numpy as np
 import xarray as xr
 
 from roms_tools import Grid
-from roms_tools.datasets.utils import convert_to_float64, extrapolate_deepest_to_bottom
+from roms_tools.datasets.utils import (
+    check_dataset,
+    convert_to_float64,
+    extrapolate_deepest_to_bottom,
+    select_relevant_times,
+    validate_start_end_time,
+)
 from roms_tools.fill import LateralFill
 from roms_tools.utils import load_data, wrap_longitudes
 from roms_tools.vertical_coordinate import (
@@ -32,10 +38,28 @@ class ROMSDataset:
 
     Parameters
     ----------
+    path: str | Path | list[str | Path]
+        Filename, or list of filenames with model output.
     grid : Grid
         Object representing the grid information.
-    path : Union[str, Path, List[Union[str, Path]]]
-        Filename, or list of filenames with model output.
+    start_time : Optional[datetime], optional
+        Start time for selecting relevant data. If not provided, no time-based filtering is applied.
+    end_time : Optional[datetime], optional
+        End time for selecting relevant data. If not provided, the dataset selects the time entry
+        closest to `start_time` within the range `[start_time, start_time + 24 hours)`.
+        If `start_time` is also not provided, no time-based filtering is applied.
+    allow_flex_time: bool, optional
+        Controls how strictly the dataset selects a time entry when `end_time` is not provided (relevant for initial conditions):
+
+        - If False (default): requires an exact match to `start_time`. Raises a ValueError if no match exists.
+        - If True: allows a +24h search window after `start_time` and selects the closest available
+          time entry within that window. Raises a ValueError if none are found.
+
+        Only used when `end_time` is None. Has no effect otherwise.
+    dim_names: dict[str, str], optional
+        Dictionary specifying the names of dimensions in the dataset.
+    var_names: dict[str, str], optional
+        Dictionary of variable names that are required in the dataset.
     model_reference_date : datetime, optional
         Reference date of ROMS simulation.
         If not specified, this is inferred from metadata of the model output
@@ -47,10 +71,27 @@ class ROMSDataset:
         Indicates whether to use dask for processing. If True, data is processed with dask; if False, data is processed eagerly. Defaults to False.
     """
 
+    path: str | Path | list[str | Path]
+    """Filename, or list of filenames with model output."""
     grid: Grid
     """Object representing the grid information."""
-    path: str | Path
-    """Filename, or list of filenames with model output."""
+    start_time: datetime | None = None
+    """Start time for selecting relevant data."""
+    end_time: datetime | None = None
+    """End time for selecting relevant data."""
+    allow_flex_time: bool = False
+    """Controls how strictly the dataset selects a time entry when `end_time` is not provided."""
+    dim_names: dict[str, str] = field(
+        default_factory=lambda: {
+            "eta_rho": "eta_rho",
+            "xi_rho": "xi_rho",
+            "s_rho": "s_rho",
+            "time": "time",
+        }
+    )
+    """Dictionary specifying the names of dimensions in the dataset."""
+    var_names: dict[str, str] | None = None
+    """Dictionary of variable names that are required in the dataset."""
     use_dask: bool = False
     """Whether to use dask for processing."""
     model_reference_date: datetime | None = None
@@ -63,10 +104,17 @@ class ROMSDataset:
     """An xarray Dataset containing the ROMS output."""
 
     def __post_init__(self):
-        ds = self._load_model_output()
-        self._infer_model_reference_date_from_metadata(ds)
+        validate_start_end_time(self.start_time, self.end_time)
+        ds = self.load_data()
[…]
