### 2025-12-11 de80240
Set required vars in IC
