+++ roms_tools/setup/initial_conditions.py
@@ -995,3 +995,90 @@ def _set_dask_chunks(location: str, chunk_size: int):
         "v": {"eta_v": chunk_size, "xi_rho": chunk_size},
     }
     return chunk_mapping.get(location, {})
+
+
+def _set_variable_mapping(self, var_type: str = "physics") -> None:
+    """
+    Set up variable name mapping for the ROMS dataset.
+
+    This mapping makes the dataset compatible with `InitialConditions`.
+
+    Parameters
+    ----------
+    var_type : str, optional
+        Type of variables to map. Supported values are:
+        - "physics" : physical variables like temperature, salinity, currents.
+        - "bgc" : biogeochemical variables like nutrients, chlorophyll, and carbon.
+        Default is "physics".
+
+    Raises
+    ------
+    ValueError
+        If `var_type` is not one of the supported types.
+    KeyError
+        If any of the mapped variables are missing from `self.ds`.
+    """
+    # Define variable mappings
+    var_mappings = {
+        "physics": {
+            "zeta": "zeta",
+            "temp": "temp",
+            "salt": "salt",
+            "u": "u",
+            "v": "v",
+        },
+        "bgc": {
+            "PO4": "PO4",
+            "NO3": "NO3",
+            "SiO3": "SiO3",
+            "NH4": "NH4",
+            "Fe": "Fe",
+            "Lig": "Lig",
+            "O2": "O2",
+            "DIC": "DIC",
+            "DIC_ALT_CO2": "DIC_ALT_CO2",
+            "ALK": "ALK",
+            "ALK_ALT_CO2": "ALK_ALT_CO2",
+            "DOC": "DOC",
+            "DON": "DON",
+            "DOP": "DOP",
+            "DOPr": "DOPr",
+            "DONr": "DONr",
+            "DOCr": "DOCr",
+            "spChl": "spChl",
+            "spC": "spC",
+            "spP": "spP",
+            "spFe": "spFe",
+            "diatChl": "diatChl",
+            "diatC": "diatC",
+            "diatP": "diatP",
+            "diatFe": "diatFe",
+            "diatSi": "diatSi",
+            "diazChl": "diazChl",
+            "diazC": "diazC",
+            "diazP": "diazP",
+            "diazFe": "diazFe",
+            "spCaCO3": "spCaCO3",
+            "zooC": "zooC",
+        },
+    }
+
+    # Validate type
+    if var_type not in var_mappings:
+        raise ValueError(
+            f"Unsupported var_type '{var_type}'. Choose from {list(var_mappings.keys())}."
+        )
+
+    # Assign variable mapping
+    self.var_names = var_mappings[var_type]
+
+    # Check all mapped variables exist in the dataset
+    missing_vars = [
+        v for v in self.var_names.values() if v not in self.ds.variables
+    ]
+    if missing_vars:
+        raise KeyError(
+            f"The following variables are missing from the dataset: {missing_vars}"
+        )
+
+
