### 2025-12-11 5620fc2
Set required vars in IC
