### 2025-12-11 b7c97db
Remove obsolete test
