### 2025-12-11 305d2bd
Correct setting of required vars
