+++ roms_tools/tests/test_setup/test_initial_conditions.py
@@ -678,74 +678,28 @@ def test_from_yaml_missing_initial_conditions(tmp_path, use_dask):
         yaml_filepath.unlink()
 
 
-# Test _set_variable_mapping
+# Test _set_required_vars
+
+def test_default_var_type():
+    vars_map = _set_required_vars()
+    # Default is "physics"
+    expected_keys = {"zeta", "temp", "salt", "u", "v"}
+    assert set(vars_map.keys()) == expected_keys
+    # Values should match keys
+    for key, val in vars_map.items():
+        assert key == val
 
 
-def make_dummy_roms_dataset(var_names):
-    """Create a minimal ROMSDataset with given variables."""
-    ds = xr.Dataset(
-        {name: (("eta_rho", "xi_rho"), np.zeros((2, 2))) for name in var_names}
-    )
-    roms_ds = ROMSDataset.__new__(ROMSDataset)
-    roms_ds.ds = ds
-    return roms_ds
+def test_bgc_var_type():
+    vars_map = _set_required_vars("bgc")
+    # Check a few expected keys exist
+    expected_keys = {"PO4", "NO3", "SiO3", "NH4", "Fe", "DIC", "spChl", "zooC"}
+    assert expected_keys.issubset(vars_map.keys())
+    # Values should match keys
+    for key, val in vars_map.items():
+        assert key == val
 
 
-def test_set_variable_mapping_physics():
-    roms_ds_missing = make_dummy_roms_dataset(
-        ["zeta", "temp", "salt", "u"]
-    )  # missing 'v'
-    with pytest.raises(KeyError, match="v"):
-        roms_ds_missing._set_variable_mapping("physics")
-
-    roms_ds = make_dummy_roms_dataset(["zeta", "temp", "salt", "u", "v"])
-    roms_ds._set_variable_mapping("physics")
-    assert roms_ds.var_names["temp"] == "temp"
-
-
-def test_set_variable_mapping_bgc():
-    bgc_vars = ["PO4", "NO3", "SiO3", "NH4", "Fe", "Lig", "O2", "DIC"]
-    roms_ds = make_dummy_roms_dataset(bgc_vars)
-    with pytest.raises(KeyError):
-        roms_ds._set_variable_mapping("bgc")
-
-    roms_ds_all = make_dummy_roms_dataset(
-        [
-            *bgc_vars,
-            "DIC_ALT_CO2",
-            "ALK",
-            "ALK_ALT_CO2",
-            "DOC",
-            "DON",
-            "DOP",
-            "DOPr",
-            "DONr",
-            "DOCr",
-            "spChl",
-            "spC",
-            "spP",
-            "spFe",
-            "diatChl",
-            "diatC",
-            "diatP",
-            "diatFe",
-            "diatSi",
-            "diazChl",
-            "diazC",
-            "diazP",
-            "diazFe",
-            "spCaCO3",
-            "zooC",
-        ]
-    )
-
-    roms_ds_all._set_variable_mapping("bgc")
-    assert "PO4" in roms_ds_all.var_names.values()
-
-
-def test_set_variable_mapping_invalid_type():
-    roms_ds = make_dummy_roms_dataset(["zeta", "temp", "salt", "u", "v"])
+def test_invalid_var_type():
     with pytest.raises(ValueError, match="Unsupported var_type"):
-        roms_ds._set_variable_mapping("invalid_type")
-
-
+        _set_required_vars("invalid_type")
