### 2025-12-11 7ad25b6
Restore functionality for non-ROMS sources
