### 2025-12-11 e37fbba
Improve adding masks
