### 2025-12-11 6b25607
Add consistency check
