### 2025-12-11 1151ec1
Introduce time_dim and time_coord separately
