+++ roms_tools/datasets/utils.py
@@ -150,6 +150,7 @@ def validate_start_end_time(
 def select_relevant_times(
     ds: xr.Dataset,
     time_dim: str,
+    time_coord: str,
     start_time: datetime,
     end_time: datetime | None = None,
     climatology: bool = False,
@@ -184,6 +185,8 @@ def select_relevant_times(
         The dataset to filter. Must contain a valid time dimension.
     time_dim : str
         Name of the time dimension in `ds`.
+    time_coord : str
+        Name of the time coordinate in `ds`.
     start_time : datetime
         Start time for filtering.
     end_time : datetime or None
@@ -221,19 +224,26 @@ def select_relevant_times(
     - Logs warnings instead of failing hard when boundary records are missing, and
       defaults to using the earliest or latest available time in such cases.
     """
-    if time_dim not in ds.variables:
+    if time_dim not in ds.dims:
         logging.warning(
             f"Dataset does not contain time dimension '{time_dim}'. "
             "Please check variable naming or dataset structure."
         )
         return ds
 
-    time_type = get_time_type(ds[time_dim])
+    if time_coord not in ds.variables:
+        logging.warning(
+            f"Dataset does not contain time coordinate '{time_coord}'. "
+            "Please check variable naming or dataset structure."
+        )
+        return ds
+
+    time_type = get_time_type(ds[time_coord])
 
     if climatology:
-        if len(ds[time_dim]) != 12:
+        if len(ds[time_coord]) != 12:
             raise ValueError(
-                f"The dataset contains {len(ds[time_dim])} time steps, but the climatology flag is set to True, which requires exactly 12 time steps."
+                f"The dataset contains {len(ds[time_coord])} time steps, but the climatology flag is set to True, which requires exactly 12 time steps."
             )
     else:
         if time_type == "int":
@@ -241,41 +251,41 @@ def select_relevant_times(
                 "The dataset contains integer time values, which are only supported when the climatology flag is set to True. However, your climatology flag is set to False."
             )
     if time_type == "cftime":
-        ds = ds.assign_coords({time_dim: convert_cftime_to_datetime(ds[time_dim])})
+        ds = ds.assign_coords({time_dim: convert_cftime_to_datetime(ds[time_coord])})
 
     if not end_time:
         # Assume we are looking for exactly one time record for initial conditions
         return _select_initial_time(
-            ds, time_dim, start_time, climatology, allow_flex_time
+            ds, time_dim, time_coord, start_time, climatology, allow_flex_time
         )
 
     if climatology:
         return ds
 
     # Identify records before or at start_time
-    before_start = ds[time_dim] <= np.datetime64(start_time)
+    before_start = ds[time_coord] <= np.datetime64(start_time)
     if before_start.any():
-        closest_before_start = ds[time_dim].where(before_start, drop=True)[-1]
+        closest_before_start = ds[time_coord].where(before_start, drop=True)[-1]
     else:
         logging.warning(f"No records found at or before the start_time: {start_time}.")
-        closest_before_start = ds[time_dim][0]
+        closest_before_start = ds[time_coord][0]
 
     # Identify records after or at end_time
-    after_end = ds[time_dim] >= np.datetime64(end_time)
+    after_end = ds[time_coord] >= np.datetime64(end_time)
     if after_end.any():
-        closest_after_end = ds[time_dim].where(after_end, drop=True).min()
+        closest_after_end = ds[time_coord].where(after_end, drop=True).min()
     else:
         logging.warning(f"No records found at or after the end_time: {end_time}.")
-        closest_after_end = ds[time_dim].max()
[…]
