### 2025-12-11 76d1af9
Make sure time selection works on ROMS datasets
