### 2025-12-11 73b1aa3
Improve docstrings
