### 2025-12-12 1526ce1
select_relevant_fields as shared utils
