### 2025-12-12 bc84dc6
Test subsetting of u and v
