### 2025-12-15 d0793e2
Avoid summing entire array
