### 2025-12-15 32aa504
Default boundary selection respects land-only boundaries (#504)
* Prohibit child grids with boundaries on land

* Correct tests with preaction of wrapping lons

* Add to release notes

* Functions for checking and validating boundaries dict

* Overwrite test data due to new defaults

* New boundaries defaults

* Change logging message

* Rerun notebook

* Fix release notes

* Fix merge conflict

* Fix merge conflict

* Update notebook

* Fix small bug

* Add to release notes

* Avoid summing entire array
