### 2025-12-15 dd0aa8e
Fix vertical regridding
