### 2025-12-15 d0fd371
Make lateral filling safe
