### 2025-12-15 b260d81
Only interpolate layer depths once
