### 2025-12-15 aeb862f
Don't convert masks to float64
