### 2025-12-15 4b8aadc
Correct new convert to float64 implementation
