### 2025-12-15 93f26e6
Bulk-drop variables for optimization
