### 2025-12-15 0299779
Remove list
