### 2025-12-15 6dbac69
No repeated pd.Timestamp conversions
