+++ roms_tools/utils.py
@@ -10,6 +10,7 @@ from pathlib import Path
 from typing import TypeAlias
 
 import numpy as np
+import pandas as pd
 import xarray as xr
 
 from roms_tools.constants import R_EARTH
@@ -1031,3 +1032,195 @@ def get_pkg_error_msg(purpose: str, package_name: str, option_name: str) -> str:
           • `pip install {package_name}` or
           • `conda install {package_name}`
         Alternatively, install `roms-tools` with conda to include all dependencies.""")
+
+
+def wrap_longitudes(ds: xr.Dataset, straddle: bool) -> xr.Dataset:
+    """
+    Safely adjust longitude variables for datasets that may or may not cross
+    the dateline. Only modifies longitude-like variables that are present.
+
+    Parameters
+    ----------
+    ds : xr.Dataset
+        Dataset containing longitude variables (e.g., lon_rho, lon_u, lon_v).
+    straddle : bool
+        - True: force longitudes into [-180, 180]
+        - False: force longitudes into [0, 360]
+
+    Returns
+    -------
+    xr.Dataset
+        A new dataset with adjusted longitude values.
+    """
+    lon_vars = ["lon_rho", "lon_u", "lon_v"]
+
+    for lon_dim in lon_vars:
+        if lon_dim not in ds:
+            continue  # skip missing coordinate
+
+        lon = ds[lon_dim]
+
+        if straddle:
+            # wrap into [-180, 180]
+            lon_wrapped = xr.where(lon > 180, lon - 360, lon)
+        else:
+            # wrap into [0, 360]
+            lon_wrapped = xr.where(lon < 0, lon + 360, lon)
+
+        # preserve attributes
+        lon_wrapped.attrs.update(lon.attrs)
+        ds[lon_dim] = lon_wrapped
+
+    return ds
+
+
+def interpolate_from_climatology(
+    field: xr.DataArray | xr.Dataset,
+    time_dim: str,
+    time_coord: str,
+    time: xr.DataArray | pd.DatetimeIndex,
+) -> xr.DataArray | xr.Dataset:
+    """Interpolates a climatological field to specified time points.
+
+    This function interpolates the input `field` based on `day_of_year` values
+    extracted from the provided `time` points. If `field` is an `xarray.Dataset`,
+    interpolation is applied to all its data variables individually.
+
+    Parameters
+    ----------
+    field : xarray.DataArray or xarray.Dataset
+        The input field to be interpolated.
+        - If `field` is an `xarray.DataArray`, it must have a time dimension identified by `time_dim_name`.
+        - If `field` is an `xarray.Dataset`, all variables within the dataset are interpolated along `time_dim_name`.
+        The time dimension is assumed to represent `day_of_year` for climatological purposes.
+    time_dim : str
+        The name of the time dimension in `field`.
+    time_coord : str
+        The name of the time coordinate in `field`.
+    time : xarray.DataArray or pandas.DatetimeIndex
+        The target time points for interpolation. These are internally converted to `day_of_year`
+        before performing interpolation.
+
+    Returns
+    -------
+    xarray.DataArray or xarray.Dataset
+        The interpolated field, maintaining the same type (`xarray.DataArray` or `xarray.Dataset`)
+        but aligned to the specified `time` values.
+
+    Notes
+    -----
+    - This function assumes that `field` represents a climatological dataset, where time is expressed as `day_of_year` (1-365).
+    - The `time` input is automatically converted to `day_of_year`, so manual conversion is not required before calling this function.
+    """
+
+    def np_times_to_fractional_days(
+        np_times: np.ndarray | pd.DatetimeIndex | np.datetime64 | pd.Timestamp,
+    ) -> np.ndarray:
+        """Convert datetime(s) to fractional day-of-year values."""
+        pd_times = pd.to_datetime(np_times)
+
[…]
