+++ roms_tools/analysis/roms_output.py
@@ -1,16 +1,11 @@
-import logging
-import re
-import warnings
-from dataclasses import dataclass, field
-from datetime import datetime, timedelta
-from pathlib import Path
+from dataclasses import dataclass
 
 import numpy as np
 import xarray as xr
 from matplotlib.axes import Axes
 
-from roms_tools import Grid
 from roms_tools.analysis.cdr_analysis import compute_cdr_metrics
+from roms_tools.datasets.roms_dataset import ROMSDataset
 from roms_tools.plot import plot, plot_uptake_efficiency
 from roms_tools.regrid import LateralRegridFromROMS, VerticalRegridFromROMS
 from roms_tools.utils import (
@@ -18,23 +13,19 @@ from roms_tools.utils import (
     infer_nominal_horizontal_resolution,
     interpolate_from_rho_to_u,
     interpolate_from_rho_to_v,
-    load_data,
-)
-from roms_tools.vertical_coordinate import (
-    compute_depth_coordinates,
 )
 
 
 @dataclass(kw_only=True)
-class ROMSOutput:
+class ROMSOutput(ROMSDataset):
     """Represents ROMS model output.
 
     Parameters
     ----------
+    path: str | Path | list[str | Path]
+        Filename, or list of filenames with model output.
     grid : Grid
         Object representing the grid information.
-    path : Union[str, Path, List[Union[str, Path]]]
-        Filename, or list of filenames with model output.
     model_reference_date : datetime, optional
         Reference date of ROMS simulation.
         If not specified, this is inferred from metadata of the model output
@@ -45,32 +36,6 @@ class ROMSOutput:
     use_dask: bool, optional
         Indicates whether to use dask for processing. If True, data is processed with dask; if False, data is processed eagerly. Defaults to False.
     """
-
-    grid: Grid
-    """Object representing the grid information."""
-    path: str | Path
-    """Filename, or list of filenames with model output."""
-    use_dask: bool = False
-    """Whether to use dask for processing."""
-    model_reference_date: datetime | None = None
-    """Reference date of ROMS simulation."""
-    adjust_depth_for_sea_surface_height: bool | None = False
-    """Whether to account for sea surface height variations when computing depth
-    coordinates."""
-
-    ds: xr.Dataset = field(init=False, repr=False)
-    """An xarray Dataset containing the ROMS output."""
-
-    def __post_init__(self):
-        ds = self._load_model_output()
-        self._infer_model_reference_date_from_metadata(ds)
-        self._check_vertical_coordinate(ds)
-        ds = self._add_absolute_time(ds)
-        ds = self._add_lat_lon_coords(ds)
-        self.ds = ds
-
-        # Dataset for depth coordinates
-        self.ds_depth_coords = xr.Dataset()
 
     def cdr_metrics(self) -> None:
         """
@@ -379,272 +344,8 @@ class ROMSOutput:
         if regridded_datasets:
             ds = xr.merge(regridded_datasets)
 
-            with warnings.catch_warnings():
-                warnings.filterwarnings("ignore", category=UserWarning)
-                ds = ds.rename({"abs_time": "time"}).set_index(time="time")
             ds["time"].attrs = {"long_name": "Time"}
             ds["lon"].attrs = {"long_name": "Longitude", "units": "Degrees East"}
             ds["lat"].attrs = {"long_name": "Latitude", "units": "Degrees North"}
 
             return ds
-
-    def _get_depth_coordinates(self, depth_type="layer", locations=["rho"]):
-        """Ensure depth coordinates are stored for a given location and depth type.
-
-        Calculates vertical depth coordinates (layer or interface) for specified locations (e.g., rho, u, v points)
-        and updates them in the dataset (`self.ds`).
-
-        Parameters
-        ----------
[…]
