+++ roms_tools/datasets/river_datasets.py
@@ -7,12 +7,11 @@ from pathlib import Path
 import numpy as np
 import xarray as xr
 
-from roms_tools.download import download_river_data
+from roms_tools.datasets.download import download_river_data
+from roms_tools.datasets.utils import check_dataset, select_relevant_times
 from roms_tools.setup.utils import (
     assign_dates_to_climatology,
-    check_dataset,
     gc_dist,
-    select_relevant_times,
 )
 from roms_tools.utils import load_data
 
@@ -198,7 +197,13 @@ class RiverDataset:
         """
         time_dim = self.dim_names["time"]
 
-        ds = select_relevant_times(ds, time_dim, self.start_time, self.end_time, False)
+        ds = select_relevant_times(
+            ds=ds,
+            time_dim=time_dim,
+            time_coord=time_dim,
+            start_time=self.start_time,
+            end_time=self.end_time,
+        )
 
         return ds
 

[…]
