### 2025-12-16 2efd849
Add test for VerticalRegrid
