### 2025-12-16 09dc699
Don't allow ini_time=None
