### 2025-12-17 cdb6222
Raise error if ini_time = None
