### 2025-12-18 bc1f789
Make to_dict compatible with new IC params
