+++ roms_tools/setup/utils.py
@@ -3,6 +3,7 @@ import logging
 import time
 import typing
 from collections.abc import Sequence
+from copy import deepcopy
 from dataclasses import asdict, fields, is_dataclass
 from datetime import datetime
 from enum import StrEnum
@@ -1368,83 +1369,108 @@ def write_to_yaml(yaml_data, filepath: str | Path) -> None:
 def to_dict(forcing_object, exclude: list[str] | None = None) -> dict:
     """Serialize a forcing object (including its grid) into a dictionary.
 
-    This function serializes a dataclass object (forcing_object) and its associated
-    `grid` attribute into a dictionary. It omits fields like `grid` and `ds`
-    that are not serializable or meant to be excluded.
+    This function serializes a forcing object (dataclass or pydantic model),
+    including its associated grid(s), into a dictionary suitable for YAML output.
 
-    The function also converts datetime fields to ISO format strings for proper
-    serialization.
+    - Top-level grids (`grid`, `parent_grid`) are serialized consistently
+    - Nested grids inside `source` and `bgc_source` are also serialized
+    - Datetime objects are converted to ISO strings
+    - Path objects are converted to strings
 
     Parameters
     ----------
     forcing_object : object
-        The object that contains the forcing data, typically a dataclass with attributes
-        such as `grid`, `start_time`, `end_time`, etc.
+        A dataclass or pydantic model representing a forcing configuration.
     exclude : list[str], optional
-        List of keys to exclude from the serialized output. Defaults to empty list. The field "ds" is always excluded by default.
+        List of field names to exclude from serialization. The fields
+        "grid", "parent_grid", and "ds" are always excluded.
 
     Returns
     -------
     dict
+        Serialized representation of the forcing object.
     """
-    # Serialize Grid data
-    if hasattr(forcing_object, "grid") and forcing_object.grid is not None:
-        grid_data = asdict(forcing_object.grid)
-        grid_yaml_data = {"Grid": pop_grid_data(grid_data)}
-    elif hasattr(forcing_object, "parent_grid"):
-        grid_data = asdict(forcing_object.parent_grid)
-        grid_yaml_data = {"ParentGrid": pop_grid_data(grid_data)}
+    exclude_list = exclude or []
+    exclude_set: set[str] = {"grid", "parent_grid", "ds", *exclude_list}
 
-    # Ensure Paths are Strings
-    def ensure_paths_are_strings(obj, key):
-        attr = getattr(obj, key, None)
-        if attr is not None and "path" in attr:
-            paths = attr["path"]
-            if isinstance(paths, list):
-                attr["path"] = [str(p) if isinstance(p, Path) else p for p in paths]
-            elif isinstance(paths, Path):
-                attr["path"] = str(paths)
-            elif isinstance(paths, dict):
-                for key, path in paths.items():
-                    attr["path"][key] = str(path)
+    def serialize_grid(grid_obj):
+        """Serialize a Grid object consistently."""
+        return pop_grid_data(asdict(grid_obj))
 
-    ensure_paths_are_strings(forcing_object, "source")
-    ensure_paths_are_strings(forcing_object, "bgc_source")
+    def serialize_paths(value):
+        """Convert Path objects (including nested containers) to strings."""
+        if isinstance(value, Path):
+            return str(value)
+        if isinstance(value, list):
+            return [serialize_paths(v) for v in value]
+        if isinstance(value, dict):
+            return {k: serialize_paths(v) for k, v in value.items()}
+        return value
 
-    # Prepare Forcing Data
-    forcing_data = {}
-    if isinstance(forcing_object, BaseModel):
-        field_names = forcing_object.model_fields
-    elif is_dataclass(forcing_object):
[…]
