### 2025-12-18 94a5798
Clean up some utils and deserialize datetime
