+++ roms_tools/setup/utils.py
@@ -1366,6 +1366,93 @@ def write_to_yaml(yaml_data, filepath: str | Path) -> None:
         )
 
 
+def serialize_paths(value: Any) -> Any:
+    """Recursively convert Path objects to strings."""
+    if isinstance(value, Path):
+        return str(value)
+    if isinstance(value, list):
+        return [serialize_paths(v) for v in value]
+    if isinstance(value, dict):
+        return {k: serialize_paths(v) for k, v in value.items()}
+    return value
+
+
+def normalize_paths(value: Any) -> Any:
+    """Recursively convert path-like strings back to Path objects.
+
+    Heuristic: strings containing '/' or ending with '.nc' are treated as paths.
+    """
+    if isinstance(value, str):
+        return Path(value) if "/" in value or value.endswith(".nc") else value
+    if isinstance(value, list):
+        return [normalize_paths(v) for v in value]
+    if isinstance(value, dict):
+        return {k: normalize_paths(v) for k, v in value.items()}
+    return value
+
+
+def serialize_datetime(value: datetime | list[datetime] | Any) -> Any:
+    """Convert datetime or list of datetimes to ISO 8601 strings."""
+    if isinstance(value, datetime):
+        return value.isoformat()
+    if isinstance(value, list) and all(isinstance(v, datetime) for v in value):
+        return [v.isoformat() for v in value]
+    return value
+
+
+def deserialize_datetime(
+    value: str | list[str] | datetime | Any,
+) -> datetime | list[datetime] | Any:
+    """Convert ISO 8601 string(s) to datetime object(s).
+
+    Returns:
+        datetime if input is string,
+        list of datetime if input is list of strings,
+        original value if parsing fails or input is already datetime.
+    """
+    if isinstance(value, list):
+        result: list[datetime | Any] = []
+        for v in value:
+            try:
+                result.append(datetime.fromisoformat(str(v)))
+            except ValueError:
+                result.append(v)
+        return result
+
+    if isinstance(value, str):
+        try:
+            return datetime.fromisoformat(value)
+        except ValueError:
+            return value
+
+    return value
+
+
+def serialize_grid(grid_obj: Any) -> dict[str, Any]:
+    """Serialize a Grid object to a dictionary, excluding non-serializable attributes."""
+    return pop_grid_data(asdict(grid_obj))
+
+
+def pop_grid_data(grid_data: dict[str, Any]) -> dict[str, Any]:
+    """Remove non-serializable or unnecessary keys from a Grid dictionary.
+
+    Removes 'ds', 'straddle', and 'verbose' keys if present.
+
+    Parameters
+    ----------
+    grid_data : dict
+        Dictionary representation of a Grid object.
+
+    Returns
+    -------
+    dict
+        Cleaned dictionary suitable for serialization.
+    """
+    for key in ("ds", "straddle", "verbose"):
+        grid_data.pop(key, None)
+    return grid_data
+
+
 def to_dict(forcing_object, exclude: list[str] | None = None) -> dict:
     """Serialize a forcing object (including its grid) into a dictionary.
 
@@ -1392,28 +1479,6 @@ def to_dict(forcing_object, exclude: list[str] | None = None) -> dict:
     """
     exclude_list = exclude or []
     exclude_set: set[str] = {"grid", "parent_grid", "ds", *exclude_list}
-
-    def serialize_grid(grid_obj):
-        """Serialize a Grid object consistently."""
-        return pop_grid_data(asdict(grid_obj))
-
-    def serialize_paths(value):
-        """Convert Path objects (including nested containers) to strings."""
-        if isinstance(value, Path):
-            return str(value)
-        if isinstance(value, list):
-            return [serialize_paths(v) for v in value]
[…]
