### 2025-12-18 1be50d2
Clean serialization and deserialization of source dict
