### 2025-12-18 50a29ab
Add to release notes
