### 2025-12-18 4055c8f
Update release notes
