### 2025-12-18 8d16b32
Improve docstrings and typehints of VerticalRegrid
