### 2025-12-20 dd49b82
Adjust depth for SSH for IC from ROMS
