### 2025-12-20 cf08f3b
Check that IC are unchanged for identical grids
