### 2025-12-20 359a780
Don't expose adjust_depth_for_sea_surface_height to user
