+++ conftest.py
@@ -340,34 +340,6 @@ def initial_conditions_on_large_grid(large_grid, use_dask):
 
 
 @pytest.fixture(scope="session")
-def initial_conditions_adjusted_for_zeta(use_dask: bool) -> InitialConditions:
-    """Fixture for creating an InitialConditions object."""
-    grid = Grid(
-        nx=2,
-        ny=2,
-        size_x=500,
-        size_y=1000,
-        center_lon=0,
-        center_lat=55,
-        rot=10,
-        N=3,  # number of vertical levels
-        theta_s=5.0,  # surface control parameter
-        theta_b=2.0,  # bottom control parameter
-        hc=250.0,  # critical depth
-    )
-
-    fname = Path(download_test_data("GLORYS_coarse_test_data.nc"))
-
-    return InitialConditions(
-        grid=grid,
-        ini_time=datetime(2021, 6, 29),
-        source={"path": fname, "name": "GLORYS"},
-        adjust_depth_for_sea_surface_height=True,
-        use_dask=use_dask,
-    )
-
-
-@pytest.fixture(scope="session")
 def initial_conditions_with_bgc(use_dask: bool) -> InitialConditions:
     """Fixture for creating an InitialConditions object."""
     grid = Grid(
@@ -392,36 +364,6 @@ def initial_conditions_with_bgc(use_dask: bool) -> InitialConditions:
         ini_time=datetime(2021, 6, 29),
         source={"path": fname, "name": "GLORYS"},
         bgc_source={"path": fname_bgc, "name": "CESM_REGRIDDED"},
-        use_dask=use_dask,
-    )
-
-
-@pytest.fixture(scope="session")
-def initial_conditions_with_bgc_adjusted_for_zeta(use_dask: bool) -> InitialConditions:
-    """Fixture for creating an InitialConditions object."""
-    grid = Grid(
-        nx=2,
-        ny=2,
-        size_x=500,
-        size_y=1000,
-        center_lon=0,
-        center_lat=55,
-        rot=10,
-        N=3,  # number of vertical levels
-        theta_s=5.0,  # surface control parameter
-        theta_b=2.0,  # bottom control parameter
-        hc=250.0,  # critical depth
-    )
-
-    fname = Path(download_test_data("GLORYS_coarse_test_data.nc"))
-    fname_bgc = Path(download_test_data("CESM_regional_test_data_one_time_slice.nc"))
-
-    return InitialConditions(
-        grid=grid,
-        ini_time=datetime(2021, 6, 29),
-        source={"path": fname, "name": "GLORYS"},
-        bgc_source={"path": fname_bgc, "name": "CESM_REGRIDDED"},
-        adjust_depth_for_sea_surface_height=True,
         use_dask=use_dask,
     )
 
@@ -505,24 +447,6 @@ def boundary_forcing(use_dask: bool, small_grid: Grid) -> BoundaryForcing:
 
 
 @pytest.fixture(scope="session")
-def boundary_forcing_adjusted_for_zeta(
-    use_dask: bool, small_grid: Grid
-) -> BoundaryForcing:
-    """Fixture for creating a BoundaryForcing object."""
-    fname1 = Path(download_test_data("GLORYS_NA_20120101.nc"))
-    fname2 = Path(download_test_data("GLORYS_NA_20121231.nc"))
-    return BoundaryForcing(
-        grid=small_grid,
-        start_time=datetime(2012, 1, 1),
-        end_time=datetime(2012, 12, 31),
-        source={"name": "GLORYS", "path": [fname1, fname2]},
-        apply_2d_horizontal_fill=False,
[…]
