+++ roms_tools/tests/test_setup/test_initial_conditions.py
@@ -1,4 +1,3 @@
-import logging
 import textwrap
 from datetime import datetime
 from pathlib import Path
@@ -46,9 +45,7 @@ def example_grid():
     "ic_fixture",
     [
         "initial_conditions",
-        "initial_conditions_adjusted_for_zeta",
         "initial_conditions_with_bgc",
-        "initial_conditions_with_bgc_adjusted_for_zeta",
         "initial_conditions_with_bgc_from_climatology",
         "initial_conditions_with_unified_bgc_from_climatology",
     ],
@@ -66,6 +63,7 @@ def test_initial_conditions_creation_with_nondefault_glorys_dataset(
         "climatology": False,
     }
     assert hasattr(ic.ds, "adjust_depth_for_sea_surface_height")
+    assert ic.ds.attrs["adjust_depth_for_sea_surface_height"] == "False"
     assert isinstance(ic.ds, xr.Dataset)
     assert ic.ds.coords["ocean_time"].attrs["units"] == "seconds"
     expected_vars = {"temp", "salt", "u", "v", "zeta", "ubar", "vbar"}
@@ -171,7 +169,6 @@ def test_initial_conditions_creation_with_duplicates(use_dask: bool) -> None:
     "ic_fixture",
     [
         "initial_conditions_with_bgc",
-        "initial_conditions_with_bgc_adjusted_for_zeta",
         "initial_conditions_with_bgc_from_climatology",
         "initial_conditions_with_unified_bgc_from_climatology",
     ],
@@ -303,90 +300,6 @@ def test_initial_conditions_default_bgc_climatology(example_grid, use_dask):
     assert initial_conditions.bgc_source["climatology"] is False
 
 
-def test_info_depth(caplog, use_dask):
-    grid = Grid(
-        nx=2,
-        ny=2,
-        size_x=500,
-        size_y=1000,
-        center_lon=0,
-        center_lat=55,
-        rot=10,
-        N=3,  # number of vertical levels
-        theta_s=5.0,  # surface control parameter
-        theta_b=2.0,  # bottom control parameter
-        hc=250.0,  # critical depth
-    )
-
-    fname = Path(download_test_data("GLORYS_coarse_test_data.nc"))
-
-    with caplog.at_level(logging.INFO):
-        InitialConditions(
-            grid=grid,
-            ini_time=datetime(2021, 6, 29),
-            source={"path": fname, "name": "GLORYS"},
-            adjust_depth_for_sea_surface_height=True,
-            use_dask=use_dask,
-        )
-    # Verify the warning message in the log
-    assert "Sea surface height will be used to adjust depth coordinates." in caplog.text
-
-    # Clear the log before the next test
-    caplog.clear()
-
-    with caplog.at_level(logging.INFO):
-        InitialConditions(
-            grid=grid,
-            ini_time=datetime(2021, 6, 29),
-            source={"path": fname, "name": "GLORYS"},
-            adjust_depth_for_sea_surface_height=False,
-            use_dask=use_dask,
-        )
-    # Verify the warning message in the log
-    assert (
-        "Sea surface height will NOT be used to adjust depth coordinates."
-        in caplog.text
-    )
-
-
-@pytest.mark.parametrize(
-    "initial_conditions_fixture",
-    [
-        "initial_conditions_adjusted_for_zeta",
-        "initial_conditions_with_bgc_adjusted_for_zeta",
-    ],
-)
-def test_correct_depth_coords_adjusted_for_zeta(
-    initial_conditions_fixture, request, use_dask
-):
-    initial_conditions = request.getfixturevalue(initial_conditions_fixture)
-
-    # compute interface depth at rho-points and write it into .ds_depth_coords
[…]

[…]
