### 2025-12-20 ab57c38
Add to release notes
