### 2025-12-22 36add1d
Don't modify original grid
