### 2025-12-22 d10180a
Also test IC from ROMS without BGC
