+++ conftest.py
@@ -435,7 +435,7 @@ def initial_conditions_with_unified_bgc_from_climatology(
 def initial_conditions_from_roms(
     use_dask: bool,
 ) -> InitialConditions:
-    grid = Grid(nx=5, ny=5, center_lon=-120, center_lat=34, size_x=100, size_y=100)
+    grid = Grid(nx=5, ny=5, center_lon=-120, center_lat=34, size_x=100, size_y=100, N=3)
 
     parent_grid = Grid(
         center_lon=-120, center_lat=30, nx=8, ny=13, size_x=3000, size_y=4000, rot=32
@@ -447,6 +447,25 @@ def initial_conditions_from_roms(
         ini_time=datetime(1998, 1, 6),
         source={"name": "ROMS", "path": fname_restart, "grid": parent_grid},  # type: ignore
         bgc_source={"name": "ROMS", "path": fname_restart, "grid": parent_grid},  # type: ignore
+        use_dask=use_dask,
+    )
+
+
+@pytest.fixture(scope="session")
+def initial_conditions_from_roms_without_bgc(
+    use_dask: bool,
+) -> InitialConditions:
+    grid = Grid(nx=5, ny=5, center_lon=-120, center_lat=34, size_x=100, size_y=100, N=3)
+
+    parent_grid = Grid(
+        center_lon=-120, center_lat=30, nx=8, ny=13, size_x=3000, size_y=4000, rot=32
+    )
+    fname_restart = Path(download_test_data("eastpac25km_rst.19980106000000.nc"))
+
+    return InitialConditions(
+        grid=grid,
+        ini_time=datetime(1998, 1, 6),
+        source={"name": "ROMS", "path": fname_restart, "grid": parent_grid},  # type: ignore
         use_dask=use_dask,
     )
 
