### 2025-12-22 e12edc7
Adjust for SSH for IC from ROMS
