### 2025-12-22 61f2735
Remove user-facing control of SSH-adjusted depth coordinates (#514)
* Don't expose adjust_depth_for_sea_surface_height to user

* Remove adjust_depth_for_sea_surface_height from docs

* Add to release notes
