### 2025-12-22 6efbed0
Make choose_subdomain consistent
