### 2025-12-22 301603c
Subset all layer depth at once
