### 2025-12-22 e11bb26
Finally found the bug!
