### 2025-12-22 f4912bc
Bump default nr buffer points
