### 2025-12-31 c060cfb
Implement safety checks for dimension order
