### 2026-01-02 8cf05db
Don't require s_rho dim
