### 2026-01-02 72c4a96
Pass dims as tuple
