### 2026-01-02 2d63d4c
Correct dim check
