+++ roms_tools/tests/test_fill.py
@@ -19,12 +19,12 @@ def test_lateral_fill_no_nans(data_fixture, request):
     data = request.getfixturevalue(data_fixture)
     lateral_fill = LateralFill(
         data.ds["mask"],
-        [data.dim_names["latitude"], data.dim_names["longitude"]],
+        (data.dim_names["latitude"], data.dim_names["longitude"]),
     )
     if "mask_vel" in data.ds.data_vars:
         lateral_fill_vel = LateralFill(
             data.ds["mask_vel"],
-            [data.dim_names["latitude"], data.dim_names["longitude"]],
+            (data.dim_names["latitude"], data.dim_names["longitude"]),
         )
 
     for var in data.var_names:
@@ -40,10 +40,10 @@ def test_lateral_fill_no_nans(data_fixture, request):
 def test_lateral_fill_correct_order_of_magnitude(coarsened_cesm_bgc_data):
     lateral_fill = LateralFill(
         coarsened_cesm_bgc_data.ds["mask"],
-        [
+        (
             coarsened_cesm_bgc_data.dim_names["latitude"],
             coarsened_cesm_bgc_data.dim_names["longitude"],
-        ],
+        ),
     )
 
     ALK = coarsened_cesm_bgc_data.ds["ALK"]
@@ -78,21 +78,21 @@ def test_lateral_fill_reproducibility(data_fixture, request):
     data = request.getfixturevalue(data_fixture)
     lateral_fill0 = LateralFill(
         data.ds["mask"],
-        [data.dim_names["latitude"], data.dim_names["longitude"]],
+        (data.dim_names["latitude"], data.dim_names["longitude"]),
     )
     if "mask_vel" in data.ds.data_vars:
         lateral_fill_vel0 = LateralFill(
             data.ds["mask_vel"],
-            [data.dim_names["latitude"], data.dim_names["longitude"]],
+            (data.dim_names["latitude"], data.dim_names["longitude"]),
         )
     lateral_fill1 = LateralFill(
         data.ds["mask"],
-        [data.dim_names["latitude"], data.dim_names["longitude"]],
+        (data.dim_names["latitude"], data.dim_names["longitude"]),
     )
     if "mask_vel" in data.ds.data_vars:
         lateral_fill_vel1 = LateralFill(
             data.ds["mask_vel"],
-            [data.dim_names["latitude"], data.dim_names["longitude"]],
+            (data.dim_names["latitude"], data.dim_names["longitude"]),
         )
 
     ds0 = data.ds.copy()
@@ -119,47 +119,47 @@ def test_lateral_fill_reproducibility(data_fixture, request):
 
 
 def test_lateralfill_raises_notimplemented_for_non2d_mask():
-    mask_3d = np.ones((3, 3, 3), dtype=bool)
+    mask_3d = xr.DataArray(np.ones((3, 3, 3)), dims=("eta_rho", "xi_rho", "s_rho"))
     with pytest.raises(
         NotImplementedError, match="LateralFill currently supports only 2D masks"
     ):
-        LateralFill(mask_3d, dims=["eta_rho", "xi_rho"])
+        LateralFill(mask_3d, dims=("eta_rho", "xi_rho"))
+
+
+def test_lateralfill_raises_valueerror_for_list_of_dims():
+    mask = xr.DataArray(np.ones((3, 3)), dims=("eta_rho", "xi_rho"))
+    with pytest.raises(TypeError, match="must be a tuple"):
+        LateralFill(mask, dims=["eta_rho", "xi_rho"])
 
 
[…]
