### 2026-01-02 5472174
Generic function for interpolation on staggered grids
