### 2026-01-02 fbc7c09
Move and expand rotate_velocities
