+++ roms_tools/setup/surface_forcing.py
@@ -28,13 +28,13 @@ from roms_tools.setup.utils import (
     group_dataset,
     min_dist_to_land,
     nan_check,
-    rotate_velocities,
     substitute_nans_by_fillvalue,
     to_dict,
     write_to_yaml,
 )
 from roms_tools.utils import (
     interpolate_from_climatology,
+    rotate_velocities,
     save_datasets,
     transpose_dimensions,
 )
