### 2026-01-05 418a647
Test rotate velocities roundtrip
