### 2026-01-05 d8c3855
Add rotation method
