### 2026-01-05 245746a
LateralFill: enforce correct dimension order for mask and input arrays (#517)
* Implement safety checks for dimension order

* Preserve dimension order when filling ROMSDataset

* Don't require s_rho dim

* Pass dims as tuple

* Correct dim check

* More detailed error message

* Add to release notes
