### 2026-01-05 04ab007
Finalize IC from ROMS
