+++ roms_tools/setup/initial_conditions.py
@@ -198,6 +198,7 @@ class InitialConditions:
         target_coords = get_target_coords(self.grid)
 
         data = self._get_data(forcing_type=type)
+        data.rotate_velocities_to_east_and_north()
 
         data.choose_subdomain(
             target_coords,
@@ -206,7 +207,6 @@ class InitialConditions:
         data.convert_to_float64()
         data.extrapolate_deepest_to_bottom()
         data.apply_lateral_fill()
-        data.rotate_velocities_to_east_and_north()
 
         self._set_variable_info(data, type=type)
         attr_name = f"variable_info_{type}"
@@ -249,7 +249,7 @@ class InitialConditions:
                 processed_fields["u"],
                 processed_fields["v"],
                 target_coords["angle"],
-                interpolate=True,
+                interpolate_after=True,
             )
 
             if "layer_depth_u" in processed_fields:
@@ -289,95 +289,153 @@ class InitialConditions:
 
         return processed_fields
 
-    def _regrid_laterally(self, data, target_coords, processed_fields, var_names):
-        """Regrid variables in data.ds laterally to rho points of ROMS grid."""
-        if isinstance(data, ROMSDataset):
-            # source data can be at three different locations: rho-, u-, v- points
-            active_locations = {info["location"] for info in var_names.values()}
+    def _regrid_laterally(
+        self,
+        data: ROMSDataset | LatLonDataset,
+        target_coords: dict[str, xr.DataArray],
+        processed_fields: dict[str, xr.DataArray],
+        var_names: dict[str, dict[str, str]],
+    ):
+        """Regrid variables in data.ds laterally to target coordinates.
 
-            # compute depth coordinates on source data and subset consistent with subsetting of data
-            for location in active_locations:
-                data._get_depth_coordinates(depth_type="layer", locations=[location])
+        Parameters
+        ----------
+        data : ROMSDataset or LatLonDataset
+            The dataset containing variables to regrid.
+        target_coords : dict[str, xr.DataArray]
+            Dictionary of target coordinates for regridding.
+        processed_fields : dict[str, xr.DataArray]
+            Dictionary where regridded variables will be stored.
+        var_names : dict[str, dict[str, str]]
+            Mapping from variable keys to dataset variable names and metadata.
+
+        Returns
+        -------
+        processed_fields : dict[str, xr.DataArray]
+            Updated dictionary with regridded variables.
+        """
+        if isinstance(data, ROMSDataset):
+            # Compute depth coordinates on source data for rho
+            data._get_depth_coordinates(depth_type="layer", locations=["rho"])
+            # Subset depth coordinate to target subdomain
             data.ds_depth_coords = choose_subdomain(
                 data.ds_depth_coords,
                 data.grid.ds,
                 target_coords,
             )
 
-            for location in ["rho", "u", "v"]:
-                filtered_vars = [
-                    var_name
-                    for var_name, info in var_names.items()
-                    if info["location"] == location
-                ]
-                if filtered_vars:
-                    ds_loc = data.ds[filtered_vars].rename(
-                        {f"lat_{location}": "lat", f"lon_{location}": "lon"}
-                    )
-                    lateral_regrid = LateralRegridFromROMS(ds_loc, target_coords)
-
-                    ds_loc = lateral_regrid.apply(ds_loc)
-                    for var_name in filtered_vars:
-                        processed_fields[var_name] = ds_loc[var_name]
-
-                    # also regrid depth coordinates
-                    processed_fields[f"layer_depth_{location}"] = lateral_regrid.apply(
-                        data.ds_depth_coords[f"layer_depth_{location}"]
-                    )
-
-        else:
-            # all variables are at the same location
[…]
