### 2026-01-05 ce2fda8
Update methods
