+++ docs/methods.md
@@ -123,16 +123,20 @@ Here, $\zeta(x,y,t)$ is the time-varying sea surface height, and $h(x,y)$ is the
 
 Example plots of the vertical stretching functions and layer depths can be found [here](grid.ipynb#The-vertical-coordinate-system).
 
-`ROMS-Tools` executes steps 1 and 2 during the grid generation process, when an instance of the {py:obj}`roms_tools.Grid` class is created. Step 3 is executed when needed, during the creation of the initial conditions and boundary forcing.
+`ROMS-Tools` executes steps 1 and 2 during grid generation, when an instance of the :py:class:`roms_tools.Grid` class is created.
+Step 3 is executed only when needed, during the creation of initial conditions and boundary forcing.
+Depending on the dataset, either the actual sea surface height or a flat sea surface height (:math:`\zeta(x,y,t)=0`) is used.
 
-During step 3, `ROMS-Tools` handles sea surface height (SSH) based on the value of `adjust_depth_for_sea_surface_height`:
+**Actual sea surface height** (:math:`\zeta \neq 0`), consistent with measuring elevation above the bottom:
 
-* If `adjust_depth_for_sea_surface_height = False` (default), a constant sea surface height is assumed: $\zeta(x,y,t) = 0$.
-* If `adjust_depth_for_sea_surface_height = True`, the depth calculations account for spatial and temporal variations in SSH.
+- Initial conditions derived from an outer-nest ROMS simulation (used for both target and source vertical coordinates).
 
-Setting `adjust_depth_for_sea_surface_height = True` increases the complexity in the depth coordinate generation, as depths become time-dependent. While this approach may seem more accurate, its effectiveness depends on how source datasets (e.g., GLORYS) define their depth coordinates. It is often unclear whether these datasets measure depth relative to the surface (consistent with `adjust_depth_for_sea_surface_height = False`) or elevation above the bottom (consistent with `adjust_depth_for_sea_surface_height = True`). Given this uncertainty and the unclear benefit of adjusting for sea surface height variations, the default setting is `adjust_depth_for_sea_surface_height = False` for simplicity.
+**Flat sea surface height** (:math:`\zeta = 0`), consistent with measuring depth relative to the surface:
 
+- Initial and boundary conditions derived from external datasets such as GLORYS or BGC data from hybrid model-obs sources.
 
+For most external datasets, it is unclear whether depth values are relative to the surface or above the bottom, and not all datasets include sea surface height. To avoid inconsistencies, `ROMS-Tools` defaults to a flat sea surface (:math:`\zeta=0`)
+when using external data.
 
 ## Tidal Forcing
 
@@ -164,9 +168,9 @@ For practical examples, see [this notebook](tides.ipynb).
 
 ## Surface Forcing
 
-The surface forcing data is sourced from **ERA5** (for meteorological forcing) and **CESM** (for biogeochemical surface forcing). The data is regridded onto the ROMS grid via the following steps:
+The surface forcing data is sourced from **ERA5** (for meteorological forcing) and **CESM** or observational/model-based sources (for biogeochemical surface forcing). The data is regridded onto the ROMS grid via the following steps:
 
-1. **Horizontal Land Fill**: Ocean values are extended into land areas using a horizontal fill process based on a [multigrid method](#multigrid-method-for-filling-land-values). This step is crucial because the ERA5/CESM and ROMS grids may have differing land masks, especially when their resolutions differ. Without applying the horizontal fill, land mask discrepancies could result in skewed values at certain ocean points in the ROMS grid that the ERA5/CESM data considers land. Surface forcing over land differs significantly from surface forcing over the ocean, making the horizontal land fill essential.
[…]
