### 2026-01-05 68a7505
Reorder operations
