### 2026-01-06 242ec85
Add pytest marker
