### 2026-01-06 857304b
Fix methods
