### 2026-01-06 e0703e7
Remove duplicate code
