### 2026-01-07 ddc66f0
Add to release notes
