### 2026-01-07 215fe40
Add eta_u and xi_v to partition dims
