### 2026-01-07 5989492
Test partitioning of eta_u, xi_v
