### 2026-01-07 7556bfc
Refactor partition function
