+++ roms_tools/tiling/partition.py
@@ -7,6 +7,61 @@ import xarray as xr
 
 from roms_tools.utils import save_datasets
 
+DIM_INFO = {
+    # eta-direction
+    "eta_rho": dict(axis="eta", ghost=2, edge="both"),
+    "eta_u": dict(axis="eta", ghost=2, edge="both"),
+    "eta_v": dict(axis="eta", ghost=1, edge="upper"),
+    "eta_psi": dict(axis="eta", ghost=3, edge="both"),
+    "eta_coarse": dict(axis="eta", ghost=2, edge="both"),
+    # xi-direction
+    "xi_rho": dict(axis="xi", ghost=2, edge="both"),
+    "xi_u": dict(axis="xi", ghost=1, edge="upper"),
+    "xi_v": dict(axis="xi", ghost=2, edge="both"),
+    "xi_psi": dict(axis="xi", ghost=3, edge="both"),
+    "xi_coarse": dict(axis="xi", ghost=2, edge="both"),
+}
+
+
+def _exact_division(size: int, nparts: int, dim: str) -> int:
+    if size % nparts != 0:
+        raise ValueError(
+            f"Dimension '{dim}' of size {size} cannot be evenly divided into {nparts} partitions."
+        )
+    return size // nparts
+
+
+def _compute_partition_sizes(
+    total_size: int,
+    nparts: int,
+    ghost: int,
+    edge: str,
+    dim: str,
+) -> list[int]:
+    """Compute per-tile sizes including ghost cells."""
+    if nparts == 1:
+        return [total_size]
+
+    core = _exact_division(total_size - ghost, nparts, dim)
+
+    sizes = [core] * nparts
+
+    if edge == "both":
+        sizes[0] += ghost // 2
+        sizes[-1] += ghost - ghost // 2
+    elif edge == "upper":
+        sizes[-1] += ghost
+    else:
+        raise ValueError(f"Unknown edge rule '{edge}' for {dim}")
+
+    return sizes
+
+
+def _cumsum(sizes: list[int]) -> np.ndarray:
+    out = np.zeros(len(sizes) + 1, dtype=int)
+    out[1:] = np.cumsum(sizes)
+    return out
+
 
 def partition(
     ds: xr.Dataset, np_eta: int = 1, np_xi: int = 1, include_coarse_dims: bool = True
@@ -71,250 +126,42 @@ def partition(
     ):
         raise ValueError("np_eta and np_xi must be positive integers")
 
-    base_dims = [
-        "eta_rho",
-        "eta_u",
-        "eta_v",
-        "eta_psi",
-        "xi_rho",
-        "xi_u",
-        "xi_v",
-        "xi_psi",
+    # Select applicable dimensions
+    dims_to_partition = [
+        d
+        for d in DIM_INFO
+        if d in ds.dims and (include_coarse_dims or "coarse" not in d)
     ]
-    coarse_dims = ["eta_coarse", "xi_coarse"]
-    partitionable_dims = base_dims + coarse_dims if include_coarse_dims else base_dims
 
-    dims_to_partition = [d for d in partitionable_dims if d in ds.dims]
+    partitioned_sizes = {}
 
-    n_eta_ghost_cells = 1
-    n_xi_ghost_cells = 1
+    for dim in dims_to_partition:
+        info = DIM_INFO[dim]
+        nparts = np_eta if info["axis"] == "eta" else np_xi
 
-    def integer_division_or_raise(a: int, b: int, dimension: str) -> int:
-        """Perform integer division and ensure that the division is exact.
-
-        Parameters
-        ----------
-        a : int
-            The numerator for the division.
-        b : int
-            The denominator for the division.
-        dimension : str
-            The name of the dimension being partitioned, used for error reporting.
-
-        Returns
[…]
