### 2026-01-07 612041b
Make sure dicts are compared
