### 2026-01-08 4a6fa3b
Refactor rotate_velocities and staggered interpolation
