### 2026-01-08 4b0228d
Add rotate_velocities_to_east_and_north methods to Dataset
