### 2026-01-08 55f4b2c
Add to release notes
