### 2026-01-08 046db59
Refactor and correct regrid method
