+++ roms_tools/analysis/roms_output.py
@@ -1,4 +1,6 @@
+from collections.abc import Sequence
 from dataclasses import dataclass
+from typing import Final
 
 import numpy as np
 import xarray as xr
@@ -11,6 +13,7 @@ from roms_tools.regrid import LateralRegridFromROMS, VerticalRegridFromROMS
 from roms_tools.utils import (
     generate_coordinate_range,
     infer_nominal_horizontal_resolution,
+    rotate_velocities,
 )
 
 
@@ -212,90 +215,144 @@ class ROMSOutput(ROMSDataset):
             cmap_name=cmap_name,
         )
 
-    def regrid(self, var_names=None, horizontal_resolution=None, depth_levels=None):
+    def regrid(
+        self,
+        var_names: Sequence[str] | None = None,
+        horizontal_resolution: float | None = None,
+        depth_levels: xr.DataArray | np.ndarray | Sequence[float] | None = None,
+    ) -> xr.Dataset:
         """Regrid the dataset both horizontally and vertically.
 
-        This method selects the specified variables, interpolates them onto a lat-lon-z horizontal grid. The horizontal target resolution and vertical target depth levels are either specified or inferred dynamically.
+        This method selects the specified variables, interpolates them onto a
+        lat-lon-depth grid, and rotates velocity components to east/north
+        directions when required.
 
         Parameters
         ----------
-        var_names : list of str, optional
-            List of variable names to be regridded. If None, all variables in the dataset
-            are used.
+        var_names : sequence of str, optional
+            Variables to regrid. If None, all variables in the dataset are used.
         horizontal_resolution : float, optional
             Target horizontal resolution in degrees. If None, the nominal horizontal resolution is inferred from the grid.
-        depth_levels : xarray.DataArray, numpy.ndarray, list, optional
+        depth_levels : xarray.DataArray, numpy.ndarray, sequence, optional
             Target depth levels. If None, depth levels are determined dynamically.
-            If provided as a list or numpy array, it is safely converted to an `xarray.DataArray`.
 
         Returns
         -------
-        xarray.Dataset
+        xr.Dataset
             The regridded dataset.
         """
-        if var_names is None:
-            var_names = list(self.ds.data_vars)
+        VELOCITY_PAIRS: Final[list[tuple[str, str]]] = [
+            ("u", "v"),
+            ("ubar", "vbar"),
+            ("u_slow", "v_slow"),
+        ]
 
-        # Check that all var_names exist in self.ds
-        missing_vars = [var for var in var_names if var not in self.ds.data_vars]
-        if missing_vars:
-            raise ValueError(
-                f"The following variables are not found in the dataset: {', '.join(missing_vars)}"
+        def _select_variables(var_names: Sequence[str] | None) -> xr.Dataset:
+            requested: set[str] = set(var_names or self.ds.data_vars)
+
+            # bidirectional map
+            DEPENDENCY_MAP: Final[dict[str, str]] = {k: v for k, v in VELOCITY_PAIRS}
+            DEPENDENCY_MAP.update({v: k for k, v in VELOCITY_PAIRS})
+
+            required: set[str] = set(requested)
+            for var in requested:
+                if var in DEPENDENCY_MAP:
+                    required.add(DEPENDENCY_MAP[var])
+
+            missing = required - set(self.ds.data_vars)
+            if missing:
+                raise ValueError(f"Variables not found: {', '.join(sorted(missing))}")
+
+            return self.ds[list(required)].copy()
+
+        def _prepare_horizontal_grid() -> dict[str, xr.DataArray]:
+            lat = self.grid.ds["lat_rho"]
[…]
