### 2026-01-09 a9ea8ba
Provide depth to VerticalRegrid as xarray.DataArray
