### 2026-01-09 3a1d9d3
Introduce source_dim parameter
