### 2026-01-09 3a6c7ed
Test VerticalRegrid for depth -> ROMS
