### 2026-01-12 c7c0aad
New vertical regridding for initial conditions
