+++ roms_tools/setup/initial_conditions.py
@@ -18,7 +18,7 @@ from roms_tools.datasets.lat_lon_datasets import (
     UnifiedBGCDataset,
 )
 from roms_tools.plot import plot
-from roms_tools.regrid import LateralRegridToROMS, VerticalRegridToROMS
+from roms_tools.regrid import LateralRegridToROMS, VerticalRegrid
 from roms_tools.setup.utils import (
     RawDataSource,
     compute_barotropic_velocity,
@@ -91,10 +91,6 @@ class InitialConditions:
         - If True: allows a +24h search window after `ini_time` and selects the closest available
           time entry within that window. Raises a ValueError if none are found.
 
-    horizontal_chunk_size : int, optional
-        The chunk size used for horizontal partitioning for the vertical regridding when `use_dask = True`. Defaults to 50.
-        A larger number results in a bigger memory footprint but faster computations.
-        A smaller number results in a smaller memory footprint but slower computations.
     bypass_validation: bool, optional
         Indicates whether to skip validation checks in the processed data. When set to True,
         the validation process that ensures no NaN values exist at wet points
@@ -129,9 +125,6 @@ class InitialConditions:
     """Whether to handle ini_time flexibly."""
     use_dask: bool = False
     """Whether to use dask for processing."""
-    horizontal_chunk_size: int = 50
-    """The chunk size used for horizontal partitioning for the vertical regridding when
-    `use_dask = True`."""
     bypass_validation: bool = False
     """Whether to skip validation checks in the processed data."""
 
@@ -259,19 +252,18 @@ class InitialConditions:
                 # Handle depth coordinates
                 self._get_depth_coordinates(zeta, location, "layer")
 
-                # Vertical regridding
-                vertical_regrid = VerticalRegridToROMS(
-                    self.ds_depth_coords[f"layer_depth_{location}"],
-                    data.ds[data.dim_names["depth"]],
+                vertical_regrid = VerticalRegrid(
+                    data.ds, source_dim=data.dim_names["depth"]
                 )
                 for var_name in filtered_vars:
                     if var_name in processed_fields:
-                        field = processed_fields[var_name]
-                        if self.use_dask:
-                            field = field.chunk(
-                                _set_dask_chunks(location, self.horizontal_chunk_size)
-                            )
-                        processed_fields[var_name] = vertical_regrid.apply(field)
+                        processed_fields[var_name] = vertical_regrid.apply(
+                            processed_fields[var_name],
+                            source_depth_coords=data.ds[data.dim_names["depth"]],
+                            target_depth_coords=self.ds_depth_coords[
+                                f"layer_depth_{location}"
+                            ],
+                        )
 
         # Compute barotropic velocities
         if "u" in var_names and "v" in var_names:
@@ -517,12 +509,6 @@ class InitialConditions:
                 if isinstance(zeta, xr.DataArray):
                     zeta = interpolate_from_rho_to_v(zeta)
 
-            if self.use_dask:
-                h = h.chunk(_set_dask_chunks(location, self.horizontal_chunk_size))
-                if isinstance(zeta, xr.DataArray):
-                    zeta = zeta.chunk(
-                        _set_dask_chunks(location, self.horizontal_chunk_size)
-                    )
             depth = compute_depth(zeta, h, self.grid.ds.attrs["hc"], Cs, sigma)
             self.ds_depth_coords[key] = depth
 
