+++ roms_tools/regrid.py
@@ -134,100 +134,6 @@ class LateralRegridFromROMS:
         return regridded
 
 
-class VerticalRegridToROMS:
-    """Interpolates data onto new vertical (depth) coordinates.
-
-    Parameters
-    ----------
-    target_depth_coords : xarray.DataArray
-        Depth coordinates for the target grid.
-    source_depth_coords : xarray.DataArray
-        Depth coordinates for the source grid.
-    """
-
-    def __init__(self, target_depth_coords, source_depth_coords):
-        """Initialize regridding factors for interpolation.
-
-        Parameters
-        ----------
-        target_depth_coords : xarray.DataArray
-            Depth coordinates for the target grid.
-        source_depth_coords : xarray.DataArray
-            Depth coordinates for the source grid.
-
-        Attributes
-        ----------
-        coeff : xarray.Dataset
-            Dataset containing:
-            - `is_below` : Boolean mask for depths just below target.
-            - `is_above` : Boolean mask for depths just above target.
-            - `upper_mask`, `lower_mask` : Masks for valid interpolation bounds.
-            - `factor` : Weight for blending values between levels.
-        """
-        self.depth_dim = source_depth_coords.dims[0]
-        source_depth = source_depth_coords
-        dims = {"dim": self.depth_dim}
-
-        dlev = source_depth - target_depth_coords
-        is_below = dlev == dlev.where(dlev >= 0).min(**dims)
-        is_above = dlev == dlev.where(dlev <= 0).max(**dims)
-        p_below = dlev.where(is_below).sum(**dims)
-        p_above = -dlev.where(is_above).sum(**dims)
-        denominator = p_below + p_above
-        denominator = denominator.where(denominator > 1e-6, 1e-6)
-        factor = p_below / denominator
-
-        upper_mask = is_above.sum(**dims) > 0
-        lower_mask = is_below.sum(**dims) > 0
-
-        self.coeff = xr.Dataset(
-            {
-                "is_below": is_below,
-                "is_above": is_above,
-                "upper_mask": upper_mask,
-                "lower_mask": lower_mask,
-                "factor": factor,
-            }
-        )
-
-    def apply(self, da, fill_nans=True):
-        """Interpolates the variable onto the new depth grid using precomputed
-        coefficients for linear interpolation between layers.
-
-        Parameters
-        ----------
-        da : xarray.DataArray
-            The input data to be regridded along the depth dimension. This should be
-            an array with the same depth coordinates as the original grid.
-        fill_nans : bool, optional
-            Whether to fill NaN values in the regridded data. If True (default),
-            forward-fill and backward-fill are applied along the 's_rho' dimension to
-            ensure there are no NaNs after interpolation.
-
-        Returns
-        -------
-        xarray.DataArray
-            The regridded data array, interpolated onto the new depth grid. NaN values
-            are replaced if `fill_nans=True`, with extrapolation allowed at the surface
-            and bottom layers to minimize gaps.
-        """
-        dims = {"dim": self.depth_dim}
-
-        da_below = da.where(self.coeff["is_below"]).sum(**dims)
-        da_above = da.where(self.coeff["is_above"]).sum(**dims)
-
-        result = da_below + (da_above - da_below) * self.coeff["factor"]
-        if fill_nans:
-            result = result.where(self.coeff["upper_mask"], da.isel({dims["dim"]: 0}))
-            result = result.where(self.coeff["lower_mask"], da.isel({dims["dim"]: -1}))
-        else:
-            result = result.where(self.coeff["upper_mask"]).where(
-                self.coeff["lower_mask"]
-            )
-
-        return result
-
-
 class VerticalRegrid:
     """Regrid ROMS variables along the vertical.
 
