### 2026-01-14 d22d5d7
Improved structure of landing page
