+++ docs/index.rst
@@ -43,26 +43,29 @@ Built on ``xarray`` and optionally powered by ``dask``, ``ROMS-Tools`` automates
 10. **CDR Forcing**: User-defined interventions that inject BGC tracers at point sources or as larger-scale Gaussian perturbations, designed to simulate CDR interventions.
 11. **Nesting**: Support for creating nested grids and parent-child configurations.
 
-Beyond input generation, ``ROMS-Tools`` provides a suite of analysis and postprocessing utilities, including regridding fields to standard latitude-longitude-depth grids and performing specialized CDR-focused analysis.
+Beyond input generation, ``ROMS-Tools`` provides a suite of **analysis and postprocessing utilities**, including merging tiled outputs, regridding fields to standard latitude-longitude-depth grids, visualization, and specialized CDR-focused analysis.
 
-``ROMS-Tools`` supports modern, reproducible workflows with YAML-based configuration, cloud-accessible datasets, optional ``dask`` parallelization, interactive Jupyter usage, and CI-tested reliability with comprehensive documentation.
+``ROMS-Tools`` supports **modern, reproducible workflows** with YAML-based configuration, cloud-accessible datasets, optional ``dask`` parallelization, interactive Jupyter usage, and CI-tested reliability with comprehensive documentation.
 
-.. note::
-  Below is a series of examples. The `end-to-end workflow <https://roms-tools.readthedocs.io/en/latest/end_to_end.html>`_ demonstrates using ``ROMS-Tools`` to prepare inputs and analyze outputs for a ROMS–MARBL simulation, and is designed to run on a laptop. Subsequent examples cover more detailed, task-specific workflows and are configured for the Perlmutter supercomputing system, so you may need to pre-download data and adjust paths when running them elsewhere.
-
+Getting Started
+---------------------------
 
 .. toctree::
    :maxdepth: 1
-   :caption: Getting Started
 
    installation
    methods
    datasets
-   end_to_end
+   End-to-end workflow (laptop) <end_to_end>
+
+
+Preparing a ROMS Simulation
+---------------------------
+
+The following examples cover task-specific workflows for the Perlmutter supercomputer. Pre-download data and adjust paths as needed. For a full end-to-end workflow designed to run on a laptop, refer to the “End-to-end workflow (laptop)” section above.
 
 .. toctree::
    :maxdepth: 1
-   :caption: Preparing a ROMS Simulation
 
    Creating a grid <grid>
    Creating tidal forcing <tides>
@@ -74,9 +77,13 @@ Beyond input generation, ``ROMS-Tools`` provides a suite of analysis and postpro
    Preparing nested simulations <nesting>
    Partitioning the input files <partition>
 
+Analyzing a ROMS Simulation
+---------------------------
+
+The following examples cover task-specific workflows for the Perlmutter supercomputer. Pre-download data and adjust paths as needed. For a full end-to-end workflow designed to run on a laptop, refer to the “End-to-end workflow (laptop)” section above.
+
 .. toctree::
    :maxdepth: 1
-   :caption: Analyzing a ROMS Simulation
 
    Joining ROMS output files <join>
    Reading ROMS Output <reading_roms_output>
@@ -84,23 +91,27 @@ Beyond input generation, ``ROMS-Tools`` provides a suite of analysis and postpro
    Regridding ROMS Output <regridding_roms_output>
    Analyzing Carbon Dioxide Removal (CDR) Metrics <cdr_analysis>
 
+
+Advanced Topics
+---------------------------
 .. toctree::
    :maxdepth: 1
-   :caption: Advanced Topics
 
    Using ROMS-Tools with Dask <using_dask>
 
 
+For Developers
+---------------------------
 .. toctree::
[…]
