### 2026-01-15 51db92c
Error handling for missing hmin
