### 2026-01-15 2c1d815
Add test that checks error is raised correctly
