### 2026-01-15 a429db2
adding emod dataset as a class
