### 2026-01-15 cd38557
adding emod dataset as a class (#530)
