### 2026-01-15 398dae5
Improve robustness of update_topography for externally generated grids (#523)
* Raise update topography

* Add to release notes

* Error handling for missing hmin

* Add test that checks error is raised correctly
