+++ roms_tools/tests/test_setup/test_grid.py
@@ -741,6 +741,23 @@ def test_hmin_criterion_and_update_topography():
     assert np.less_equal(grid.hmin, grid.ds.h.min())
 
 
+def test_update_topography_raises_if_grid_loaded_from_file_has_no_source_info():
+    fname = download_test_data("grid_created_with_matlab.nc")
+    grid = Grid.from_file(fname)
+
+    with pytest.raises(
+        ValueError,
+        match="Topography source information is not available",
+    ):
+        grid.update_topography(hmin=15)
+
+    with pytest.raises(
+        ValueError,
+        match="Minimal ocean depth is not available",
+    ):
+        grid.update_topography(topography_source={"name": "ETOPO5"})
+
+
 # Mask tests
 
 
