### 2026-01-20 097e384
Add minimial EMOD example
