+++ roms_tools/setup/topography.py
@@ -7,7 +7,11 @@ import gcm_filters
 import numpy as np
 import xarray as xr
 
-from roms_tools.datasets.lat_lon_datasets import ETOPO5Dataset, SRTM15Dataset, EMODDataset
+from roms_tools.datasets.lat_lon_datasets import (
+    EMODDataset,
+    ETOPO5Dataset,
+    SRTM15Dataset,
+)
 from roms_tools.regrid import LateralRegridToROMS
 from roms_tools.setup.utils import handle_boundaries
 
