### 2026-01-20 6c8c7f5
Apply lateral fill to topo data
