### 2026-01-20 413c737
Assign mask to EMOD dataset
