### 2026-01-20 c56b2f9
Shorten paper and add new required sections
