+++ paper.md
@@ -1,5 +1,5 @@
 ---
-title: 'ROMS-Tools: A Python Package for Preparing and Analyzing ROMS Simulations'
+title: 'ROMS-Tools: Reproducible preprocessing and analysis for regional ocean modeling'
 tags:
   - Python
   - ocean modeling
@@ -54,20 +54,19 @@ affiliations:
    index: 3
 date: 3 November 2025
 bibliography: docs/references.bib
-
----
+—
 
 # Summary
 
-The ocean shapes Earth’s climate and sustains marine ecosystems by circulating and storing vast amounts of heat, oxygen, carbon, and nutrients, while exchanging heat and gases with the atmosphere. To understand these complex dynamics, scientists rely on ocean models, powerful computer simulations of physical circulation and biogeochemical (BGC) processes. These models represent the ocean on a grid of cells, where finer grid spacing (more, smaller cells) provides higher fidelity and greater detail at the cost of  more computing power. While global ocean models simulate the entire ocean, **regional ocean models** focus computational resources on a specific area to achieve much finer grid spacing than is computationally feasible over the global domain. This finer grid spacing enables regional ocean models to explicitly resolve fine-scale phenomena, like mesoscale (10-100 km) and submesoscale (0.1-10 km) features, tidal dynamics, coastal currents, upwelling, and detailed BGC processes. Capturing these dynamics and processes at high fidelity is essential for applications in environmental management, fisheries, for assessing regional impacts of climate change, and for evaluating ocean-based carbon dioxide removal (CDR) strategies.
+The ocean regulates Earth’s climate and sustains marine ecosystems by circulating and storing heat, carbon, oxygen, and nutrients, while exchanging gases with the atmosphere. Scientists study these processes using ocean models, which simulate the ocean on a grid.
+**Regional ocean models** focus computational resources on a limited geographical area with fine grid spacing, resolving fine-scale phenomena such as mesoscale and submesoscale features, tidal dynamics, coastal currents, upwelling, and detailed biogeochemical (BGC) processes. A widely used regional ocean model is the **Regional Ocean Modeling System (ROMS)** [@shchepetkin_regional_2005]. ROMS has been coupled to the Marine Biogeochemistry Library (MARBL) [@long_simulations_2021; @ucla-roms] to link physical and BGC processes. ROMS-MARBL supports research on environmental management, fisheries, regional climate impacts, and ocean-based carbon dioxide removal (CDR) strategies.
 
-A widely used regional ocean model is the **Regional Ocean Modeling System (ROMS)** [@shchepetkin_regional_2005]. To connect physical circulation with ecosystem dynamics and the ocean carbon cycle, ROMS has been coupled to a BGC model called the Marine Biogeochemistry Library (MARBL) [@long_simulations_2021; @ucla-roms]. This coupled framework allows researchers to explore a variety of scientific and practical questions.  For example, it can be used to investigate the potential of ocean-based carbon removal strategies, such as adding alkaline materials to the ocean to sequester atmospheric carbon dioxide. It can also be used to study how physical processes drive ecosystem dynamics, such as how nutrient-rich waters from upwelling fuel the phytoplankton blooms that form the base of the marine food web [@gruber_eddy-resolving_2006].
[…]
