### 2026-01-21 8cf5aca
Shorten sentence
