### 2026-01-21 32e6ba3
Include GitHub stats
