### 2026-01-21 0abbc9b
adding emod dataset as a class (#531)
* adding emod dataset as a class (#530)

* Add EMOD dataset hash

* Add minimial EMOD example

* Fix pre-commit errors

* Apply lateral fill to topo data

* Assign mask to EMOD dataset

---------

Co-authored-by: Nora Loose <nora.loose@gmail.com>
