+++ paper.md
@@ -1,5 +1,5 @@
 ---
-title: 'ROMS-Tools: Reproducible and Scalable Preprocessing and Analysis for Regional Ocean Modeling with ROMS'
+title: 'ROMS-Tools: Reproducible, Scalable Preprocessing and Analysis for Regional Ocean Modeling with ROMS'
 tags:
   - Python
   - ocean modeling
@@ -59,10 +59,10 @@ bibliography: docs/references.bib
 
 # Summary
 
-The ocean regulates Earth’s climate and sustains marine ecosystems by circulating and storing heat, carbon, oxygen, and nutrients, while exchanging gases with the atmosphere. Scientists study these processes using ocean models, which simulate the ocean on a grid.
+The ocean regulates Earth’s climate and sustains marine ecosystems by circulating and storing heat, carbon, oxygen, and nutrients, while exchanging heat and gases with the atmosphere. Scientists study these processes using ocean models, which simulate the ocean on a grid.
 **Regional ocean models** focus computational resources on a limited geographical area with fine grid spacing, and can resolve fine-scale phenomena such as mesoscale and submesoscale features, tidal dynamics, coastal currents, upwelling, and detailed biogeochemical (BGC) processes. A widely used regional ocean model is the **Regional Ocean Modeling System (ROMS)** [@shchepetkin_regional_2005]. ROMS has been coupled to the Marine Biogeochemistry Library (MARBL) [@long_simulations_2021; @ucla-roms] to link physical and BGC processes. ROMS-MARBL supports research on environmental management, fisheries, regional climate impacts, and ocean-based carbon dioxide removal (CDR) strategies.
 
-Configuring a regional ocean model like ROMS-MARBL is technically challenging. Setting up a model requires initializing and forcing it with oceanic and atmospheric data from multiple external sources in diverse formats, which can reach several petabytes for global datasets. These data must be subsetted, processed, and mapped onto the target domain’s geometry, producing input datasets of 10–100 terabytes for large regional models. Generating these input files is time-consuming, error-prone, and hard to reproduce, creating a bottleneck for both new and experienced users. The Python package `ROMS-Tools` addresses this challenge by providing efficient, `dask`-backed [@dask], user-friendly tools that can be installed via Conda or PyPI and run interactively from Jupyter notebooks. It supports creating regional grids, preprocessing all required model inputs, and postprocessing and analysis. Current capabilities are fully compatible with UCLA-ROMS [@ucla-roms; @ucla-roms-cworthy], with potential support for other ROMS versions, such as Rutgers ROMS [@rutgers-roms], in the future.
+Configuring a regional ocean model like ROMS-MARBL is technically challenging; it requires initialization and time-dependent forcing from oceanic and atmospheric data drawn from multiple external sources in diverse formats. These global datasets can reach several petabytes and must be subsetted, processed, and mapped onto the target domain’s geometry, yielding input datasets of 10–100 terabytes for large regional models. Generating these input files is time-consuming, error-prone, and hard to reproduce, creating a bottleneck for both new and experienced users. The Python package `ROMS-Tools` addresses this challenge by providing efficient, `dask`-backed [@dask], user-friendly tools that can be installed via Conda or PyPI and run interactively from Jupyter notebooks. It supports creating regional grids, preprocessing all required model inputs, and postprocessing and analysis. Current capabilities are fully compatible with UCLA-ROMS [@ucla-roms; @ucla-roms-cworthy], with potential support for other ROMS versions, such as Rutgers ROMS [@rutgers-roms], in the future.
 
 
 ## Input Data and Preprocessing
@@ -73,21 +73,21 @@ Configuring a regional ocean model like ROMS-MARBL is technically challenging. S
 2. **Bathymetry**: Derived from **SRTM15** [@tozer_global_2019].
[…]
