### 2026-01-21 54b4714
Add print statement
