### 2026-01-21 1e80a53
Make get_time_type more robust to all datetime types
