+++ roms_tools/datasets/utils.py
@@ -412,7 +412,13 @@ def get_time_type(data_array: xr.DataArray) -> str:
     TypeError
         If the values in the DataArray are not of type numpy.ndarray or list.
     """
-    # List of cftime datetime types
+    values = data_array.values
+
+    # 1. numpy datetime64 (any precision)
+    if np.issubdtype(values.dtype, np.datetime64):
+        return "datetime"
+
+    # 2. cftime objects
     cftime_types = (
         cftime.DatetimeNoLeap,
         cftime.DatetimeJulian,
@@ -420,23 +426,17 @@ def get_time_type(data_array: xr.DataArray) -> str:
         cftime.Datetime360Day,
         cftime.DatetimeProlepticGregorian,
     )
-
-    values = data_array.values
-
-    print(values)
-    # numpy datetime64
-    if values.dtype == "datetime64[ns]":
-        return "datetime"
-
-    # cftime objects (stored as object dtype)
-    if any(isinstance(value, cftime_types) for value in values):
+    if values.dtype == object and any(isinstance(v, cftime_types) for v in values):
         return "cftime"
 
-    # integer time axis
+    # 3. integer axis
     if np.issubdtype(values.dtype, np.integer):
         return "int"
 
-    raise ValueError("Unsupported data type for time values in input dataset.")
+    raise ValueError(
+        f"Unsupported data type for time values: {values.dtype}. "
+        "Expected datetime64, cftime objects, or integer."
+    )
 
 
 def convert_cftime_to_datetime(data_array: np.ndarray) -> np.ndarray:
