### 2026-01-21 4c38a53
Remove duplicate section
