+++ paper.md
@@ -121,44 +121,19 @@ The original MATLAB scripts were powerful but required users to edit source code
 
 ## Design Trade-Offs
 
-A central design trade-off in `ROMS-Tools` is between **automation** and **user control**. Rather than enforcing a fixed workflow, the package exposes key choices such as physical options (e.g., radiation or wind corrections), interpolation and fill methods, and computational backends. This approach contrasts with opinionated frameworks that fix defaults and directory structures to maximize automation. While users must make explicit decisions, some steps remain automated to prevent errors. For example, bathymetry smoothing is applied automatically with a non-tunable parameter, since overly small smoothing factors could produce rough bathymetry and crash simulations due to pressure gradient errors. This design choice directly addresses issues new users faced in the MATLAB scripts, balancing **flexibility, reproducibility, and safety**.
+A central design trade-off in `ROMS-Tools` is between **automation** and **user control**. Rather than enforcing a fixed workflow, the package exposes key choices such as physical options (e.g., radiation or wind corrections), interpolation and fill methods, and computational backends. This approach contrasts with opinionated frameworks that fix defaults and directory structures to maximize automation. While users must make explicit decisions, some steps remain automated to prevent errors. For example, bathymetry smoothing is applied automatically with a non-tunable parameter, since overly small smoothing factors could produce rough bathymetry and crash simulations due to pressure gradient errors. This design choice directly addresses issues new users faced in the MATLAB scripts, and balances **flexibility, reproducibility, and safety**, enabling transparent experimentation withoutexposing users to avoidable pitfalls.
 
 Another key trade-off is between **monolithic workflows** and **incremental, modular steps**. `ROMS-Tools` uses small, composable components (such as generating initial conditions, boundary forcing, and surface forcing) that can be executed, saved, and revisited independently. This avoids unnecessary recomputation when only some inputs change. To ensure reproducibility despite a modular workflow, configuration choices are stored in compact, text-based YAML files. These files are version-controllable, easy to share, and remove the need to transfer large model input NetCDF datasets. This directly remedies the MATLAB scripts’ lack of explicit workflow tracking.
 
 ## Architecture
 
-At the user-facing level, `ROMS-Tools` provides high-level objects such as `Grid`, `InitialConditions`, and `BoundaryForcing`. Each object exposes a consistent interface (`.ds`, `.plot()`, `.save()`, `.to_yaml()`), so users can always call the same methods in sequence or inspect attributes that are guaranteed to exist. This design reduces cognitive overhead and prevents the need for new users to manipulate raw scripts or track intermediate files manually.
+At the user-facing level, `ROMS-Tools` provides high-level objects such as `Grid`, `InitialConditions`, and `BoundaryForcing`. Each object exposes a consistent interface (`.ds`, `.plot()`, `.save()`, `.to_yaml()`), allowing users to call the same methods in sequence or inspect attributes that are always present. This design reduces cognitive overhead, makes workflows predictable and easy to follow, and eliminates the need for new users to manipulate raw scripts or track intermediate files manually.
 
[…]
