+++ paper.md
@@ -140,9 +140,9 @@ Internally, `ROMS-Tools` follows a **layered, modular architecture**. Low-level 
 `ROMS-Tools` is used by two primary research communities. First, regional ocean modelers use it to generate reproducible input datasets for ROMS simulations; external users include researchers at **PNNL**, **WHOI**, and **UCLA**. Second, researchers in the ocean-based carbon dioxide removal (CDR) community use `ROMS-Tools` to configure reproducible ROMS–MARBL simulations of climate intervention scenarios, with adopters including **[C]Worthy**, **Carbon to Sea**, **Ebb Carbon**, and **SCCWRP**.
 All of these groups have contacted the developers directly or engaged with the project through GitHub or offline discussions.
 
-Additional evidence of community uptake comes from public usage metrics. At the time of writing, the GitHub repository shows **119 unique cloners in the past 14 days**, with stars from users at institutions including the University of Waikato, NCAR, University of Maryland, National Oceanography Centre, McGill University, UC Santa Cruz, and others. Distribution statistics indicate **over 3,100 conda-forge downloads in the past six months**, including **68 downloads of the most recent release (v3.3.0)**, and **more than 48,000 total PyPI downloads** (noting that PyPI counts include automated CI usage, whereas conda downloads do not).
+Beyond standalone use, ROMS-Tools is integrated into broader scientific workflows, including C-Star [@cstar], an open-source platform that provides scientifically credible monitoring, reporting, and verification (MRV) for the emerging marine carbon market.
 
-`ROMS-Tools` is also integrated into broader workflows, including **C-Star** [@cstar], an open-source platform to provide scientifically credible monitoring, reporting, and verification (MRV) for the emerging marine carbon market.
+Additional evidence of community uptake comes from public usage metrics. At the time of writing, the GitHub repository shows **119 unique cloners in the past 14 days**, with stars from users at institutions including the University of Waikato, NCAR, University of Maryland, National Oceanography Centre, McGill University, UC Santa Cruz, and others. Distribution statistics indicate **over 3,100 conda-forge downloads in the past six months**, including **68 downloads of the most recent release** (`v3.3.0`), and **more than 48,000 total PyPI downloads**. PyPI counts include automated continuous integration (CI) usage by `ROMS-Tools`, in addition to direct user installations. In contrast, conda-forge downloads of `v3.3.0` reflect exclusively human-initiated installs, as C-Star's CI workflows currently pin pre-`v3.3.0` releases of `ROMS-Tools`.
 
 # AI Usage Disclosure
 
