+++ docs/releases.md
@@ -19,6 +19,7 @@
 
 * Default boundary selection respects land-only boundaries for `BoundaryForcing` and `ChildGrid` ([#504](https://github.com/CWorthy-ocean/roms-tools/pull/504))
 * Partition `eta_u` and `xi_v` dimension if present ([#526](https://github.com/CWorthy-ocean/roms-tools/pull/526))
+* Add support for EMOD bathymetry ([#531](https://github.com/CWorthy-ocean/roms-tools/pull/531))
 
 ### Breaking Changes
 
