### 2026-01-22 412ad05
Docs: minor fixes and clarifications (#538)
* Link to C-Worthy ucla-roms

* Complete release notes

* Clarify container executable in end to end example

* Point to cworthy ucla-roms
