+++ paper.md
@@ -80,13 +80,13 @@ Configuring a regional ocean model like ROMS-MARBL is technically challenging; i
 9. **River Forcing:** Freshwater runoff derived from **Dai & Trenberth** [@dai_estimates_2002] or user-provided custom files.
 10. **CDR Forcing**: User-defined interventions that inject BGC tracers at point sources or as larger-scale Gaussian perturbations to simulate CDR interventions. The CDR forcing is prescribed as volume and tracer fluxes (e.g., alkalinity for ocean alkalinity enhancement, iron for iron fertilization, or other BGC constituents).  Users can control the magnitude, spatial footprint, and temporal evolution, allowing flexible representation of CDR interventions.
 
-Some source datasets are accessed automatically by the package, including Natural Earth, Dai & Trenberth runoff, and ERA5 meteorology, while users must manually download SRTM15, GSHHG, GLORYS, the BGC datasets, and TPXO tidal files. Although these are the datasets currently supported, the package’s modular design makes it straightforward to add new source datasets in the future.
+Some source datasets are accessed automatically by the `ROMS-Tools`, including Natural Earth, Dai & Trenberth runoff, and ERA5 meteorology, while users must manually download SRTM15, GSHHG, GLORYS, the BGC datasets, and TPXO tidal files. Although these are the datasets currently supported, the modular design of `ROMS-Tools` makes it straightforward to add new source datasets in the future.
 
 To generate the model inputs, `ROMS-Tools` automates several intermediate processing steps, including:
 
 * **Bathymetry processing**: The bathymetry is smoothed in two stages, first across the entire model domain and then locally in areas with steep slopes, to ensure local steepness ratios do not exceed a prescribed threshold in order to reduce pressure-gradient errors. A minimum depth is enforced to prevent water levels from becoming negative during large tidal excursions.
 * **Mask definition**: The land-sea mask is generated by comparing the ROMS grid’s horizontal coordinates with a coastline dataset using the `regionmask` package [@hauser_regionmaskregionmask_2024]. Enclosed basins are subsequently filled with land.
-* **Land value handling**: Land values are filled via an algebraic multigrid method using `pyamg` [@pyamg2023] prior to horizontal regridding. This extends ocean values into land areas to resolve discrepancies between source data and ROMS land masks, preventing land-originating values from appearing in ocean cells.
+* **Land value handling**: Land values are filled via an algebraic multigrid method using `pyamg` [@pyamg2023] prior to horizontal regridding. This extends ocean values into land areas to reconcile discrepancies between source data and ROMS land masks, ensuring that no NaNs or land-originating values contaminate ocean grid cells.
 * **Regridding**: Ocean and atmospheric fields are horizontally and vertically regridded from standard latitude-longitude-depth grids to the model’s curvilinear grid with a terrain-following vertical coordinate using `xarray` [@hoyer2017xarray] and `xgcm` [@xgcm]. Velocities are rotated to align with the curvilinear ROMS grid.
 * **Longitude conventions**: `ROMS-Tools` handles differences in longitude conventions, converting between [-180°, 180°] and [0°, 360°] as needed.
 * **River locations**: Rivers that fall within the model domain are automatically identified and relocated to the nearest coastal grid cell. Rivers that need to be shifted manually or span multiple cells can be configured by the user.
@@ -121,7 +121,7 @@ The legacy MATLAB preprocessing scripts were powerful but required users to edit
 
 ## Design Trade-Offs
 
[…]
