### 2026-01-26 caa05e1
Correct figure caption
