### 2026-01-26 b29f9c5
Update paper (#533)
* Shorten paper and add new required sections

* Add refs

* Correct format

* (auto) Paper PDF Draft

* Change title

* (auto) Paper PDF Draft

* Shorten sentence

* Modify sentence

* (auto) Paper PDF Draft

* Include GitHub stats

* Add C-Star reference

* (auto) Paper PDF Draft

* Tweak text

* (auto) Paper PDF Draft

* Remove duplicate section

* (auto) Paper PDF Draft

* Paper update

* (auto) Paper PDF Draft

* Final tweaks

* (auto) Paper PDF Draft

* Clarify download stats

* (auto) Paper PDF Draft

* Mention manuscripts in prep

* (auto) Paper PDF Draft

* Sam's c
[…]
