+++ docs/references.bib
@@ -495,3 +495,16 @@ MARBL is a flexible, plankton functional type model with a modular architecture 
 	pages = {8741--8743},
 	file = {Snapshot:/Users/noraloose/Zotero/storage/SK2BUP9P/96JB00104.html:text/html},
 }
+
+@misc{cstar,
+	title = {C-Star},
+	copyright = {Apache-2.0},
+	url = {https://github.com/CWorthy-ocean/C-Star},
+	abstract = {C-Star is a python package for setting up and running ocean model simulations, with a particular focus on marine carbon dioxide removal (mCDR) applications.},
+	urldate = {2026-01-21},
+	publisher = {[C]Worthy},
+	author = {Dafydd Stephenson and contributors},
+	month = dec,
+	year = {2025},
+	note = {original-date: 2024-08-08T23:01:29Z},
+}
