### 2026-01-26 #9492 @NoraLoose
@kthyng I have updated the paper, and it now includes the additional sections "Software Design", "Research Impact Statement" and "AI Usage Disclosure". Thanks!
