### 2026-01-27 446cd63
Refactor `rotate_velocities` and staggered interpolation (#527)
* Refactor rotate_velocities and staggered interpolation

* Add rotate_velocities_to_east_and_north methods to Dataset

* Add blank lines

* Add to release notes

* Add blank line
