### 2026-01-27 6b27681
Make chunking safe
