### 2026-01-29 f599c03
Rewrite paper to include State of Field section
