+++ paper.md
@@ -61,7 +61,7 @@ bibliography: docs/references.bib
 
 The ocean regulates Earth’s climate and sustains marine ecosystems by circulating and storing heat, carbon, oxygen, and nutrients, while exchanging heat and gases with the atmosphere. Scientists study these processes using ocean models, which simulate the ocean on a grid. **Regional ocean models** focus computational resources on a limited geographical area with fine grid spacing, and can resolve fine-scale phenomena such as mesoscale and submesoscale features, tidal dynamics, coastal currents, upwelling, and detailed biogeochemical (BGC) processes. A widely used regional ocean model is the **Regional Ocean Modeling System (ROMS)** [@shchepetkin_regional_2005]. ROMS has been coupled to the Marine Biogeochemistry Library (MARBL) [@long_simulations_2021; @ucla-roms] to link physical and BGC processes. ROMS-MARBL supports research on environmental management, fisheries, regional climate impacts, and ocean-based carbon dioxide removal (CDR) strategies.
 
-`ROMS-Tools` is a Python package that streamlines the **preparation and analysis of ROMS-MARBL simulations** by enabling users to generate regional grids, prepare model inputs, and analyze outputs efficiently. The package integrates with `xarray` [@xarray] for labeled, multi-dimensional data handling, and supports large input and output datasets via parallel computation with `dask` [@dask], making workflows scalable from laptops to high-performance computing clusters. By providing a modern, user-friendly interface, `ROMS-Tools` lowers technical barriers, improves reproducibility, and allows scientists to focus on research rather than data preparation. The package is installable via Conda or PyPI and can be run interactively in Jupyter notebooks. Its modular design facilitates extension to new datasets.
+`ROMS-Tools` is a Python package that streamlines the **preparation and analysis of ROMS-MARBL simulations** by enabling users to generate regional grids, prepare model inputs, and analyze outputs efficiently. By providing a modern, user-friendly interface, `ROMS-Tools` lowers technical barriers, improves reproducibility, and allows scientists to focus on research rather than data preparation. The package is installable via Conda or PyPI and can be run interactively in Jupyter notebooks. Its modular design facilitates extension to new datasets.
 
 # Statement of Need
 
@@ -70,11 +70,13 @@ Model setup requires initialization and time-dependent forcing from multiple oce
 
 Existing preprocessing tools within the ocean modeling ecosystem do not fully address these challenges for ROMS-MARBL. Legacy MATLAB-based scripts and packages such as `pyroms` [@pyroms] provide critical functionality but rely on low-level, manually coordinated steps that limit reproducibility, maintainability, and accessibility. In contrast, preprocessing frameworks developed for other ocean models (e.g., MOM6) cannot be directly applied to ROMS due to fundamental differences in grid geometry, vertical coordinates, and model input data requirements. As a result, ROMS-MARBL users lack a modern, integrated framework for reproducible model setup and analysis.
 
[…]
