+++ paper.md
@@ -68,7 +68,7 @@ The ocean regulates Earth’s climate and sustains marine ecosystems by circulat
 Regional ocean models are essential tools for research in marine ecosystems, climate dynamics, and ocean-based CDR. However, configuring a regional ocean model like ROMS-MARBL is technically demanding.
 Model setup requires initialization and time-dependent forcing from oceanic and atmospheric datasets, drawn from multiple external sources in diverse formats. These global source datasets can span petabytes and must be subsetted, processed, and mapped onto the target model grid, producing 10–100 terabytes of input data for large regional domains. Generating these input files is time-consuming, error-prone, and difficult to reproduce. These challenges create a bottleneck for both new and experienced users, slow down science, and limit collaboration across groups.
 
-Existing tools within the ocean modeling ecosystem do not fully address these challenges for ROMS or ROMS-MARBL. Legacy MATLAB-based scripts and packages such as `pyroms` [@pyroms] provide critical functionality but rely on low-level, manually coordinated steps that limit reproducibility, maintainability, and accessibility. In contrast, frameworks developed for other ocean models (e.g., MOM6) cannot be directly applied to ROMS due to fundamental differences in grid geometry, vertical coordinates, and model input data requirements. As a result, ROMS-MARBL users lack a modern, integrated framework for reproducible model setup and analysis that is specifically designed for ROMS and ROMS-MARBL.
+Existing tools within the ocean modeling ecosystem do not fully address these challenges for ROMS-MARBL or ROMS. While legacy MATLAB-based scripts developed at UCLA [@ucla-matlab] and Python packages such as `pyroms` [@pyroms] provide critical functionality, both but rely on low-level, manually coordinated steps that limit reproducibility, maintainability, and accessibility. In contrast, frameworks developed for other ocean models (e.g., MOM6) cannot be directly applied to ROMS due to fundamental differences in grid geometry, vertical coordinates, and model input data requirements. As a result, ROMS-MARBL users lack a modern, integrated framework for reproducible model setup and analysis that is specifically designed for ROMS and ROMS-MARBL.
 
 `ROMS-Tools` was developed to fill this gap. It is an open-source Python framework designed for researchers and practitioners who run ROMS or ROMS-MARBL regional ocean simulations, including users in physical oceanography, marine biogeochemistry, and ocean-based CDR applications. Current capabilities are fully compatible with UCLA-ROMS [@ucla-roms; @ucla-roms-cworthy], with potential support for other ROMS implementations, such as Rutgers ROMS [@rutgers-roms], in the future.
 The package provides high-level APIs that automate and standardize preprocessing workflows and manage complex model configuration state via YAML, supporting reproducible simulation setup.
