### 2026-01-29 d9a0dc8
Remove sentence
