### 2026-01-29 bfbf8ed
Include "state of field" section in JOSS paper (#543)
* Rewrite paper to include State of Field section

* (auto) Paper PDF Draft

* Some more fixes

* (auto) Paper PDF Draft

* More tweaks

* (auto) Paper PDF Draft

* Further tweaks

* (auto) Paper PDF Draft

* Typos

* (auto) Paper PDF Draft

* Remove sentence

* (auto) Paper PDF Draft

* Final edit

* (auto) Paper PDF Draft
