+++ roms_tools/tests/test_setup/test_validation.py
@@ -104,13 +104,17 @@ def test_check_results(
     [
         "tidal_forcing",
         "initial_conditions_with_bgc_from_climatology",
+        "initial_conditions_with_unified_bgc_from_climatology",
+        pytest.param("initial_conditions_from_roms", marks=skip_xesmf),
         "surface_forcing",
         "coarse_surface_forcing",
         "corrected_surface_forcing",
         "bgc_surface_forcing",
         "bgc_surface_forcing_from_climatology",
+        "bgc_surface_forcing_from_unified_climatology",
         "boundary_forcing",
         "bgc_boundary_forcing_from_climatology",
+        "bgc_boundary_forcing_from_unified_climatology",
     ],
 )
 def test_dask_vs_no_dask(
