### 2026-02-02 366a0b0
Safe way to convert to seconds regardless of xarray version
