+++ roms_tools/setup/initial_conditions.py
@@ -755,7 +755,7 @@ class InitialConditions:
         model_reference_date = np.datetime64(self.model_reference_date)
 
         # Convert the time coordinate to the format expected by ROMS (seconds since model reference date)
-        ocean_time = (ds["time"] - model_reference_date).astype("float64") * 1e-9
+        ocean_time = (ds["time"] - model_reference_date).dt.total_seconds()
         ds = ds.assign_coords(ocean_time=("time", ocean_time.data.astype("float64")))
         ds["ocean_time"].attrs["long_name"] = (
             f"relative time: seconds since {self.model_reference_date!s}"
