### 2026-02-02 02d3b25
Initial conditions from outer nest ROMS restart file (#513)
* Introduce ROMSDataset class

* Move tests

* Move dataset modules into their own directory

* Shared utility function

* Test shared utility function

* Move wrap_longitude function

* Adjust longitude range

* Make wrap_longitudes function safe

* Test choose_subdomain

* Test extrapolation to bottom

* Move fill module

* Use shared extrapolation function

* Move one_dim_fill function

* Correct import

* Test lateral fill

* Test variable_mapping

* Correct API ref

* Add to release notes

* First stab

* Shared validate_start_en
[…]
