+++ roms_tools/setup/mask.py
@@ -1,4 +1,5 @@
 import logging
+import time
 import warnings
 from pathlib import Path
 
@@ -12,8 +13,18 @@ from roms_tools.setup.utils import handle_boundaries
 from roms_tools.utils import interpolate_from_rho_to_u, interpolate_from_rho_to_v
 
 
-def add_mask(ds: xr.Dataset, shapefile: str | Path | None = None) -> xr.Dataset:
+def add_mask(
+    ds: xr.Dataset,
+    shapefile: str | Path | None = None,
+    close_narrow_channels: bool = False,
+    verbose: bool = False,
+) -> xr.Dataset:
     """Adds a land/water mask to the dataset at rho-points.
+
+    These are the steps:
+    1. Infer mask from coastlines
+    2. Close narrow channels if requested
+    3. Fill enclosed basins
 
     Parameters
     ----------
@@ -23,12 +34,21 @@ def add_mask(ds: xr.Dataset, shapefile: str | Path | None = None) -> xr.Dataset:
     shapefile: str or Path | None
         Path to a coastal shapefile to determine the land mask. If None, NaturalEarth 10m is used.
 
+    close_narrow_channels : bool, optional
+        Whether to close narrow water channels in the mask after it is generated.
+        The default is False.
+
+    verbose: bool, optional
+        Indicates whether to print mask generation steps with timing. Defaults to False.
+
     Returns
     -------
     xarray.Dataset
         The original dataset with an added 'mask_rho' variable, representing land/water mask.
     """
-    # Suppress specific warning
+    # Infer mask from coastlines
+    if verbose:
+        start_time = time.time()
     with warnings.catch_warnings():
         warnings.filterwarnings(
             "ignore", message="No gridpoint belongs to any region.*"
@@ -61,9 +81,36 @@ def add_mask(ds: xr.Dataset, shapefile: str | Path | None = None) -> xr.Dataset:
             mask = land_mask.isnull()
 
     ds = _add_coastlines_metadata(ds, shapefile)
+    if verbose:
+        logging.info(
+            f"Inferring the mask from coastlines: {time.time() - start_time:.3f} seconds"
+        )
+
+    mask = mask.astype(int)  # convert from bool to int
+
+    # Close narrow channels if requested
+    if close_narrow_channels:
+        if verbose:
+            start_time = time.time()
+        _close_narrow_channels(
+            mask,
+            max_iterations=10,
+        )
+        if verbose:
+            logging.info(
+                "Closing narrow channels: {time.time() - start_time:.3f} seconds"
+            )
+    ds.attrs["close_narrow_channels"] = str(close_narrow_channels)
 
     # fill enclosed basins with land
+    if verbose:
+        start_time = time.time()
     mask = _fill_enclosed_basins(mask.values)
+    if verbose:
+        logging.info(
+            "Filling enclosed basins with land: {time.time() - start_time:.3f} seconds"
+        )
+
     # adjust mask boundaries by copying values from adjacent cells
     mask = handle_boundaries(mask)
 
@@ -72,8 +119,6 @@ def add_mask(ds: xr.Dataset, shapefile: str | Path | None = None) -> xr.Dataset:
         "long_name": "Mask at rho-points",
         "units": "land/water (0/1)",
     }
-
-    ds = add_velocity_masks(ds)
 
     return ds
 
@@ -103,27 +148,30 @@ def _add_coastlines_metadata(
     return ds
 
 
-def _fill_enclosed_basins(mask) -> np.ndarray:
-    """Fills enclosed basins in the mask with land (value = 1).
+def _fill_enclosed_basins(mask: np.ndarray) -> np.ndarray:
+    """Fills enclosed basins in the mask with land (value = 0).
 
[…]

[…]
