### 2026-02-02 06552ca
adding fill_narrows functionality
