### 2026-02-02 4ece8ca
Add `fill_narrow_passages` method to `Grid` class for improved mask handling
* Introduced a new method to fill narrow passages and holes in the ROMS mask.
* Updated velocity masks automatically after modifying the mask.
* Enhanced documentation with detailed parameters and examples for clarity.
