### 2026-02-02 54c6a2d
updating the fill narrows test to accomodate the grid halo
