### 2026-02-02 52fc8e6
Add logger initialization in mask.py for improved debugging
