### 2026-02-03 6f5a77d
Update release notes (#544)
* Update release notes

* Exclude docs from coverage
