+++ docs/releases.md
@@ -1,10 +1,8 @@
 # Release notes
 
-## v3.5.0 (unreleased)
+## v3.6.0 (unreleased)
 
 ### New Features
-
-* Initialize inner-nest initial conditions from outer-nest ROMS restart file ([513](https://github.com/CWorthy-ocean/roms-tools/pull/513/files))
 
 ### Breaking Changes
 
@@ -14,6 +12,11 @@
 
 ### Bugfixes
 
+## v3.5.0
+
+### New Features
+
+* Initialize inner-nest initial conditions from outer-nest ROMS restart file ([513](https://github.com/CWorthy-ocean/roms-tools/pull/513/files))
 
 ## v3.4.0
 
