### 2026-02-03 a1e121b
change grid documentation
