### 2026-02-05 965e9a5
Rename `fill_narrow_passages` to `_close_narrow_channels` for clarity and update related documentation. Enhance `Grid` class to include an option for closing narrow channels after mask generation. Update tests and notebook documentation to reflect these changes.
