### 2026-02-05 580b9dd
Enhance `_close_narrow_channels` function in `mask.py` to include a `verbose` parameter for detailed logging. Update `Grid` class's `update_mask` method to conditionally close narrow channels based on a new `close_narrow_channels` parameter, improving flexibility in mask updates.
