### 2026-02-05 caa1b02
Refactor import order in `grid.py` for consistency and update `ChildGrid` class's `update_mask` method to include a new `close_narrow_channels` parameter, enhancing flexibility in mask updates.
