### 2026-02-05 e916231
Include time records strictly outside start/end bounds (#547)
* Include time record strictly before start time and strictly after end time

* Add to release notes

* Overwrite test data due to more time stamps

* Refactor and correct compare hash tests

* Fix tests
