### 2026-02-05 6dd2f03
Use open boundaries as default
