+++ test_close_narrow_channels_notebook.py
@@ -1,0 +1,216 @@
+# Test and visualize close_narrow_channels functionality
+# Paste this into a Jupyter notebook cell
+
+import numpy as np
+import matplotlib.pyplot as plt
+from roms_tools import Grid
+from roms_tools.setup.mask import _close_narrow_channels, add_velocity_masks
+
+# ============================================================================
+# Test 1: Closing narrow 1-pixel water channels
+# Note: In ROMS masks, 1 = OCEAN (water) and 0 = LAND
+# ============================================================================
+print("=" * 80)
+print("Test 1: Closing narrow 1-pixel water channels")
+print("=" * 80)
+
+# Create a small grid with close_narrow_channels=False to avoid closing during init
+grid1 = Grid(
+    nx=15,
+    ny=15,
+    size_x=100,
+    size_y=100,
+    center_lon=-20,
+    center_lat=64,
+    rot=0,
+    N=3,
+    close_narrow_channels=False,
+)
+
+# Get the actual shape of mask_rho (grid adds boundary cells, so it's nx+2, ny+2)
+mask_shape = grid1.ds["mask_rho"].shape
+
+# Create a mask with narrow 1-pixel water channels
+# Start with all land (0)
+mask = np.zeros(mask_shape, dtype=np.int32)
+
+# Create a narrow 1-pixel water channel in north-south direction
+mask[8, 8] = 1  # water cell (ocean)
+mask[7, 8] = 0  # land above
+mask[9, 8] = 0  # land below
+
+# Create a narrow 1-pixel water channel in east-west direction
+mask[11, 11] = 1  # water cell (ocean)
+mask[11, 10] = 0  # land left
+mask[11, 12] = 0  # land right
+
+# Set the mask in the grid
+grid1.ds["mask_rho"].values[:] = mask
+
+# Save original mask for comparison (as DataArray for plotting)
+mask_before_1 = grid1.ds.mask_rho.copy()
+
+print(f"\nBefore closing:")
+print(f"  NS water channel at [8, 8]: {mask_before_1.values[8, 8]} (should be 1 = ocean)")
+print(f"  EW water channel at [11, 11]: {mask_before_1.values[11, 11]} (should be 1 = ocean)")
+
+# Close narrow channels directly (as update_mask would do)
+grid1.ds = _close_narrow_channels(
+    grid1.ds,
+    mask_var="mask_rho",
+    max_iterations=10,
+    connectivity=4,
+    min_region_fraction=0.1,
+    inplace=True,
+    verbose=True,
+)
+# Update velocity masks after modifying mask_rho
+grid1.ds = add_velocity_masks(grid1.ds)
+
+# Get the mask after closing (as DataArray for plotting)
+mask_after_1 = grid1.ds.mask_rho.copy()
+
+print(f"\nAfter closing:")
+print(f"  NS water channel at [8, 8]: {mask_after_1.values[8, 8]} (should be 0 = land, channel closed)")
+print(f"  EW water channel at [11, 11]: {mask_after_1.values[11, 11]} (should be 0 = land, channel closed)")
+
+# ============================================================================
+# Test 2: Filling small lakes (removing small isolated water regions)
+# Note: In ROMS masks, 1 = OCEAN (water) and 0 = LAND
+# ============================================================================
+print("\n" + "=" * 80)
+print("Test 2: Filling small lakes - removing small isolated water regions")
+print("=" * 80)
+
[…]
