### 2026-02-05 6636d2c
Option to include BGC in boundary outputs
