+++ roms_tools/setup/nesting.py
@@ -53,6 +53,7 @@ class ChildGrid(Grid):
 
         - `"prefix"` (str): Prefix for variable names in `ds_nesting`. Defaults to `"child"`.
         - `"period"` (float): Temporal resolution for boundary outputs in seconds. Defaults to 3600 (hourly).
+        - `"include_bgc"` (bool): Whether to include BGC variables in boundary outputs. Defaults to `False`.
     verbose: bool, optional
         Indicates whether to print grid generation steps with timing. Defaults to False.
     """
@@ -64,7 +65,11 @@ class ChildGrid(Grid):
     """Specifies which child grid boundaries (south, east, north, west) should be
     adjusted for topography/mask and included in `ds_nesting`."""
     metadata: dict[str, Any] = field(
-        default_factory=lambda: {"prefix": "child", "period": 3600.0}
+        default_factory=lambda: {
+            "prefix": "child",
+            "period": 3600.0,
+            "include_bgc": False,
+        }
     )
     """Dictionary configuring the boundary nesting process."""
     verbose: bool = False
@@ -97,6 +102,7 @@ class ChildGrid(Grid):
                 self.boundaries,
                 self.metadata["prefix"],
                 self.metadata["period"],
+                self.metadata["include_bgc"],
             )
 
             self.ds_nesting = ds_nesting
@@ -380,6 +386,7 @@ def map_child_boundaries_onto_parent_grid_indices(
     boundaries: dict = {"south": True, "east": True, "north": True, "west": True},
     prefix: str = "child",
     period: float = 3600.0,
+    include_bgc: bool = False,
     update_land_indices: bool = True,
 ):
     """Maps child grid boundary points onto absolute indices of the parent grid.
@@ -410,6 +417,9 @@ def map_child_boundaries_onto_parent_grid_indices(
     period : float, optional
         The output period (in seconds) to be assigned to the mapped boundary indices.
         Defaults to `3600.0` (1 hour).
+
+    include_bgc: bool, optional
+        Whether to include BGC variables in the boundary outputs.
 
     update_land_indices : bool, optional
         If `True`, updates indices that fall on land in the parent grid to nearby ocean points.
@@ -472,7 +482,10 @@ def map_child_boundaries_onto_parent_grid_indices(
                         f"{grid_location}-points of {direction}ern child boundary mapped onto parent (absolute) grid indices"
                     )
                     ds[var_name].attrs["units"] = "non-dimensional"
-                    ds[var_name].attrs["output_vars"] = "zeta, temp, salt"
+                    if include_bgc:
+                        ds[var_name].attrs["output_vars"] = "zeta, temp, salt, bgc"
+                    else:
+                        ds[var_name].attrs["output_vars"] = "zeta, temp, salt"
                 else:
                     angle_child = child_grid_ds[names["angle"]].isel(
                         **bdry_coords[direction]
