### 2026-02-05 90ea587
fixing the broken close_narrow_passage test by making a new test mask
