+++ roms_tools/tests/test_setup/test_grid.py
@@ -829,9 +829,10 @@ def test_update_mask():
 
 def test_close_narrow_channels():
     """Test that close_narrow_channels closes 1-pixel wide water channels.
-    
-    Creates an H-shaped land mask where the distance between the two vertical
-    legs is one grid cell, creating a narrow water channel that should be closed.
+
+    Creates a mask with a vertical line of ocean, a small lake connected by a narrow
+    channel.
+    All narrow channels should be closed by the algorithm and the lake should be filled in.
     Note: In ROMS masks, 1 = OCEAN (water) and 0 = LAND.
     """
     # Create a small grid with close_narrow_channels=False to avoid closing during init
@@ -850,34 +851,46 @@ def test_close_narrow_channels():
     # Get the actual shape of mask_rho (grid adds boundary cells, so it's nx+2, ny+2)
     mask_shape = grid.ds["mask_rho"].shape
 
-    # Start with all ocean (1 = water)
-    mask = np.ones(mask_shape, dtype=np.int32)
+    # Start with all land (0 = land)
+    mask = np.zeros(mask_shape, dtype=np.int32)
 
-    # Create an H-shaped land mask (0 = land)
-    # The H has two vertical legs and one horizontal bar connecting them
-    # The distance between the two vertical legs is one grid cell
-    
-    # Left vertical leg of H
-    left_xi = 7
-    right_xi = 9  # One cell to the right of left leg
+    # Create a single vertical line of ocean (1 = water)
+    # This creates narrow water channels on either side that should be closed
+
+    # Single vertical line
+    line_xi = 8
     top_eta = 5
     bottom_eta = 10
-    mid_eta = 7  # Where the horizontal bar connects
-    
-    mask[top_eta:bottom_eta+1, left_xi] = 0  # Left leg
-    mask[top_eta:bottom_eta+1, right_xi] = 0  # Right leg
-    mask[mid_eta, left_xi:right_xi+1] = 0  # Horizontal bar connecting them
-    
-    # This creates narrow water channels at xi=8 (between the legs)
+
+    mask[top_eta : bottom_eta + 1, line_xi] = 1  # Single vertical line (ocean)
+
+    # Set the entire lower half of the mask to ocean
+    mid_point = mask_shape[0] // 2
+    mask[mid_point:, :] = 1  # Lower half is all ocean
+
+    # Create a small lake in the upper half, connected to the ocean by a narrow channel
+    lake_eta_start = 3
+    lake_eta_end = 5
+    lake_xi_start = 4
+    lake_xi_end = 6
+    mask[lake_eta_start : lake_eta_end + 1, lake_xi_start : lake_xi_end + 1] = 1  # Small lake
+
+    # Create a narrow channel connecting the lake to the vertical ocean line
+    # The channel is 1 cell wide at xi=7, connecting from the lake to the vertical line
+    channel_eta = 4  # Middle of the lake vertically
+    mask[channel_eta, lake_xi_end : line_xi] = 1  # Narrow channel connecting lake to ocean
+
+    # This creates narrow water channels at xi=7 and xi=9 
     # The channels are one cell wide and should be closed by the algorithm
-    # Specifically, the water at xi=8, eta=6 and eta=8 (above and below the horizontal bar)
+    # Also, the narrow channel connecting the lake to the ocean should be closed
 
     # Set the mask in the grid
     grid.ds["mask_rho"].values[:] = mask
 
     # Verify the narrow water channels exist before closing
-    assert mask[6, 8] == 1, "Narrow water channel should exist above horizontal bar"
-    assert mask[8, 8] == 1, "Narrow water channel should exist below horizontal bar"
+    assert mask[6, 7] == 1, "Narrow water channel should exist at [6, 7]"
[…]
