+++ test_close_narrow_channels_viz.py
@@ -1,194 +1,0 @@
-"""
-Visualization script for close_narrow_channels functionality.
-This script demonstrates the effect of closing narrow channels on ROMS masks.
-"""
-
-import numpy as np
-import matplotlib.pyplot as plt
-from roms_tools import Grid
-from roms_tools.setup.mask import _close_narrow_channels, add_velocity_masks
-from roms_tools.plot import plot
-
-# ============================================================================
-# Test 1: Closing narrow 1-pixel passages
-# ============================================================================
-print("=" * 80)
-print("Test 1: Closing narrow 1-pixel passages")
-print("=" * 80)
-
-# Create a small grid with close_narrow_channels=False to avoid closing during init
-grid1 = Grid(
-    nx=15,
-    ny=15,
-    size_x=100,
-    size_y=100,
-    center_lon=-20,
-    center_lat=64,
-    rot=0,
-    N=3,
-    close_narrow_channels=False,
-)
-
-# Get the actual shape of mask_rho (grid adds boundary cells, so it's nx+2, ny+2)
-mask_shape = grid1.ds["mask_rho"].shape
-
-# Create a mask with narrow 1-pixel land passages
-# Start with all water (0)
-mask = np.zeros(mask_shape, dtype=np.int32)
-
-# Create a narrow 1-pixel land passage in north-south direction
-# A land cell with water above and below (vertically isolated)
-mask[8, 8] = 1  # land cell
-mask[7, 8] = 0  # water above
-mask[9, 8] = 0  # water below
-
-# Create a narrow 1-pixel land passage in east-west direction
-# A land cell with water left and right (horizontally isolated)
-mask[11, 11] = 1  # land cell
-mask[11, 10] = 0  # water left
-mask[11, 12] = 0  # water right
-
-# Set the mask in the grid
-grid1.ds["mask_rho"].values[:] = mask
-
-# Save original mask for comparison
-mask_before_1 = grid1.ds.mask_rho.values.copy()
-
-print(f"\nBefore closing:")
-print(f"  NS passage at [8, 8]: {mask_before_1[8, 8]} (should be 1)")
-print(f"  EW passage at [11, 11]: {mask_before_1[11, 11]} (should be 1)")
-
-# Close narrow channels directly (as update_mask would do)
-grid1.ds = _close_narrow_channels(
-    grid1.ds,
-    mask_var="mask_rho",
-    max_iterations=10,
-    connectivity=4,
-    min_region_fraction=0.1,
-    inplace=True,
-    verbose=True,
-)
-# Update velocity masks after modifying mask_rho
-grid1.ds = add_velocity_masks(grid1.ds)
-
-# Get the mask after closing
-mask_after_1 = grid1.ds.mask_rho.values.copy()
-
-print(f"\nAfter closing:")
-print(f"  NS passage at [8, 8]: {mask_after_1[8, 8]} (should be 0)")
-print(f"  EW passage at [11, 11]: {mask_after_1[11, 11]} (should be 0)")
-
-# ============================================================================
-# Test 2: Hole filling (removing small isolated regions)
-# ============================================================================
-print("\n" + "=" * 80)
-print("Test 2: Hole filling - removing small isolated regions")
-print("=" * 80)
-
-# Create a small grid with close_narrow_channels=False to avoid closing during init
-grid2 = Grid(
-    nx=20,
-    ny=20,
-    size_x=100,
-    size_y=100,
[…]

[…]
