### 2026-02-08 5de7f9a
changing verbose log output and simplifying the narrow channels test
