+++ roms_tools/tests/test_setup/test_grid.py
@@ -848,54 +848,19 @@ def test_close_narrow_channels():
         close_narrow_channels=False,
     )
 
-    # Get the actual shape of mask_rho (grid adds boundary cells, so it's nx+2, ny+2)
     mask_shape = grid.ds["mask_rho"].shape
-
-    # Start with all land (0 = land)
     mask = np.zeros(mask_shape, dtype=np.int32)
 
-    # Create a single vertical line of ocean (1 = water)
-    # This creates narrow water channels on either side that should be closed
-
-    # Single vertical line
+    # Create test mask: vertical ocean line, lower half ocean, small lake with narrow channel
     line_xi = 8
-    top_eta = 5
-    bottom_eta = 10
+    mask[5:11, line_xi] = 1
+    mask[mask_shape[0] // 2 :, :] = 1
+    mask[3:6, 4:7] = 1  # Small lake
+    mask[4, 6:8] = 1  # Narrow channel connecting lake to ocean
 
-    mask[top_eta : bottom_eta + 1, line_xi] = 1  # Single vertical line (ocean)
-
-    # Set the entire lower half of the mask to ocean
-    mid_point = mask_shape[0] // 2
-    mask[mid_point:, :] = 1  # Lower half is all ocean
-
-    # Create a small lake in the upper half, connected to the ocean by a narrow channel
-    lake_eta_start = 3
-    lake_eta_end = 5
-    lake_xi_start = 4
-    lake_xi_end = 6
-    mask[lake_eta_start : lake_eta_end + 1, lake_xi_start : lake_xi_end + 1] = 1  # Small lake
-
-    # Create a narrow channel connecting the lake to the vertical ocean line
-    # The channel is 1 cell wide at xi=7, connecting from the lake to the vertical line
-    channel_eta = 4  # Middle of the lake vertically
-    mask[channel_eta, lake_xi_end : line_xi] = 1  # Narrow channel connecting lake to ocean
-
-    # This creates narrow water channels at xi=7 and xi=9 
-    # The channels are one cell wide and should be closed by the algorithm
-    # Also, the narrow channel connecting the lake to the ocean should be closed
-
-    # Set the mask in the grid
     grid.ds["mask_rho"].values[:] = mask
-
-    # Verify the narrow water channels exist before closing
-    assert mask[6, 7] == 1, "Narrow water channel should exist at [6, 7]"
-    assert mask[6, 9] == 1, "Narrow water channel should exist at [6, 9]"
-    assert mask[4, 7] == 1, "Narrow channel connecting lake should exist at [4, 7]"
-
-    # Save original mask for comparison
     mask_before = grid.ds.mask_rho.values.copy()
 
-    # Close narrow channels directly (as update_mask would do)
     grid.ds = _close_narrow_channels(
         grid.ds,
         mask_var="mask_rho",
@@ -904,27 +869,14 @@ def test_close_narrow_channels():
         min_region_fraction=0.1,
         inplace=True,
     )
-    # Update velocity masks after modifying mask_rho
     grid.ds = add_velocity_masks(grid.ds)
 
-    # Get the mask after closing
     mask_after = grid.ds.mask_rho.values.copy()
 
-    # Verify that the narrow water channels were closed (converted to land)
-    assert mask_after[6, 7] == 0, (
-        "Narrow water channel at [6, 7] should be closed (converted to land)"
-    )
-    assert mask_after[6, 9] == 0, (
-        "Narrow water channel at [6, 9] should be closed (converted to land)"
-    )
-    assert mask_after[4, 7] == 0, (
-        "Narrow channel connecting lake at [4, 7] should be closed (converted to land)"
-    )
-
-    # Verify that the mask changed
[…]
