### 2026-02-08 c990799
fixing bug in verbose output
