### 2026-02-09 cbc57ca
Merge user metadata with defaults
