+++ roms_tools/tests/test_setup/test_nesting.py
@@ -52,7 +52,7 @@ def small_grid_that_straddles():
 
 
 @pytest.fixture()
-def child_grid(big_grid):
+def child_grid_with_bgc(big_grid):
     return ChildGrid(
         parent_grid=big_grid,
         nx=10,
@@ -62,6 +62,7 @@ def child_grid(big_grid):
         rot=-20,
         size_x=500,
         size_y=500,
+        metadata={"include_bgc": True},
     )
 
 
@@ -76,6 +77,21 @@ def child_grid_that_straddles(big_grid_that_straddles):
         rot=0,
         size_x=50,
         size_y=200,
+        metadata={"prefix": "child"},
+    )
+
+
+@pytest.fixture()
+def child_grid_with_eastern_boundary_closed(big_grid):
+    return ChildGrid(
+        parent_grid=big_grid,
+        nx=10,
+        ny=10,
+        center_lon=-23,
+        center_lat=65,
+        rot=-20,
+        size_x=200,
+        size_y=100,
     )
 
 
@@ -370,12 +386,18 @@ class TestModifyChid:
 
 class TestNesting:
     @pytest.mark.parametrize(
-        "child_grid_fixture",
-        ["child_grid", "child_grid_that_straddles"],
+        "fixture_name, include_bgc, expected_r_vars",
+        [
+            ("child_grid_with_bgc", True, "zeta, temp, salt, bgc"),
+            ("child_grid_that_straddles", False, "zeta, temp, salt"),
+        ],
     )
-    def test_successful_initialization(self, child_grid_fixture, request):
-        child_grid = request.getfixturevalue(child_grid_fixture)
+    def test_successful_initialization(
+        self, request, fixture_name, include_bgc, expected_r_vars
+    ):
+        child_grid = request.getfixturevalue(fixture_name)
 
+        # Basic metadata + structure checks
         assert child_grid.boundaries == {
             "south": True,
             "east": True,
@@ -384,31 +406,32 @@ class TestNesting:
         }
         assert child_grid.metadata["prefix"] == "child"
         assert child_grid.metadata["period"] == 3600.0
+        assert child_grid.metadata["include_bgc"] is include_bgc
         assert isinstance(child_grid.ds, xr.Dataset)
         assert isinstance(child_grid.ds_nesting, xr.Dataset)
 
+        # Nesting dataset checks
         ds = child_grid.ds_nesting
         for direction in ["south", "east", "north", "west"]:
             for location in ["r", "u", "v"]:
-                assert f"child_{direction}_{location}" in ds.data_vars
-                assert (
-                    ds[f"child_{direction}_{location}"].attrs["output_period"] == 3600.0
-                )
+                varname = f"child_{direction}_{location}"
+                assert varname in ds.data_vars
+                assert ds[varname].attrs["output_period"] == 3600.0
+
                 if location == "r":
-                    assert (
-                        ds[f"child_{direction}_{location}"].attrs["output_vars"]
-                        == "zeta, temp, salt"
-                    )
+                    assert ds[varname].attrs["output_vars"] == expected_r_vars
                 elif location == "u":
-                    assert (
-                        ds[f"child_{direction}_{location}"].attrs["output_vars"]
-                        == "ubar, u, up"
-                    )
+                    assert ds[varname].attrs["output_vars"] == "ubar, u, up"
                 elif location == "v":
-                    assert (
-                        ds[f"child_{direction}_{location}"].attrs["output_vars"]
-                        == "vbar, v, vp"
-                    )
+                    assert ds[varname].attrs["output_vars"] == "vbar, v, vp"
+
[…]
