### 2026-02-09 3c2c54f
Include child grid fixture in regression tests
