### 2026-02-11 9dc097c
Improve `ChildGrid` boundary defaults and add `include_bgc` metadata option (#550)
* Use open boundaries as default

* Option to include BGC in boundary outputs

* Merge user metadata with defaults

* Test new behaviors

* Document new feature

* Include warning

* Add to release notes

* Include child grid fixture in regression tests

* Add to release notes

* Add new test data

* Remove leftover text in docs
