### 2026-02-13 2c4dedd
Consolidate vertical regridding objects (#528)
* VerticalRegrid for ROMS->ROMS

* Consolidate VerticalRegrid, VerticalRegridFromROMS

* Provide depth to VerticalRegrid as xarray.DataArray

* Introduce source_dim parameter

* Test VerticalRegrid for depth -> ROMS

* Change default

* Include new vertical regridding in BoundaryForcing

* New vertical regridding for initial conditions

* Remove deprecated VerticalRegridToROMS class

* Add to release notes

* Update using_dask notebook
