+++ docs/releases.md
@@ -6,12 +6,14 @@
 
 * Include time records strictly outside start/end bounds for `SurfaceForcing`, `BoundaryForcing` ([#547](https://github.com/CWorthy-ocean/roms-tools/pull/547))
 * `ChildGrid` now infers open boundaries from the mask by default and supports optional BGC boundary outputs via metadata ([#550](https://github.com/CWorthy-ocean/roms-tools/pull/550))
+* Memory savings and speedup for vertical regridding in `InitialConditions` and `BoundaryForcing` ([#528](https://github.com/CWorthy-ocean/roms-tools/pull/528))
 
 ### Breaking Changes
 
 ### Internal Changes
 
 * Regression tests for `ChildGrid` ([#550](https://github.com/CWorthy-ocean/roms-tools/pull/550))
+* Consolidate vertical regridding objects ([#528](https://github.com/CWorthy-ocean/roms-tools/pull/528))
 
 ### Documentation
 
@@ -34,6 +36,7 @@
 ### Breaking Changes
 
 * Remove user-facing control of SSH-adjusted depth coordinates (`adjust_depth_for_sea_surface_height` parameter) for `InitialConditions` and `BoundaryForcing`, enforcing consistent and physically sensible defaults ([#514](https://github.com/CWorthy-ocean/roms-tools/pull/514))
+* Remove `horizontal_chunk_size` parameter from `InitialConditions`; the new vertical regridding scheme is sufficiently memory-efficient without it ([#528](https://github.com/CWorthy-ocean/roms-tools/pull/528))
 
 ### Internal Changes
 
