### 2026-02-13 77c2d23
Rotate ROMS velocities before regridding (#522)
* Rotate velocities giving vel pair names

* Correct regridding

* Formatting

* Refactor rotate_velocities and staggered interpolation

* Add rotate_velocities_to_east_and_north methods to Dataset

* Add blank lines

* Add to release notes

* Add blank line

* Refactor and correct regrid method

* Test more variable combos

* Include u-plot in regridding notebook

* Add to release notes

* Provide comment for why -angle
