### 2026-02-20 #9492 @kthyng
@NoraLoose Thank you for your patience. Your submission has passed all of our original and new scope requirements at a high level so this can move forward to getting an editor.

One note: Please reduce to length of your paper to up to about 1750 words. Paper format information here: https://joss.readthedocs.io/en/latest/paper.html. You can check paper length (and a few other items) with the command `@editorialbot check repository` though note it needs to be the first item in a comment to run correctly.

Do your best on paper length but we are still figuring out if this is adequate.
