### 2026-02-23 #9492 @jatkinson1000
Hi @NoraLoose I'll be editing this for you.

As @kthyng notes, the paper is currently [much longer than we'd like to see in JOSS](https://joss.readthedocs.io/en/latest/paper.html).
Typically we suggest 750-1750 words.

Having read through there are no glaring issues with the content and I'm happy for this to sit at the longer end of this range.
However, it will need reducing a bit please.

To help you I have the following suggestions:

- The processing description could perhaps be detailed in the docs, with a note to say this in the paper instead focussing on any salient points of the process or with less elaboration here.
- Similarly for input-file generation, perhaps provide extensive details in the docs noting here that several required files are generated and highlighting any unique or overarching points.
- The general verbosity could be reduced a little throughout the paper.
- We expect reviewers (and readers) to use and follow quality documentation for the software alongside papers. Consider if anything might be more suitable to be detailed there to leave the main messages in the paper.

Once you've taken a look at this I'm happy to start inviting reviewers.
