+++ docs/index.rst
@@ -24,28 +24,74 @@
 .. image:: https://static.pepy.tech/personalized-badge/roms-tools?period=total&units=INTERNATIONAL_SYSTEM&left_color=BLACK&right_color=GREEN&left_text=downloads
    :target: https://pepy.tech/projects/roms-tools
 
-ROMS-Tools: Preparing and Analyzing ROMS Simulations
-=========================================================================
 
-**ROMS-Tools** is a Python package that automates the creation, management, and analysis of all input and output files needed for regional ocean simulations with `UCLA-ROMS <https://github.com/CWorthy-ocean/ucla-roms>`_, with optional `MARBL biogeochemistry (BGC) <https://marbl-ecosys.github.io/versions/latest_release/index.html>`_.
+ROMS-Tools: Reproducible Preprocessing and Analysis for Regional Ocean Modeling with ROMS
+==========================================================================================
+
+**ROMS-Tools** is a Python package for **preparing and analyzing ROMS simulations**, with optional `MARBL biogeochemistry (BGC) <https://marbl-ecosys.github.io/versions/latest_release/index.html>`_ support.
+
+``ROMS-Tools`` enables users to generate regional grids, prepare model inputs, and analyze outputs through a modern, user-friendly interface that standardizes common workflows and reduces data-preparation overhead. The package is designed for reproducible research, with YAML-based configuration, optional ``dask`` parallelization, interactive Jupyter support, and CI-tested reliability with comprehensive documentation.
+
+Current capabilities are fully compatible with UCLA-ROMS :cite:`ucla-roms,ucla-roms-cworthy`, with potential support for other ROMS implementations, such as Rutgers ROMS :cite:`rutgers-roms`, in the future.
+
+
+Overview of ROMS-Tools Functionality
+--------------------------------------
+
+``ROMS-Tools`` provides a comprehensive workflow for generating, processing, and analyzing ROMS-MARBL model inputs and outputs, as detailed below.
+
+Input Data and Preprocessing
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
 Built on ``xarray`` and optionally powered by ``dask``, ``ROMS-Tools`` automates the generation of all major ROMS–MARBL inputs, including:
 
 1. **Model Grid**: Customizable, curvilinear, and orthogonal grid designed to maintain a nearly uniform horizontal resolution across the domain. The grid is rotatable to align with coastlines and features a terrain-following vertical coordinate.
-2. **Bathymetry**: Derived from **SRTM15**.
-3. **Land Mask**: Inferred from coastlines provided by **Natural Earth** or **GSHHG**.
-4. **Physical Ocean Conditions**:  Initial and open boundary conditions for sea surface height, temperature, salinity, and velocities derived from **GLORYS**.
-5. **BGC Ocean Conditions**: Initial and open boundary conditions for dissolved inorganic carbon, alkalinity, and other BGC tracers from **CESM** output or hybrid observational-model sources.
-6. **Meteorological forcing**: Wind, radiation, precipitation, and air temperature/humidity processed from **ERA5** with optional corrections for radiation bias and coastal wind.
-7. **BGC surface forcing**: Partial pressure of carbon dioxide, as well as iron, dust, and nitrogen deposition from **CESM** output or hybrid observational-model sources.
-8. **Tidal Forcing:** Tidal potential, elevation, and velocities derived from **TPXO** including self-attraction and loading (SAL) corrections.
-9. **River Forcing:** Freshwater runoff derived from **Dai & Trenberth** or user-provided custom files.
[…]
