### 2026-03-02 94ffd2b
Remove badges from docs landing page
