+++ docs/references.bib
@@ -90,7 +90,6 @@
 	abstract = {An updated global bathymetry and topography grid is presented using a spatial sampling interval of 15 arc sec. The bathymetry is produced using a combination of shipboard soundings and depths predicted using satellite altimetry. New data consists of {\textgreater}33.6 million multibeam and singlebeam measurements collated by several institutions, namely, the National Geospatial-Intelligence Agency, Japan Agency for Marine-Earth Science and Technology, Geoscience Australia, Center for Coastal and Ocean Mapping, and Scripps Institution of Oceanography. New altimetry data consists of 48, 14, and 12 months of retracked range measurements from Cryosat-2, SARAL/AltiKa, and Jason-2, respectively. With respect to SRTM15\_PLUS (Olson et al., 2016), the inclusion of these new data results in a ∼1.4-km improvement in the minimum wavelength recovered for sea surface free-air gravity anomalies, a small increase in the accuracy of altimetrically derived predicted depths, and a 1.24\% increase, from 9.60\% to 10.84\%, in the total area of ocean floor that is constrained by shipboard soundings at 15-arc sec resolution. Bathymetric grid cells constrained by satellite altimetry have estimated uncertainties of ±150 m in the deep oceans and ±180 m between coastlines and the continental rise. Onshore, topography data are sourced from previously published digital elevation models, predominately SRTM-CGIAR V4.1 between 60◦N and 60◦S. ArcticDEM is used above 60◦N, while Reference Elevation Model of Antarctica is used below 62◦S. Auxiliary grids illustrating shipboard data coverage, marine free-air gravity anomalies, and vertical gradient gradients are also provided in common data formats.},
 	language = {en},
 	number = {10},
-	urldate = {2024-11-13},
 	journal = {Earth and Space Science},
 	author = {Tozer, B. and Sandwell, D. T. and Smith, W. H. F. and Olson, C. and Beale, J. R. and Wessel, P.},
 	month = oct,
@@ -121,7 +120,6 @@ year = {2020}
 	doi = {10.1175/1520-0426(2002)019<0183:EIMOBO>2.0.CO;2},
[…]
