### 2026-03-02 3a5506f
Remove overview of functionality from paper
