### 2026-03-02 7770665
Add affiliations of more external users
