### 2026-03-02 d32028e
Paper formatting
