### 2026-03-02 f43854c
Paper formatting
