### 2026-03-02 b7a9e6c
Adjust figure caption
