### 2026-03-02 6eedbe5
Move "overview of ROMS-Tools functionality" section from paper to docs (#554)
* Overview of ROMS-Tools functionality moved to docs

* Remove badges from docs landing page

* Clean up references

* Remove overview of functionality from paper

* (auto) Paper PDF Draft

* Add affiliations of more external users

* (auto) Paper PDF Draft

* Paper formatting

* Add to release notes

* (auto) Paper PDF Draft

* Paper formatting

* (auto) Paper PDF Draft

* Adjust figure caption

* (auto) Paper PDF Draft
