### 2026-03-02 #9492 @NoraLoose
Hi @jatkinson1000, I followed your suggestion and moved the entire "Overview of ROMS-Tools functionality" section from the paper to the docs landing page. The paper should be much shorter now. What do you think?
