### 2026-03-06 #9492 @jatkinson1000
Thanks @NoraLoose

That looks a lot better. It's still a little on the long side so I'd encourage you to consider where it might be reduced further, but I'm happy to proceed with looking for reviewers.

It looks like there are [some missing and malformed DOIs in your references](https://github.com/openjournals/joss-reviews/issues/9492#issuecomment-4011834673). For the latter two please could you remove the https from your bib entries. For the former please check the suggested DOIs to see if they are a match for the missing entries and add them to the bib entry igf appropriate.
