### 2026-03-06 48a0b9e
Rework paper
