+++ docs/references.bib
@@ -105,7 +105,7 @@ volume = {146},
 number = {730},
 pages = {1999-2049},
 keywords = {climate reanalysis, Copernicus Climate Change Service, data assimilation, ERA5, historical observations},
-doi = {https://doi.org/10.1002/qj.3803},
+doi = {10.1002/qj.3803},
 url = {https://rmets.onlinelibrary.wiley.com/doi/abs/10.1002/qj.3803},
 eprint = {https://rmets.onlinelibrary.wiley.com/doi/pdf/10.1002/qj.3803},
 abstract = {Abstract Within the Copernicus Climate Change Service (C3S), ECMWF is producing the ERA5 reanalysis which, once completed, will embody a detailed record of the global atmosphere, land surface and ocean waves from 1950 onwards. This new reanalysis replaces the ERA-Interim reanalysis (spanning 1979 onwards) which was started in 2006. ERA5 is based on the Integrated Forecasting System (IFS) Cy41r2 which was operational in 2016. ERA5 thus benefits from a decade of developments in model physics, core dynamics and data assimilation. In addition to a significantly enhanced horizontal resolution of 31 km, compared to 80 km for ERA-Interim, ERA5 has hourly output throughout, and an uncertainty estimate from an ensemble (3-hourly at half the horizontal resolution). This paper describes the general set-up of ERA5, as well as a basic evaluation of characteristics and performance, with a focus on the dataset from 1979 onwards which is currently publicly available. Re-forecasts from ERA5 analyses show a gain of up to one day in skill with respect to ERA-Interim. Comparison with radiosonde and PILOT data prior to assimilation shows an improved fit for temperature, wind and humidity in the troposphere, but not the stratosphere. A comparison with independent buoy data shows a much improved fit for ocean wave height. The uncertainty estimate reflects the evolution of the observing systems used in ERA5. The enhanced temporal and spatial resolution allows for a detailed evolution of weather systems. For precipitation, global-mean correlation with monthly-mean GPCP data is increased from 67\% to 77\%. In general, low-frequency variability is found to be well represented and from 10 hPa downwards general patterns of anomalies in temperature match those from the ERA-Interim, MERRA-2 and JRA-55 reanalyses.},
@@ -134,6 +134,7 @@ year = {2020}
 	shorttitle = {Estimates of {Freshwater} {Discharge} from {Continents}},
 	journal = {Journal of Hydrometeorology},
 	url = {https://journals.ametsoc.org/view/journals/hydr/3/6/1525-7541_2002_003_0660_eofdfc_2_0_co_2.xml},
+	doi = {10.1175/1525-7541(2002)003<0660:EOFDFC>2.0.CO;2},
[…]
