### 2026-03-06 e1b6268
Rework paper (#557)
* Rework paper

* (auto) Paper PDF Draft

* Fix typo

* Fix DOIs

* (auto) Paper PDF Draft
