### 2026-03-06 #9492 @NoraLoose
Thanks @jatkinson1000! I’ve fixed the DOIs and made a few edits to shorten the paper. I may do another round of edits over the next few days to tighten it further.
