### 2026-03-10 ec5db87
ensure ChildGrid mask is int32 (#559)
* ensure ChildGrid mask is int32

* update releases.md

---------

Co-authored-by: Dafydd Stephenson <“dafydd_stephenson@hotmail.com”>
