### 2026-03-12 aad1960
update regionmask dependency version to >=0.11.0 (#565)
Co-authored-by: Dafydd Stephenson <“dafydd_stephenson@hotmail.com”>
