+++ pyproject.toml
@@ -33,7 +33,7 @@ dependencies = [
     "PyYAML",
     "pyamg",
     "bottleneck",
-    "regionmask",
+    "regionmask>=0.11.0",
     "xgcm>=0.9.0",
     "numba>=0.61.2",
     "pydantic>2,<3",
