### 2026-03-13 97356fc
POC of from_file mods
