### 2026-03-13 #9492 @NoraLoose
Hi @jatkinson1000, here are a few additional suggestions for potential reviewers:

* `@manishvenu` (Manish Venumuddula): involved in developing the [regional-mom6](https://github.com/COSIMA/regional-mom6) package,  a package analogous to ROMS-Tools but for MOM6.
* `@ashjbarnes` (Ashley Barnes): involved in developing the [regional-mom6](https://github.com/COSIMA/regional-mom6) package,  a package analogous to ROMS-Tools but for MOM6.
* `@raphaeldussin` (Raphael Dussin): previously contributed to [pyroms](https://github.com/ESMG/pyroms), and wrote [this blog post](https://raphaeldussin.medium.com/modern-python-tools-for-the-roms-ocean-model-bfca8642db01) on modern Python tools for ROMS
* `@hga007` (Hernan G. Arango): one of the main developers of (Rutgers) ROMS
