### 2026-03-13 #9492 @jatkinson1000
Hi there @hklion, @garrettstaller, @esclapez!

Just bumping this request to review a ROMS Oceanography paper for JOSS.
Please let me know if you would be available to review.
