### 2026-03-13 #9492 @jatkinson1000
Hi there @manishvenu, @hga007, and @dcherian!

This paper has been submitted to JOSS, the Journal of Open Source Software.
It is a toolbox for preparing inputs for the ROMS ocean model and analysing the outputs.

Would you be willing to review it?
Please indicate by replying to this thread.

You can find information about reviewing for JOSS [here](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html).

Please note that the scope and review criteria for JOSS has recently changed.
You can read more about this in the https://github.com/openjournals/joss-reviews/issues/9492#issuecomment-3746345306 and the associated links.

If you are unavailable but you know someone who would be a suitable reviewer please do suggest them.
