### 2026-03-13 4c81e9a
removing fill enclosed basins functionality from the close narrow passages functionality
