### 2026-03-13 7ead2bb
store filepath. write to yaml
