### 2026-03-16 #9492 @esclapez
Hi @jatkinson1000, I'd be happy to review it but I'll admit I don't have much experience with ROMS.
