### 2026-03-16 cb02dbf
Refactor `test_close_narrow_channels` tests by removing unused parameters `connectivity` and `min_region_fraction`, simplifying the test setup for better clarity.
