### 2026-03-16 2323595
fixing labeling issue in _fill_enclosed_basins
