### 2026-03-16 0f63e1a
changing the test to accurately reflect the new close_channels functionality
