### 2026-03-16 d1bc86b
Refactor `test_close_narrow_channels` by removing the test implementation to streamline the test suite and improve clarity.
