+++ roms_tools/tests/test_setup/test_grid.py
@@ -879,77 +879,6 @@ def test_close_narrow_channels():
     assert "mask_v" in grid.ds.variables
 
 
-def test_close_narrow_channels():
-    """Test that close_narrow_channels fills narrow channels.
-
-    """
-    # Create a small grid with close_narrow_channels=False to avoid closing during init
-    grid = Grid(
-        nx=20,
-        ny=20,
-        size_x=100,
-        size_y=100,
-        center_lon=-20,
-        center_lat=64,
-        rot=0,
-        N=3,
-        close_narrow_channels=False,
-    )
-
-    # Get the actual shape of mask_rho
-    mask_shape = grid.ds["mask_rho"].shape
-
-    # Create a mask with a large ocean region and a small isolated lake (water region)
-    mask = np.zeros(mask_shape, dtype=np.int32)  # Start with all land (0)
-
-    # Create a large ocean region (1) - this should be preserved as the largest region
-    # Account for boundary cells: indices are shifted by 1
-    mask[6:16, 6:16] = 1  # Large ocean region (5+1:15+1, 5+1:15+1)
-
-    # Create a small isolated lake (water region, 1) - this should be filled (converted to land)
-    mask[3:5, 3:5] = 1  # Small 2x2 isolated lake (2+1:4+1, 2+1:4+1)
-
-    # Set the mask in the grid
-    grid.ds["mask_rho"].values[:] = mask
-
-    # Save original mask
-    mask_before = grid.ds.mask_rho.values.copy()
-
-    # Verify the small isolated lake exists before closing
-    assert mask_before[3:5, 3:5].all() == 1, (
-        "Small isolated lake should exist before closing"
-    )
-
-    # Close narrow channels directly (as update_mask would do)
-    # The small isolated lake should be filled (converted to land), but the large ocean region should be preserved
-    grid.ds = _close_narrow_channels(
-        grid.ds,
-        mask_var="mask_rho",
-        max_iterations=10,
-        inplace=True,
-    )
-    # Update velocity masks after modifying mask_rho
-    grid.ds = add_velocity_masks(grid.ds)
-
-    # Get the mask after closing
-    mask_after = grid.ds.mask_rho.values.copy()
-
-    # Verify that channel was closed
-    assert mask_after[5:7, 7:8].all() == 0, (
-        "Narrow channel should be closed"
-    )
-
-    # Verify that the large ocean region was preserved
-    assert mask_after[6:16, 0:16].sum() == 100, "Large ocean region should be preserved"
-
-    # Verify that the mask changed
-    assert np.any(mask_before != mask_after), "Mask should have changed after closing"
-
-    # Verify that velocity masks were updated
-    assert "mask_u" in grid.ds.variables
-    assert "mask_v" in grid.ds.variables
-
-
 # Boundary tests
 
 
