### 2026-03-18 007a122
todos and delete some code we don't want
