+++ roms_tools/setup/grid.py
@@ -1,7 +1,6 @@
 import copy
 import importlib.metadata
 import logging
-import re
 from dataclasses import asdict, dataclass, field
 from pathlib import Path
 from typing import Any
@@ -17,7 +16,6 @@ from roms_tools.setup.mask import add_mask, add_velocity_masks
 from roms_tools.setup.topography import add_topography
 from roms_tools.setup.utils import (
     Timed,
-    extract_single_value,
     gc_dist,
     get_target_coords,
     pop_grid_data,
@@ -121,6 +119,10 @@ class Grid:
     verbose: bool = False
     """Whether to print grid generation steps with timing."""
     filename: str | Path | None = None
+    """An external file to load instead of generating the grid from params"""
+
+    # TODO: consider adding a "load_clean" param that forces us to JUST load and do nothing else.
+    # we may want to ask if this is actually needed before implementing
 
     ds: xr.Dataset = field(init=False, repr=False)
     """An xarray Dataset containing post-processed variables ready for input into
@@ -130,9 +132,9 @@ class Grid:
 
     def __post_init__(self):
         # TODO: add checks that raises error if you supply both a filename
-        # and the nx/ny/etc. args; only allow one or the other
+        #  and the nx/ny/etc. args; only allow one or the other
         if self.filename is not None:
-            self = self.from_file(
+            self = Grid.from_file(
                 filepath=self.filename,
                 theta_s=self.theta_s,
                 theta_b=self.theta_b,
@@ -147,56 +149,61 @@ class Grid:
             self.theta_b = self.theta_b or 2.0
             self.hc = self.hc or 300.0
 
-            self._input_checks()
+        self._input_checks()
 
-            # Horizontal grid
-            self._create_horizontal_grid()
+        # Horizontal grid
+        # todo don't do this for external file
+        self._create_horizontal_grid()
 
-            # Check if the Greenwich meridian goes through the domain.
-            self._straddle()
+        # Check if the Greenwich meridian goes through the domain.
+        self._straddle()
 
-            # Mask
-            self.update_mask(mask_shapefile=self.mask_shapefile, verbose=self.verbose)
+        # Mask
+        # todo don't do this for external file
+        self.update_mask(mask_shapefile=self.mask_shapefile, verbose=self.verbose)
 
-            # Coarsen the dataset if needed
-            self._coarsen()
+        # Coarsen the dataset if needed
+        self._coarsen()
 
-            # Topography
-            self.update_topography(
-                topography_source=self.topography_source,
-                hmin=self.hmin,
-                verbose=self.verbose,
-            )
+        # Topography
+        # todo don't do this for external file
+        self.update_topography(
+            topography_source=self.topography_source,
+            hmin=self.hmin,
+            verbose=self.verbose,
+        )
 
-            # Vertical coordinate system
-            self.update_vertical_coordinate(
-                N=self.N,
-                theta_s=self.theta_s,
-                theta_b=self.theta_b,
-                hc=self.hc,
-                verbose=self.verbose,
-            )
+        # Vertical coordinate system
+
+        self.update_vertical_coordinate(
+            N=self.N,
+            theta_s=self.theta_s,
+            theta_b=self.theta_b,
+            hc=self.hc,
+            verbose=self.verbose,
+        )
 
     def _input_checks(self):
-        for var in ("nx", "ny", "size_x", "size_y", "center_lon", "center_lat"):
-            if getattr(self, var) is None:
+        if not self.filename:
+            for var in ("nx", "ny", "size_x", "size_y", "center_lon", "center_lat"):
+                if getattr(self, var) is None:
+                    raise ValueError(
[…]
