### 2026-03-18 d1be3af
add optional 'include_pressure_fluxes' key to ChildGrid metadata dictionary to toggle 'up','vp' in 'output_vars'
