### 2026-03-18 c6000fc
make from_file private and clean up vertical stuff
