### 2026-03-19 #9492 @jatkinson1000
Thanks @esclapez
I don't believe excessive knowledge of ROMS is required - the package is Python-based and focusses on preparing and post-processing data, but perhaps @NoraLoose can confirm how much running/using of ROMS would be required to review.
