### 2026-03-19 #9492 @NoraLoose
> perhaps @NoraLoose can confirm how much running/using of ROMS would be required to review.

None required. The documentation includes a single [end-to-end example](https://roms-tools.readthedocs.io/en/latest/end_to_end.html) that prepares the inputs, runs ROMS in a container (with a pre-compiled executable), and analyzes the outputs. To rerun this workflow, you only need to install `podman`. All other examples focus on generating inputs or analyzing existing outputs.
