### 2026-03-19 b4a90d8
add optional 'include_pressure_fluxes' key to ChildGrid metadata dictionary to toggle 'up','vp' in 'output_vars' (#568)
Co-authored-by: Dafydd Stephenson <“dafydd_stephenson@hotmail.com”>
