### 2026-03-19 cd43787
Slight refactor of update_mask
