+++ roms_tools/setup/mask.py
@@ -1,4 +1,5 @@
 import logging
+import time
 import warnings
 from pathlib import Path
 
@@ -16,11 +17,19 @@ from roms_tools.utils import (
     interpolate_from_rho_to_v,
 )
 
-logger = logging.getLogger(__name__)
 
-
-def add_mask(ds: xr.Dataset, shapefile: str | Path | None = None) -> xr.Dataset:
+def add_mask(
+    ds: xr.Dataset,
+    shapefile: str | Path | None = None,
+    close_narrow_channels: bool = False,
+    verbose: bool = False,
+) -> xr.Dataset:
     """Adds a land/water mask to the dataset at rho-points.
+
+    These are the steps:
+    1. Infer mask from coastlines
+    2. Close narrow channels if requested
+    3. Fill enclosed basins
 
     Parameters
     ----------
@@ -30,12 +39,21 @@ def add_mask(ds: xr.Dataset, shapefile: str | Path | None = None) -> xr.Dataset:
     shapefile: str or Path | None
         Path to a coastal shapefile to determine the land mask. If None, NaturalEarth 10m is used.
 
+    close_narrow_channels : bool, optional
+        Whether to close narrow water channels in the mask after it is generated.
+        The default is False.
+
+    verbose: bool, optional
+        Indicates whether to print mask generation steps with timing. Defaults to False.
+
     Returns
     -------
     xarray.Dataset
         The original dataset with an added 'mask_rho' variable, representing land/water mask.
     """
-    # Suppress specific warning
+    # Infer mask from coastlines
+    if verbose:
+        start_time = time.time()
     with warnings.catch_warnings():
         warnings.filterwarnings(
             "ignore", message="No gridpoint belongs to any region.*"
@@ -68,9 +86,33 @@ def add_mask(ds: xr.Dataset, shapefile: str | Path | None = None) -> xr.Dataset:
             mask = land_mask.isnull()
 
     ds = _add_coastlines_metadata(ds, shapefile)
+    if verbose:
+        logging.info(
+            f"Inferring the mask from coastlines: {time.time() - start_time:.3f} seconds"
+        )
+
+    # Close narrow channels if requested
+    if close_narrow_channels:
+        if verbose:
+            start_time = time.time()
+        _close_narrow_channels(
+            mask,
+            max_iterations=10,
+        )
+        if verbose:
+            logging.info(
+                "Closing narrow channels: {time.time() - start_time:.3f} seconds"
+            )
 
     # fill enclosed basins with land
+    if verbose:
+        start_time = time.time()
     mask = _fill_enclosed_basins(mask.values)
+    if verbose:
+        logging.info(
+            "Filling enclosed basins with land: {time.time() - start_time:.3f} seconds"
+        )
+
     # adjust mask boundaries by copying values from adjacent cells
     mask = handle_boundaries(mask)
 
@@ -108,7 +150,7 @@ def _add_coastlines_metadata(
     return ds
 
 
-def _fill_enclosed_basins(mask) -> np.ndarray:
+def _fill_enclosed_basins(mask: np.ndarray) -> np.ndarray:
     """Fills enclosed basins in the mask with land (value = 0).
 
     This function identifies the largest connected region in the mask, which is assumed to represent
@@ -143,7 +185,7 @@ def _fill_enclosed_basins(mask) -> np.ndarray:
     return mask
 
 
-def add_velocity_masks(ds):
+def add_velocity_masks(ds: xr.Dataset) -> xr.Dataset:
     """Adds velocity masks for u- and v-points based on the rho-point mask.
 
     This function generates masks for u- and v-points by interpolating the rho-point land/water mask.
@@ -175,12 +217,9 @@ def add_velocity_masks(ds):
 
 
 def _close_narrow_channels(
[…]
