+++ roms_tools/datasets/roms_dataset.py
@@ -246,7 +246,6 @@ class ROMSDataset:
             use_dask=self.use_dask,
             decode_times=False,
             decode_timedelta=False,
-            time_chunking=True,
             force_combine_nested=True,
         )
 
