### 2026-03-23 5eb05ce
.rechunk() --> .chunk()
