+++ roms_tools/setup/nesting.py
@@ -9,17 +9,16 @@ import xarray as xr
 from scipy.interpolate import griddata, interp1d
 
 from roms_tools import Grid
-from roms_tools.plot import plot_nesting
+#from roms_tools.plot import plot_nesting
 from roms_tools.setup.topography import clip_depth
 from roms_tools.setup.utils import (
     Timed,
     check_and_set_boundaries,
-    from_yaml,
     get_boundary_coords,
     pop_grid_data,
-    to_dict,
-    write_to_yaml,
 )
+#    to_dict,
+
 from roms_tools.utils import (
     interpolate_from_rho_to_u,
     interpolate_from_rho_to_v,
@@ -76,272 +75,120 @@ class ChildGrid(Grid):
     """An xarray Dataset containing boundary mappings, where child grid boundaries are
     mapped onto parent grid indices."""
 
-    def _map_child_boundaries_onto_parent_grid_indices(self, verbose: bool = False):
-        """Maps child grid boundary points onto absolute indices of the parent grid."""
-        with Timed(
-            "=== Mapping the child grid boundary points onto the indices of the parent grid ===",
-            verbose=verbose,
-        ):
-            # Default metadata
-            defaults = {
-                "prefix": "child",
-                "period": 3600.0,
-                "include_bgc": False,
-            }
-            # Merge user metadata on top of defaults
-            self.metadata = {**defaults, **self.metadata}
+#    def update_mask(
+#        self, mask_shapefile: str | Path | None = None, verbose: bool = False
+#    ) -> None:
+#        """
+#        Update the child grid mask and ensure consistency with the parent grid.
+#
+#        This method performs the following steps:
+#
+#        1. Derives the child mask from the provided ``mask_shapefile`` (or from the
+#           default Natural Earth 10m coastline if ``None``).
+#        2. Updates the mapping of child boundaries to parent-grid indices.
+#           This mapping depends on the updated mask, since masked (land) points may
+#           extend outside the parent grid.
+#        3. Adjusts the child mask to ensure consistency with the parent mask.
+#
+#        Parameters
+#        ----------
+#        mask_shapefile : str or Path, optional
+#            Path to a coastal shapefile used to derive the land mask. If ``None``,
+#            a default coastline dataset is used.
+#        verbose : bool, default False
+#            If True, prints timing and progress information.
+#
+#        Returns
+#        -------
+#        None
+#            Updates the internal datasets (``self.ds`` and ``self.ds_nesting``) in place,
+#            modifying the mask and ensuring consistent parent-child boundary mapping.
+#        """
+#        super().update_mask(mask_shapefile=mask_shapefile, verbose=verbose)
+#        self._map_child_boundaries_onto_parent_grid_indices(verbose=verbose)
+#        self._modify_child_mask(verbose=verbose)
 
-            self.boundaries = check_and_set_boundaries(
-                self.boundaries, self.ds.mask_rho
-            )
+#    def update_topography(
+#        self,
+#        topography_source: dict | None = None,
+#        hmin: float | None = None,
+#        verbose: bool = False,
+#    ) -> None:
+#        """
+#        Update the child grid topography and ensure consistency with the parent grid.
+#
+#        This method performs the following operations:
+#
+#        1. Regrids the topography from the specified source.
+#        2. Applies domain-wide and local smoothing.
+#        3. Enforces the minimum depth constraint ``hmin``.
+#        4. Adjusts the child grid topography to maintain consistency with the
+#           parent grid.
+#        5. Updates the internal dataset (``self.ds``) with the processed bathymetry.
+#
+#        Parameters
+#        ----------
+#        topography_source : dict, optional
[…]
