+++ roms_tools/setup/nesting_SE.py
@@ -1,992 +1,0 @@
-import logging
-from collections.abc import Callable
-from dataclasses import dataclass, field
-from pathlib import Path
-from typing import Any
-
-import numpy as np
-import xarray as xr
-from scipy.interpolate import griddata, interp1d
-
-from roms_tools import Grid
-from roms_tools.plot import plot_nesting
-from roms_tools.setup.topography import clip_depth
-from roms_tools.setup.utils import (
-    Timed,
-    check_and_set_boundaries,
-    from_yaml,
-    get_boundary_coords,
-    pop_grid_data,
-    to_dict,
-    write_to_yaml,
-)
-from roms_tools.utils import (
-    interpolate_from_rho_to_u,
-    interpolate_from_rho_to_v,
-    save_datasets,
-    wrap_longitudes,
-)
-
-
-@dataclass(kw_only=True)
-class ChildGrid(Grid):
-    """Represents a ROMS child grid that is compatible with the provided parent grid.
-
-    This class establishes the relationship between a parent grid and a child grid in ROMS simulations.
-    It generates two datasets:
-
-    1. `ds`: Contains child grid variables, ensuring compatibility with the parent grid.
-       The child grid topography and mask are adjusted to match the parent grid at the boundaries.
-
-    2. `ds_nesting`: Contains boundary mappings, linking the boundaries of the child grid
-       to the corresponding parent grid indices.
-
-    Parameters
-    ----------
-    parent_grid : Grid
-        The parent grid object, providing the reference for the topography and mask of the child grid.
-    boundaries : Dict[str, bool]
-        Specifies which child grid boundaries ('south', 'east', 'north', 'west') should be adjusted for topography/mask
-        and included in `ds_nesting`. If not provided, valid (non-land) boundaries are enabled automatically.
-    metadata : Dict[str, Any]
-        Dictionary configuring the boundary nesting process, including:
-
-        - `"prefix"` (str): Prefix for variable names in `ds_nesting`. Defaults to `"child"`.
-        - `"period"` (float): Temporal resolution for boundary outputs in seconds. Defaults to 3600 (hourly).
-        - `"include_bgc"` (bool): Whether to include BGC variables in boundary outputs. Defaults to `False`.
-    verbose: bool, optional
-        Indicates whether to print grid generation steps with timing. Defaults to False.
-    """
-
-    parent_grid: Grid
-    """The parent grid object, providing the reference for the child topography
-    and mask of the child grid."""
-    boundaries: dict[str, bool] | None = None
-    """Specifies which child grid boundaries (south, east, north, west) should be
-    adjusted for topography/mask and included in `ds_nesting`."""
-    metadata: dict[str, Any] = field(default_factory=dict)
-    """Dictionary configuring the boundary nesting process."""
-    verbose: bool = False
-    """Whether to print grid generation steps with timing."""
-
-    ds: xr.Dataset = field(init=False, repr=False)
-    """An xarray Dataset containing child grid variables aligned with the
-    topography and mask of the parent grid at the boundaries."""
-    ds_nesting: xr.Dataset = field(init=False, repr=False)
-    """An xarray Dataset containing boundary mappings, where child grid boundaries are
-    mapped onto parent grid indices."""
-
-    def _map_child_boundaries_onto_parent_grid_indices(self, verbose: bool = False):
-        """Maps child grid boundary points onto absolute indices of the parent grid."""
-        with Timed(
-            "=== Mapping the child grid boundary points onto the indices of the parent grid ===",
-            verbose=verbose,
-        ):
-            # Default metadata
-            defaults = {
-                "prefix": "child",
-                "period": 3600.0,
-                "include_bgc": False,
-            }
-            # Merge user metadata on top of defaults
[…]
