### 2026-03-24 233f8db
Adapt _close_narrow_channels for xr.DataArray
