### 2026-03-24 5683fda
Convert mask from bool to int before filling channels/basins
