### 2026-03-24 0fb5369
Make sure closing channels works with from_file
