### 2026-03-24 24ec1f8
Add integration tests with filling closed channels
