### 2026-03-24 71de017
nesting done. w/ typehints
