+++ roms_tools/setup/nesting.py
@@ -26,146 +26,6 @@ from roms_tools.utils import (
     wrap_longitudes,
 )
 
-
-@dataclass(kw_only=True)
-class ChildGrid(Grid):
-    """Represents a ROMS child grid that is compatible with the provided parent grid.
-
-    This class establishes the relationship between a parent grid and a child grid in ROMS simulations.
-    It generates two datasets:
-
-    1. `ds`: Contains child grid variables, ensuring compatibility with the parent grid.
-       The child grid topography and mask are adjusted to match the parent grid at the boundaries.
-
-    2. `ds_nesting`: Contains boundary mappings, linking the boundaries of the child grid
-       to the corresponding parent grid indices.
-
-    Parameters
-    ----------
-    parent_grid : Grid
-        The parent grid object, providing the reference for the topography and mask of the child grid.
-    boundaries : Dict[str, bool]
-        Specifies which child grid boundaries ('south', 'east', 'north', 'west') should be adjusted for topography/mask
-        and included in `ds_nesting`. If not provided, valid (non-land) boundaries are enabled automatically.
-    metadata : Dict[str, Any]
-        Dictionary configuring the boundary nesting process, including:
-
-        - `"prefix"` (str): Prefix for variable names in `ds_nesting`. Defaults to `"child"`.
-        - `"period"` (float): Temporal resolution for boundary outputs in seconds. Defaults to 3600 (hourly).
-        - `"include_bgc"` (bool): Whether to include BGC variables in boundary outputs. Defaults to `False`.
-    verbose: bool, optional
-        Indicates whether to print grid generation steps with timing. Defaults to False.
-    """
-
-    parent_grid: Grid
-    """The parent grid object, providing the reference for the child topography
-    and mask of the child grid."""
-    boundaries: dict[str, bool] | None = None
-    """Specifies which child grid boundaries (south, east, north, west) should be
-    adjusted for topography/mask and included in `ds_nesting`."""
-    metadata: dict[str, Any] = field(default_factory=dict)
-    """Dictionary configuring the boundary nesting process."""
-    verbose: bool = False
-    """Whether to print grid generation steps with timing."""
-
-    ds: xr.Dataset = field(init=False, repr=False)
-    """An xarray Dataset containing child grid variables aligned with the
-    topography and mask of the parent grid at the boundaries."""
-    ds_nesting: xr.Dataset = field(init=False, repr=False)
-    """An xarray Dataset containing boundary mappings, where child grid boundaries are
-    mapped onto parent grid indices."""
-
-#    def update_mask(
-#        self, mask_shapefile: str | Path | None = None, verbose: bool = False
-#    ) -> None:
-#        """
-#        Update the child grid mask and ensure consistency with the parent grid.
-#
-#        This method performs the following steps:
-#
-#        1. Derives the child mask from the provided ``mask_shapefile`` (or from the
-#           default Natural Earth 10m coastline if ``None``).
-#        2. Updates the mapping of child boundaries to parent-grid indices.
-#           This mapping depends on the updated mask, since masked (land) points may
-#           extend outside the parent grid.
-#        3. Adjusts the child mask to ensure consistency with the parent mask.
-#
-#        Parameters
-#        ----------
-#        mask_shapefile : str or Path, optional
-#            Path to a coastal shapefile used to derive the land mask. If ``None``,
-#            a default coastline dataset is used.
-#        verbose : bool, default False
-#            If True, prints timing and progress information.
-#
-#        Returns
-#        -------
-#        None
[…]
