### 2026-03-24 f467d2f
Make sure bool --> int conversion happens for all shapefiles
