### 2026-03-24 d2a2011
Make sure that channel closing works for GSHHS topo as well
