### 2026-03-25 cd32297
Optional reset of chunking on choose_subdomain()
We should be able to load a slice of the whole, and optionally reset chunking to require the whole subdomain (if subsequent operations require the full stride of the subdomain are needed (like 2D/3D regridding, interpolation, etc.)
