### 2026-03-25 d0bfef4
Fixes for non-class use of roms_dataset choose_domain()
