+++ roms_tools/setup/boundary_forcing.py
@@ -505,8 +505,7 @@ class BoundaryForcing:
         if isinstance(self.source["path"], bool):
             raise ValueError('source["path"] cannot be a boolean here')
 
-        # TODO: Evaluate whether this chunking strategy improves performance
-        # performance gains/losses may also differ for GLORYS vs. BGC dataset
+        # Leave initial chunking to dask for efficient sliced reading from file
         chunks = {"time": 1}
 
         # Alternative: chunks=None (current behavior), which results in
