### 2026-03-25 2ff5050
2 more t dos. handles filename initiation + to_yaml, from_yaml
