+++ roms_tools/setup/grid.py
@@ -17,6 +17,7 @@ from roms_tools.setup.mask import add_mask, add_velocity_masks
 from roms_tools.setup.topography import add_topography
 from roms_tools.setup.utils import (
     Timed,
+    extract_single_value,
     gc_dist,
     get_target_coords,
     pop_grid_data,
@@ -138,7 +139,15 @@ class Grid:
         # TODO: add checks that raises error if you supply both a filename
         #  and the nx/ny/etc. args; only allow one or the other
         if self.filename is not None:
-            self = self._from_file(
+            # Check to confirm other parameters are not provided with filename
+            min_param = {"rot", "hmin", "verbose", "filename"}
+            param_provided = [k for k, v in vars(self).items() if v is not None]
+            extra = set(param_provided) - min_param
+            if extra:
+                raise ValueError(f"Unexpected parameter included with `filename`: {extra}.\n" +
+                    f" Either `filename` or (`nx`, `ny`, `size_x`, `size_y`, `center_lon`, `center_lat`) are accepted.")
+
+            loaded = type(self)._from_file(
                 filepath=self.filename,
                 theta_s=self.theta_s,
                 theta_b=self.theta_b,
@@ -146,6 +155,9 @@ class Grid:
                 N=self.N,
                 verbose=self.verbose,
             )
+
+            self.__dict__.update(loaded.__dict__)
+
         else:
             # assign defaults here for non-required params; need Nones in the above section
             self.N = self.N or 100
@@ -153,29 +165,29 @@ class Grid:
             self.theta_b = self.theta_b or 2.0
             self.hc = self.hc or 300.0
 
-        self._input_checks()
+            self._input_checks()
 
-        # Horizontal grid
-        # todo don't do this for external file
-        self._create_horizontal_grid()
+            # Horizontal grid
+            # todo don't do this for external file
+            self._create_horizontal_grid()
 
-        # Check if the Greenwich meridian goes through the domain.
-        self._straddle()
+            # Check if the Greenwich meridian goes through the domain.
+            self._straddle()
 
-        # Mask
-        # todo don't do this for external file
-        self.update_mask(mask_shapefile=self.mask_shapefile, verbose=self.verbose)
+            # Mask
+            # todo don't do this for external file
+            self.update_mask(mask_shapefile=self.mask_shapefile, verbose=self.verbose)
 
-        # Coarsen the dataset if needed
-        self._coarsen()
+            # Coarsen the dataset if needed
+            self._coarsen()
 
-        # Topography
-        # todo don't do this for external file
-        self.update_topography(
-            topography_source=self.topography_source,
-            hmin=self.hmin,
-            verbose=self.verbose,
-        )
+            # Topography
+            # todo don't do this for external file
+            self.update_topography(
+                topography_source=self.topography_source,
+                hmin=self.hmin,
+                verbose=self.verbose,
+            )
 
         # Vertical coordinate system
 
@@ -188,26 +200,25 @@ class Grid:
         )
 
     def _input_checks(self):
-        if not self.filename:
-            for var in ("nx", "ny", "size_x", "size_y", "center_lon", "center_lat"):
-                if getattr(self, var) is None:
-                    raise ValueError(
-                        f"{var} is required unless loading from an existing grid."
-                    )
-
-            if self.topography_source is None:
-                self.topography_source = {"name": "ETOPO5"}
-
-            if "name" not in self.topography_source:
+        for var in ("nx", "ny", "size_x", "size_y", "center_lon", "center_lat"):
+            if getattr(self, var) is None:
                 raise ValueError(
[…]
