### 2026-03-26 fc4e80d
Remove chunking inside lateralFill - it already happens upstream if needed
