+++ roms_tools/datasets/lat_lon_datasets.py
@@ -628,9 +628,6 @@ class LatLonDataset:
         point to the same variable in the dataset.
         """
         if self.needs_lateral_fill:
-            self.ds = self.ds.chunk(
-                {self.dim_names["latitude"]: -1, self.dim_names["longitude"]: -1}
-            )
 
             lateral_fill = LateralFill(
                 self.ds["mask"],
