+++ paper.md
@@ -42,6 +42,9 @@ authors:
   - name: M. Jeroen Molemaker
     orcid:
     affiliation: 3
+  - name: Benjamin Saenz
+    orcid:
+    affiliation: 1
   - name: Abigale Wyatt
     orcid: 0000-0003-2428-4370
     affiliation: 1
