+++ paper.md
@@ -43,7 +43,7 @@ authors:
     orcid:
     affiliation: 3
   - name: Benjamin Saenz
-    orcid:
+    orcid: 0000-0003-3724-3033
     affiliation: 1
   - name: Abigale Wyatt
     orcid: 0000-0003-2428-4370
