### 2026-03-26 d9e8c1c
metadata removed. does not handle Nones, needed?
