### 2026-03-26 b3cfd3e
made saving ds_nesting optional
