### 2026-03-26 8a7c869
Re-implement pressure flux setting
