+++ roms_tools/setup/nesting.py
@@ -197,6 +197,7 @@ def make_edata(
     prefix: str = "child",
     period: float = 3600.0,
     include_bgc: bool = False,
+    include_pressure_fluxes: bool = False,
     boundaries: dict | None = None,
     verbose: bool = False,
 ) -> xr.Dataset:
@@ -228,6 +229,8 @@ def make_edata(
         Temporal resolution for boundary outputs in seconds. Defaults to 3600 (hourly).
     include_bgc : bool
         Whether to include BGC variables in boundary outputs. Defaults to `False`.
+    include_pressure_fluxes: bool, optional
+        Whether to include baroclinic pressure fluxes in boundary outputs. Defaults to `False`.
     boundaries : dict, optional
         A dictionary specifying which child boundaries should be mapped onto the parent grid.
         Keys should be `"south"`, `"east"`, `"north"`, and `"west"`, with boolean values
@@ -253,7 +256,8 @@ def make_edata(
         prefix: str = "child",
         period: float = 3600.0,
         include_bgc: bool = False,
-        verbose: bool = False,
+        include_pressure_fluxes: bool = False,
+        verbose: bool = False
     ) -> xr.Dataset:
         """Maps child grid boundary points onto absolute indices of the parent grid."""
         with Timed(
@@ -273,6 +277,7 @@ def make_edata(
                 prefix,
                 period,
                 include_bgc,
+                include_pressure_fluxes
             )
 
             return ds_nesting
@@ -672,6 +677,7 @@ def map_child_boundaries_onto_parent_grid_indices(
     prefix: str = "child",
     period: float = 3600.0,
     include_bgc: bool = False,
+    include_pressure_fluxes: bool = False,
     update_land_indices: bool = True,
 ) -> xr.Dataset:
     """Maps child grid boundary points onto absolute indices of the parent grid.
@@ -705,6 +711,9 @@ def map_child_boundaries_onto_parent_grid_indices(
 
     include_bgc: bool, optional
         Whether to include BGC variables in the boundary outputs.
+
+    include_pressure_fluxes: bool, optional
+        Whether to include baroclinic pressure fluxes in boundary outputs. Defaults to `False`.
 
     update_land_indices : bool, optional
         If `True`, updates indices that fall on land in the parent grid to nearby ocean points.
@@ -782,9 +791,13 @@ def map_child_boundaries_onto_parent_grid_indices(
                     ds[var_name].attrs["units"] = "non-dimensional and radian"
 
                     if grid_location == "u":
-                        ds[var_name].attrs["output_vars"] = "ubar, u, up"
+                        ds[var_name].attrs["output_vars"] = (
+                            "ubar, u, up" if include_pressure_fluxes else "ubar, u"
+                        )
                     elif grid_location == "v":
-                        ds[var_name].attrs["output_vars"] = "vbar, v, vp"
+                        ds[var_name].attrs["output_vars"] = (
+                            "vbar, v, vp" if include_pressure_fluxes else "vbar, v"
+                        )
 
                 ds[var_name].attrs["output_period"] = period
 
