+++ roms_tools/setup/grid.py
@@ -1,6 +1,8 @@
 import copy
 import importlib.metadata
 import logging
+import os
+import re
 from dataclasses import asdict, dataclass, field
 from pathlib import Path
 from typing import TYPE_CHECKING, Any
@@ -8,7 +10,6 @@ from typing import TYPE_CHECKING, Any
 import numpy as np
 import xarray as xr
 import yaml
-import os
 from matplotlib.axes import Axes
 
 from roms_tools.constants import MAXIMUM_GRID_SIZE, R_EARTH
@@ -31,7 +32,7 @@ from roms_tools.utils import (
 from roms_tools.vertical_coordinate import compute_depth_coordinates, sigma_stretch
 
 if TYPE_CHECKING:
-    from nesting import ChildGrid
+    pass
 
 
 @dataclass(kw_only=True)
@@ -139,8 +140,10 @@ class Grid:
             param_provided = [k for k, v in vars(self).items() if v is not None]
             extra = set(param_provided) - min_param
             if extra:
-                raise ValueError(f"Unexpected parameter included with `filename`: {extra}.\n" +
-                    f" Either `filename` or (`nx`, `ny`, `size_x`, `size_y`, `center_lon`, `center_lat`) are accepted.")
+                raise ValueError(
+                    f"Unexpected parameter included with `filename`: {extra}.\n"
+                    + " Either `filename` or (`nx`, `ny`, `size_x`, `size_y`, `center_lon`, `center_lat`) are accepted."
+                )
 
             loaded = type(self)._from_file(
                 filepath=self.filename,
@@ -445,7 +448,7 @@ class Grid:
         """
         # todo check for these keys
         # NOTE: THERE IS ALREADY A CHECK FOR ALL OR NONE TO DO COARSEN UNDER FROM_FILE
-        # THERE'S NO CHECK IN THE post_init, but presumably, we need the coarsening to happen then. 
+        # THERE'S NO CHECK IN THE post_init, but presumably, we need the coarsening to happen then.
         # i.e. this check may be unecessary unless we want individual checks:
         d = {
             "angle": "angle_coarse",
@@ -455,7 +458,7 @@ class Grid:
         }
 
         # Only coarsen variables that have not yet been coarsened
-        vars_coarsen = [(k,v) for k,v in d.items() if v not in self.ds]
+        vars_coarsen = [(k, v) for k, v in d.items() if v not in self.ds]
 
         ds = self.ds
 
@@ -788,7 +791,7 @@ class Grid:
                     "Could not extract 'center_lon' from title attribute. "
                     "Expected format: '... Lon: <value> ...'"
                 )
-        elif self.filename is not None:
+        elif filepath is not None:
             center_lon = None
         else:
             raise ValueError(
@@ -810,7 +813,7 @@ class Grid:
                     "Could not extract 'center_lat' from title attribute. "
                     "Expected format: '... Lon: <value> ...'"
                 )
-        elif self.filename is not None:
+        elif filepath is not None:
             center_lat = None
         else:
             raise ValueError(
@@ -832,8 +835,8 @@ class Grid:
                     "Could not extract 'rot' from title attribute. "
                     "Expected format: '... rotate: <value> ...'"
                 )
-        elif self.filename is not None:
-            rot= None
+        elif filepath is not None:
+            rot = 0
         else:
             raise ValueError(
                 "Missing grid information: 'rot' attribute, 'rotate' variable, or 'rotate:' in 'title' attribute "
@@ -888,10 +891,12 @@ class Grid:
 
         # If parameters needed for ROMS-Tools replication are missing, store only filepath
         # Only if externally-generated grid
-        if any(data[k] is None for k in ['size_x','size_y','topography_source','hmin']):
[…]
