### 2026-03-26 f92e0fa
Fix enclosed basin fill off-by-one
