### 2026-03-26 d3c202c
Clean up of PR comments, add docstrings for chunking functionality
