### 2026-03-27 #10234 @jatkinson1000
Hi All, just doing a weekly check-in here.

It seems like things are underway for @esclapez, thank you.

@manishvenu please could you give us an update on the state of your review?
I see you have generated your checklist but only checked off the inital items so far.

@dcherian I see you haven't generated your checklist yet using `@editorialbot generate my checklist`.
Please could you do this and make a start through the items

If you could please find some time to work through reviews and raise any feedback for the authors over the next 2 weeks that would be appreciated, thank you. JOSS tries to aim for a 4-week turnaround in reviews, including time for authors to address/implement feedback.

Any questions please ask.
