### 2026-03-27 28e3755
nesting nb docs and check wet points fnctn
