### 2026-03-27 530e8a2
Update releases.md
adding mask bugfix related to number of ocean regions
