### 2026-03-27 ddb2032
typo in release notes
