+++ docs/releases.md
@@ -26,7 +26,7 @@
 
 * Rotate ROMS velocities before regridding ([#522](https://github.com/CWorthy-ocean/roms-tools/pull/522))
 
-* Corrected enclosed-basin filling in mask generation by iterating connected-component labels `1..nreg` in `_fill_enclosed_basins`, preventing spurious interior lakes; updated the enclosed-region test to expect a single connected wet region. ([#522] https://github.com/CWorthy-ocean/roms-tools/pull/577)
+* Corrected enclosed-basin filling in mask generation by iterating connected-component labels `1..nreg` in `_fill_enclosed_basins`, preventing spurious interior lakes; updated the enclosed-region test to expect a single connected wet region. ([#522] (https://github.com/CWorthy-ocean/roms-tools/pull/577))
 
 ## v3.5.0
 
