### 2026-03-30 e044c24
Doc-string additions, better test for dask use, consistent passing of dims in choose_domain
