### 2026-03-30 0aaa9ec
Fix enclosed basin labeling and update connectivity test (#577)
* Fix enclosed basin labeling and update connectivity test

* adding release note

* Update releases.md
