### 2026-03-30 6a12d27
most of tests. plot_nesting fix
