+++ roms_tools/tests/test_setup/test_nesting.py
@@ -1,5 +1,6 @@
 import logging
 from pathlib import Path
+from unittest.mock import patch
 
 import numpy as np
 import pytest
@@ -7,6 +8,7 @@ import xarray as xr
 
 from conftest import calculate_file_hash
 from roms_tools import Grid
+from roms_tools.plot import plot_nesting
 from roms_tools.setup.nesting import (
     align_grids,
     compute_boundary_distance,
@@ -373,10 +375,11 @@ class TestModifyChid:
 
 class TestNesting:
     @pytest.mark.parametrize(
-        "fixture_name, include_bgc, include_pressure_fluxes, expected_r_vars, expected_u_vars, expected_v_vars",
+        "fixture_name_child, fixture_name_parent, include_bgc, include_pressure_fluxes, expected_r_vars, expected_u_vars, expected_v_vars",
         [
             (
                 "child_grid_with_bgc",
+                "big_grid",
                 True,
                 False,
                 "zeta, temp, salt, bgc",
@@ -385,6 +388,7 @@ class TestNesting:
             ),
             (
                 "child_grid_that_straddles",
+                "big_grid_that_straddles",
                 False,
                 False,
                 "zeta, temp, salt",
@@ -393,6 +397,7 @@ class TestNesting:
             ),
             (
                 "child_grid_with_pflx",
+                "big_grid",
                 False,
                 True,
                 "zeta, temp, salt",
@@ -404,30 +409,31 @@ class TestNesting:
     def test_successful_initialization(
         self,
         request,
-        fixture_name,
+        fixture_name_child,
+        fixture_name_parent,
         include_bgc,
         include_pressure_fluxes,
         expected_r_vars,
         expected_u_vars,
         expected_v_vars,
     ):
-        child_grid = request.getfixturevalue(fixture_name)
+        child_grid = request.getfixturevalue(fixture_name_child)
+        parent_grid = request.getfixturevalue(fixture_name_parent)
 
+        ds_nesting = make_edata(
+            parent_grid,
+            child_grid,
+            include_bgc=include_bgc,
+            include_pressure_fluxes=include_pressure_fluxes,
+        )
+
+        ### NEED TO ADD A TEST TO COMPARE STDERR CALL FOR BOUNDARIES. correctly assigning open/closed.
         # Basic metadata + structure checks
-        assert child_grid.boundaries == {
-            "south": True,
-            "east": True,
-            "north": True,
-            "west": True,
-        }
-        assert child_grid.metadata["prefix"] == "child"
-        assert child_grid.metadata["period"] == 3600.0
-        assert child_grid.metadata["include_bgc"] is include_bgc
         assert isinstance(child_grid.ds, xr.Dataset)
-        assert isinstance(child_grid.ds_nesting, xr.Dataset)
+        assert isinstance(ds_nesting, xr.Dataset)
 
         # Nesting dataset checks
-        ds = child_grid.ds_nesting
+        ds = ds_nesting
         for direction in ["south", "east", "north", "west"]:
             for location in ["r", "u", "v"]:
                 varname = f"child_{direction}_{location}"
@@ -441,13 +447,7 @@ class TestNesting:
                 elif location == "v":
                     assert ds[varname].attrs["output_vars"] == expected_v_vars
 
-    def test_default_boundaries(self, child_grid_with_eastern_boundary_closed):
-        assert child_grid_with_eastern_boundary_closed.boundaries == {
-            "south": True,
-            "east": False,
-            "north": True,
-            "west": True,
-        }
[…]
