### 2026-04-01 bccd5ab
fixed typo in closing narrow channels logging
