### 2026-04-01 0671c45
Fill narrows (#545)
* adding fill_narrows functionality

* Add `fill_narrow_passages` method to `Grid` class for improved mask handling

* Introduced a new method to fill narrow passages and holes in the ROMS mask.
* Updated velocity masks automatically after modifying the mask.
* Enhanced documentation with detailed parameters and examples for clarity.

* updating the fill narrows test to accomodate the grid halo

* Add logger initialization in mask.py for improved debugging

* Enhance documentation for `fill_narrow_passages` method in `grid.ipynb`

* Added detailed explanations on the functi
[…]
