+++ roms_tools/tests/test_setup/test_grid.py
@@ -1006,3 +1006,54 @@ def test_from_file_partial_parameters_raises_error(grid, tmp_path):
 
     with pytest.raises(ValueError, match="must provide all of"):
         Grid.from_file(path, theta_s=5.0)
+
+
+def test_build_rho_ckdtree_and_nearest_indices(counter_clockwise_rotated_grid):
+    grid = counter_clockwise_rotated_grid
+    lat_rho = grid.ds["lat_rho"].values
+    lon_rho = grid.ds["lon_rho"].values
+    eta_m = lat_rho.shape[0] // 2
+    xi_m = lat_rho.shape[1] // 2
+    lat_q = float(lat_rho[eta_m, xi_m])
+    lon_q = float(lon_rho[eta_m, xi_m])
+
+    grid.build_rho_ckdtree()
+    assert grid._rho_ckdtree is not None
+
+    e, x = grid.nearest_rho_indices(lat_q, lon_q)
+    assert e == eta_m
+    assert x == xi_m
+
+    lats = np.array([lat_q, lat_q])
+    lons = np.array([lon_q, lon_q])
+    ee, xx = grid.nearest_rho_indices(lats, lons)
+    np.testing.assert_array_equal(ee, np.array([eta_m, eta_m]))
+    np.testing.assert_array_equal(xx, np.array([xi_m, xi_m]))
+
+
+def test_rho_index_bounds_from_latlon_bounds(counter_clockwise_rotated_grid):
+    grid = counter_clockwise_rotated_grid
+    lat_rho = grid.ds["lat_rho"].values
+    lon_rho = grid.ds["lon_rho"].values
+    bounds = {
+        "latitude": (float(lat_rho.min()), float(lat_rho.max())),
+        "longitude": (float(lon_rho.min()), float(lon_rho.max())),
+    }
+    bb = grid.rho_index_bounds_from_latlon_bounds(bounds)
+    assert bb["eta_rho"] == (0, lat_rho.shape[0] - 1)
+    assert bb["xi_rho"] == (0, lat_rho.shape[1] - 1)
+
+
+def test_rho_index_bounds_from_latlon_bounds_empty_raises(counter_clockwise_rotated_grid):
+    grid = counter_clockwise_rotated_grid
+    with pytest.raises(ValueError, match="No rho points fall inside"):
+        grid.rho_index_bounds_from_latlon_bounds(
+            {"latitude": (70.0, 80.0), "longitude": (-10.0, 10.0)}
+        )
+
+
+def test_copy_with_ds_clears_rho_ckdtree(counter_clockwise_rotated_grid):
+    grid = counter_clockwise_rotated_grid
+    grid.build_rho_ckdtree()
+    other = grid.copy_with_ds(grid.ds)
+    assert other._rho_ckdtree is None
