+++ roms_tools/setup/grid.py
@@ -4,10 +4,11 @@ import logging
 import re
 from dataclasses import asdict, dataclass, field
 from pathlib import Path
-from typing import Any
+from typing import Any, TypedDict
 
 import numpy as np
 import xarray as xr
+from scipy.spatial import cKDTree
 import yaml
 from matplotlib.axes import Axes
 
@@ -26,9 +27,25 @@ from roms_tools.setup.utils import (
 from roms_tools.utils import (
     interpolate_from_rho_to_u,
     interpolate_from_rho_to_v,
+    normalize_longitude,
     save_datasets,
 )
 from roms_tools.vertical_coordinate import compute_depth_coordinates, sigma_stretch
+
+
+class RhoIndexBounds(TypedDict):
+    """Index ranges along ROMS rho dimensions."""
+
+    eta_rho: tuple[int, int]
+    xi_rho: tuple[int, int]
+
+
+def _lon_array_for_bounds(lon: np.ndarray, straddle: bool) -> np.ndarray:
+    """Put longitudes in the same convention as :func:`normalize_longitude`."""
+    lon = np.asarray(lon, dtype=np.float64) % 360.0
+    if straddle:
+        return np.where(lon > 180.0, lon - 360.0, lon)
+    return lon
 
 
 @dataclass(kw_only=True)
@@ -157,6 +174,8 @@ class Grid:
             hc=self.hc,
             verbose=self.verbose,
         )
+
+        self._clear_rho_ckdtree_cache()
 
     def _input_checks(self):
         if self.topography_source is None:
@@ -813,6 +832,8 @@ class Grid:
 
         grid.mask_shapefile = mask_shapefile
 
+        grid._clear_rho_ckdtree_cache()
+
         return grid
 
     def to_yaml(self, filepath: str | Path) -> None:
@@ -1269,6 +1290,148 @@ class Grid:
 
         return z_centers, z_faces
 
+    def _clear_rho_ckdtree_cache(self) -> None:
+        self._rho_ckdtree = None
+        self._rho_ckdtree_shape: tuple[int, int] | None = None
+
+    def build_rho_ckdtree(self, *, rebuild: bool = False) -> None:
+        """Build and cache a :class:`scipy.spatial.cKDTree` over rho-point geometry.
+
+        Tree points are unit-sphere Cartesian coordinates derived from ``lat_rho`` and
+        ``lon_rho``, so nearest-neighbor queries follow great-circle distance on the
+        sphere. Indices map to ``eta_rho`` (row) and ``xi_rho`` (column) in row-major
+        (C) order.
+
+        Parameters
+        ----------
+        rebuild : bool, optional
+            If True, replace an existing cached tree. Default False.
+        """
+        if getattr(self, "_rho_ckdtree", None) is not None and not rebuild:
+            return
+
+        lat = np.asarray(self.ds["lat_rho"].values, dtype=np.float64)
+        lon = np.asarray(self.ds["lon_rho"].values, dtype=np.float64)
+        if lat.ndim != 2 or lon.shape != lat.shape:
+            raise ValueError("`lat_rho` and `lon_rho` must be 2D arrays of equal shape.")
+        ny, nx = lat.shape
+        lat_r = np.deg2rad(lat.ravel())
+        lon_r = np.deg2rad(lon.ravel())
+        x = np.cos(lat_r) * np.cos(lon_r)
+        y = np.cos(lat_r) * np.sin(lon_r)
+        z = np.sin(lat_r)
+        pts = np.column_stack([x, y, z])
+        self._rho_ckdtree = cKDTree(pts)
+        self._rho_ckdtree_shape = (ny, nx)
+
+    def nearest_rho_indices(
+        self,
[…]
