### 2026-04-02 #10234 @jatkinson1000
Brilliant, thanks for the update all.
It looks like this is progressing smoothly enough so I'll leave you to it.
Please do raise issues for the author as soon as you come across them, and tag `openjournals/joss-reviews#10234` so that they link back here.

@dcherian a weakness of editoralbot is that you have to start a comment with the command, so you'll need to retry with just `@editorialbot generate my checklist`.
