### 2026-04-02 595c98e
Loosened initial chunking on dask datasets to improve dask loading and processing speed
* Option to pass dask chunks to LatLonDataset

* Default dask chunking only in time dimension

* In current datasets, 'reset_chunking' boolean is used to ensure full dimension chunking before operations that require whole dimensions (e.g. regridding) 

* This update should be transparent to users, and potentially do things faster with on-disk data that is chunked or that dask otherwise knows how to slice.
