+++ roms_tools/datasets/roms_dataset.py
@@ -19,6 +19,7 @@ from roms_tools.datasets.utils import (
 )
 from roms_tools.fill import LateralFill
 from roms_tools.utils import (
+    dataset_using_dask,
     get_dask_chunks,
     load_data,
     rotate_velocities,
@@ -246,7 +247,6 @@ class ROMSDataset:
             use_dask=self.use_dask,
             decode_times=False,
             decode_timedelta=False,
-            time_chunking=True,
             force_combine_nested=True,
         )
 
@@ -553,6 +553,7 @@ class ROMSDataset:
         self,
         target_coords: dict[str, Any],
         buffer_points: int = DEFAULT_NR_BUFFER_POINTS,
+        reset_chunking: bool = False,
     ) -> None:
         """Selects a subdomain from the xarray Dataset based on specified target
         coordinates, extending the selection by a defined buffer. Adjusts longitude
@@ -567,6 +568,9 @@ class ROMSDataset:
         buffer_points : int
             The number of grid points to extend beyond the specified latitude and longitude
             ranges when selecting the subdomain. Defaults to 20.
+        reset_chunking : bool
+            Optionally set the dask chunking of the dataset to tell dask that full (non-time)
+            dimensions are required for subsequent operations. Defaults to False.
 
         Returns
         -------
@@ -579,12 +583,22 @@ class ROMSDataset:
             If the selected latitude or longitude range does not intersect with the dataset.
         """
         subdomain = choose_subdomain(
-            self.ds, self.grid.ds, target_coords, buffer_points
+            self.ds,
+            self.grid.ds,
+            target_coords,
+            buffer_points,
+            reset_chunking=reset_chunking,
+            dim_names=self.dim_names,
         )
         self.ds = subdomain
 
         subdomain_grid_ds = choose_subdomain(
-            self.grid.ds, self.grid.ds, target_coords, buffer_points
+            self.grid.ds,
+            self.grid.ds,
+            target_coords,
+            buffer_points,
+            reset_chunking=reset_chunking,
+            dim_names=self.dim_names,
         )
 
         self.grid = self.grid.copy_with_ds(subdomain_grid_ds)
@@ -720,6 +734,8 @@ def choose_subdomain(
     ds_grid: xr.Dataset,
     target_coords: dict[str, Any],
     buffer_points: int = DEFAULT_NR_BUFFER_POINTS,
+    reset_chunking: bool = False,
+    dim_names: dict[str, str] | None = None,
 ):
     """Selects a subdomain from the xarray Dataset based on specified target
     coordinates, extending the selection by a defined buffer. Adjusts longitude
@@ -738,7 +754,12 @@ def choose_subdomain(
     buffer_points : int
         The number of grid points to extend beyond the specified latitude and longitude
         ranges when selecting the subdomain. Defaults to 20.
-
+    reset_chunking : bool
+            Optionally set the dask chunking of the dataset to tell dask that full (non-time)
+            dimensions are required for subsequent operations. Defaults to False.
+    dim_names: dict[str, str] = None
+        Dictionary mapping logical dimension names to dataset dimension names.
+        Only used if reset_chunking is True. Defaults to None.
     Returns
     -------
     xr.Dataset
@@ -831,6 +852,14 @@ def choose_subdomain(
                 "eta_v": slice(first_eta, last_eta),
             }
         )
+
+    # if subsequent operations require this entire chunk, and dask if currently used,
+    # reset the chunking to load the rest of the dataset
+    if reset_chunking and dataset_using_dask(ds):
+        dims = {} if dim_names is None else dim_names  # catch possible None
+        chunks = get_dask_chunks(dims, time_chunking=False)
[…]
