### 2026-04-03 0018c1b
extract method for grid value lookup
