+++ roms_tools/tests/test_setup/test_grid.py
@@ -384,7 +384,10 @@ def test_roundtrip_netcdf(grid_fixture, tmp_path, request):
             filepath = Path(filepath)
 
             # Load the grid from the file
-            grid_from_file = Grid.from_file(filepath.with_suffix(".nc"))
+            grid_from_file = Grid(filename=filepath.with_suffix(".nc"))
+
+            # for this test case, filename won't match; that's ok
+            grid.filename = grid_from_file.filename
 
             assert grid == grid_from_file
             xr.testing.assert_equal(grid.ds, grid_from_file.ds)
@@ -443,7 +446,7 @@ def test_roundtrip_from_file_yaml(grid_fixture, tmp_path, request):
     filepath = Path(tmp_path / "test.nc")
     grid.save(filepath)
 
-    grid_from_file = Grid.from_file(filepath)
+    grid_from_file = Grid(filename=filepath)
 
     filepath_yaml = Path(tmp_path / "test.yaml")
     grid_from_file.to_yaml(filepath_yaml)
@@ -767,7 +770,7 @@ def test_hmin_criterion_and_update_topography():
 
 def test_update_topography_raises_if_grid_loaded_from_file_has_no_source_info():
     fname = download_test_data("grid_created_with_matlab.nc")
-    grid = Grid.from_file(fname)
+    grid = Grid(filename=fname)
 
     with pytest.raises(
         ValueError,
@@ -897,7 +900,7 @@ def test_grid_copy_with_ds_does_not_mutate_original(grid):
 def test_compatability_with_matlab_grid(tmp_path):
     fname = download_test_data("grid_created_with_matlab.nc")
 
-    grid = Grid.from_file(fname)
+    grid = Grid(filename=fname)
 
     assert not grid.straddle
     assert grid.theta_s == 5.0
@@ -939,7 +942,10 @@ def test_compatability_with_matlab_grid(tmp_path):
             filepath = Path(filepath)
 
             # Load the grid from the file
-            grid_from_file = Grid.from_file(filepath.with_suffix(".nc"))
+            grid_from_file = Grid(filename=filepath.with_suffix(".nc"))
+
+            # it won't get the filenames right and that's ok; overwrite one
+            grid.filename = grid_from_file.filename
 
             # Assert that the initial grid and the loaded grid are equivalent (including the 'ds' attribute)
             assert grid == grid_from_file
@@ -961,7 +967,7 @@ def test_from_file_with_vertical_coords(grid, tmp_path):
     path = tmp_path / "grid.nc"
     grid_copy.save(path)
 
-    grid_from_file = Grid.from_file(path, theta_s=theta_s, theta_b=theta_b, hc=hc, N=N)
+    grid_from_file = Grid(filename=path, theta_s=theta_s, theta_b=theta_b, hc=hc, N=N)
     assert np.allclose(grid_from_file.ds.Cs_r, Cs_r)
     assert np.allclose(grid_from_file.ds.Cs_w, Cs_w)
     assert grid_from_file.theta_s == theta_s
@@ -983,21 +989,10 @@ def test_from_file_with_conflicting_vertical_coords(grid, tmp_path):
     grid_copy.save(path)
 
     with pytest.raises(ValueError, match="inconsistent with the provided N"):
-        Grid.from_file(path, theta_s=5.0, theta_b=2.0, hc=300.0, N=100)
+        Grid(filename=path, theta_s=5.0, theta_b=2.0, hc=300.0, N=100)
 
     with pytest.raises(ValueError, match="inconsistent with the provided theta_s, "):
-        Grid.from_file(path, theta_s=5.0, theta_b=2.0, hc=300.0, N=10)
-
-
[…]
