### 2026-04-03 52e00ca
WIP
WIP on external_grid
