### 2026-04-07 8446956
pull out inner functions, use passed hmin
