+++ roms_tools/setup/nesting.py
@@ -22,6 +22,220 @@ from roms_tools.utils import (
 )
 
 
+def align_grids(
+    parent_grid: Grid,
+    child_grid: Grid,
+    boundaries: dict | None = None,
+    hmin: float | None = None,
+    verbose: bool = False,
+) -> "Grid":
+    """This function establishes the relationship between a parent grid and a child grid in ROMS simulations.
+    It alters the child_grid dataset, ``child_grid.ds`` ensuring compatibility with the parent grid by adjusting
+    the child grid mask and topography to match the parent grid at the boundaries.
+
+    To update the child grid mask and ensure consistency with the parent grid, this function performs
+    the following steps:
+
+       1. Adjusts the child mask to ensure consistency with the parent mask.
+
+    To update the child grid topography, this function performs the following operations:
+
+       1. Enforces the minimum depth constraint ``hmin``.
+       2. Adjusts the child grid topography to maintain consistency with the
+          parent grid.
+       3. Updates the internal dataset (``child_grid.ds``) with the processed bathymetry.
+
+    Parameters
+    ----------
+    parent_grid : Grid
+        The parent grid object, providing the reference for the topography and mask of the child grid.
+    child_grid : Grid
+        The child grid object to be adapted and mapped onto the parent_grid.
+    boundaries : Dict[str, bool]
+        Specifies which child grid boundaries ('south', 'east', 'north', 'west') should be adjusted for topography/mask
+        and included in `ds_nesting`. If not provided, valid (non-land) boundaries are enabled automatically.
+    hmin : float, optional
+        Minimum allowable ocean depth (meters). If ``None``, the existing value
+        is retained.
+    verbose: bool, optional
+        Indicates whether to print grid generation steps with timing. Defaults to False.
+
+    Returns
+    -------
+    child_grid : Grid
+        Updates the internal dataset (``child_grid.ds``) in place, modifying the mask
+         and topography, and ensuring consistent parent-child boundary mapping.
+    """
+    boundaries = check_and_set_boundaries(boundaries, child_grid.ds.mask_rho)
+
+    _check_child_outside_parent(parent_grid.ds, child_grid.ds, boundaries)
+
+    # Call Grid update_mask, then modify the child mask:
+    child_grid.ds = _apply_child_modification(
+        parent_grid,
+        child_grid,
+        modifier=lambda p, c: modify_child_mask(p, c, boundaries),  # type: ignore[arg-type]
+        modifier_name="mask",
+        verbose=verbose,
+    )
+
+    # re-coarsen new mask
+    child_grid._coarsen()
+
+    # Call Grid update_topography, then modify the child topography:
+    child_grid.ds = _apply_child_modification(
+        parent_grid,
+        child_grid,
+        modifier=lambda p, c: modify_child_topography(
+            p, c, boundaries, hmin or child_grid.hmin
+        ),
+        modifier_name="topography",
+        verbose=verbose,
+    )
+
+    child_grid.parent_info = parent_grid.grid_to_dict()
+
+    return child_grid
+
+
+def make_edata(
+    parent_grid: Grid,
+    child_grid: Grid,
+    filepath: str | None = None,
+    prefix: str = "child",
+    period: float = 3600.0,
+    include_bgc: bool = False,
+    include_pressure_fluxes: bool = False,
+    boundaries: dict | None = None,
+    verbose: bool = False,
+) -> xr.Dataset:
+    """Maps child grid boundary points onto absolute indices of the parent grid, and saves
+    the nesting information to netCDF4 files.
+
+    This function updates the mapping of child boundaries to parent-grid indices. This
+    mapping depends on the updated mask from ``align_grid``, since masked (land) points may
[…]
