### 2026-04-07 52ce8ee
add test cases
