+++ roms_tools/tests/test_setup/test_nesting.py
@@ -44,8 +44,29 @@ def big_grid_that_straddles():
 
 
 @pytest.fixture()
+def big_grid_that_wraps():
+    return Grid(
+        nx=5, ny=7, center_lon=10, center_lat=61, rot=20, size_x=18000, size_y=2400
+    )
+
+
+@pytest.fixture()
+def big_grid_that_straddles_other_side():
+    return Grid(
+        nx=5, ny=7, center_lon=180, center_lat=61, rot=20, size_x=1800, size_y=2400
+    )
+
+
+@pytest.fixture()
 def small_grid_that_straddles():
     return Grid(nx=10, ny=10, center_lon=0, center_lat=61, rot=0, size_x=50, size_y=200)
+
+
+@pytest.fixture()
+def small_grid_that_straddles_other_side():
+    return Grid(
+        nx=10, ny=10, center_lon=180, center_lat=61, rot=0, size_x=50, size_y=200
+    )
 
 
 @pytest.fixture()
@@ -61,6 +82,91 @@ def child_grid_that_straddles(big_grid_that_straddles):
     )
     child_grid = align_grids(big_grid_that_straddles, child_grid)
     make_edata(big_grid_that_straddles, child_grid, prefix="child")
+
+    return child_grid
+
+
+@pytest.fixture()
+def child_grid_that_straddles_big_wrap_grid(big_grid_that_wraps):
+    child_grid = Grid(
+        nx=10,
+        ny=10,
+        center_lon=0,
+        center_lat=61,
+        rot=0,
+        size_x=50,
+        size_y=200,
+    )
+    child_grid = align_grids(big_grid_that_wraps, child_grid)
+    make_edata(big_grid_that_wraps, child_grid, prefix="child")
+
+    return child_grid
+
+
+@pytest.fixture()
+def child_grid_that_straddles_other_side_big_wrap_grid(big_grid_that_wraps):
+    child_grid = Grid(
+        nx=10,
+        ny=10,
+        center_lon=180,
+        center_lat=61,
+        rot=0,
+        size_x=50,
+        size_y=200,
+    )
+    child_grid = align_grids(big_grid_that_wraps, child_grid)
+    make_edata(big_grid_that_wraps, child_grid, prefix="child")
+
+    return child_grid
+
+
+@pytest.fixture()
+def child_grid_that_straddles2(big_grid_that_straddles_other_side):
+    child_grid = Grid(
+        nx=10,
+        ny=10,
+        center_lon=0,
+        center_lat=61,
+        rot=0,
+        size_x=50,
+        size_y=200,
+    )
+    child_grid = align_grids(big_grid_that_straddles_other_side, child_grid)
+    make_edata(big_grid_that_straddles_other_side, child_grid, prefix="child")
+
+    return child_grid
+
+
+@pytest.fixture()
+def child_grid_that_straddles_other_side(big_grid_that_straddles):
+    child_grid = Grid(
+        nx=10,
+        ny=10,
+        center_lon=180,
+        center_lat=61,
+        rot=0,
+        size_x=50,
+        size_y=200,
+    )
+    child_grid = align_grids(big_grid_that_straddles, child_grid)
+    make_edata(big_grid_that_straddles, child_grid, prefix="child")
+
+    return child_grid
+
+
+@pytest.fixture()
+def both_grids_that_straddles_other_side(big_grid_that_straddles_other_side):
+    child_grid = Grid(
+        nx=10,
+        ny=10,
+        center_lon=180,
+        center_lat=61,
+        rot=0,
+        size_x=50,
+        size_y=200,
+    )
+    child_grid = align_grids(big_grid_that_straddles_other_side, child_grid)
+    make_edata(big_grid_that_straddles_other_side, child_grid, prefix="child")
 
     return child_grid
 
[…]
