+++ conftest.py
@@ -301,7 +301,7 @@ def child_grid_with_bgc(big_grid):
         size_y=500,
     )
     child_grid = align_grids(big_grid, child_grid)
-
+    make_edata(big_grid, child_grid, include_bgc=True)
     return child_grid
 
 
