### 2026-04-08 4ed1465
External grid sje (#580)
* initial edits

* extract method for grid value lookup

* getting tests working

* WIP

WIP on external_grid

* adjust logging

* pull out inner functions, use passed hmin

* clean up some pass-through wrappers

* more superficial restructuring

* revert plot change?

* remove comment

* undo some changes

* Revert "revert plot change?"

This reverts commit 4f505317fb079c07b6f8c470239cdefa581da878.

* fix test plots
