+++ roms_tools/setup/nesting.py
@@ -26,8 +26,6 @@ def align_grids(
     parent_grid: Grid,
     child_grid: Grid,
     boundaries: dict | None = None,
-    mask_shapefile: str | Path | None = None,
-    topography_source: dict | None = None,
     hmin: float | None = None,
     verbose: bool = False,
 ) -> "Grid":
@@ -38,18 +36,14 @@ def align_grids(
     To update the child grid mask and ensure consistency with the parent grid, this function performs
     the following steps:
 
-       1. Derives the child mask from the provided ``mask_shapefile`` (or from the
-          default Natural Earth 10m coastline if ``None``).
-       2. Adjusts the child mask to ensure consistency with the parent mask.
+       1. Adjusts the child mask to ensure consistency with the parent mask.
 
     To update the child grid topography, this function performs the following operations:
 
-       1. Regrids the topography from the specified source.
-       2. Applies domain-wide and local smoothing.
-       3. Enforces the minimum depth constraint ``hmin``.
-       4. Adjusts the child grid topography to maintain consistency with the
+       1. Enforces the minimum depth constraint ``hmin``.
+       2. Adjusts the child grid topography to maintain consistency with the
           parent grid.
-       5. Updates the internal dataset (``child_grid.ds``) with the processed bathymetry.
+       3. Updates the internal dataset (``child_grid.ds``) with the processed bathymetry.
 
     Parameters
     ----------
@@ -60,16 +54,6 @@ def align_grids(
     boundaries : Dict[str, bool]
         Specifies which child grid boundaries ('south', 'east', 'north', 'west') should be adjusted for topography/mask
         and included in `ds_nesting`. If not provided, valid (non-land) boundaries are enabled automatically.
-    mask_shapefile : str or Path, optional
-        Path to a coastal shapefile used to derive the land mask. If ``None``,
-        a default coastline dataset is used.
-    topography_source : dict, optional
-        Dictionary describing the topography data source. Expected keys:
-
-        - ``"name"`` (str): Name of the source dataset (e.g., ``"SRTM15"``).
-        - ``"path"`` (str or Path): Path to the dataset file.
-
-        If ``None``, the previously configured topography source is used.
     hmin : float, optional
         Minimum allowable ocean depth (meters). If ``None``, the existing value
         is retained.
@@ -82,198 +66,31 @@ def align_grids(
         Updates the internal dataset (``child_grid.ds``) in place, modifying the mask
          and topography, and ensuring consistent parent-child boundary mapping.
     """
-
-    def _check_child_outside_parent(
-        parent_grid_ds: xr.Dataset,
-        child_grid_ds: xr.Dataset,
-        boundaries=dict[str, bool],
-    ):
-        """Interpolate the parent indices to the child grid boundary.
-
-        Uses the parent grid ``lon_rho``/``lat_rho`` coordinates to compute
-        fractional i/j indices at child-boundary longitude/latitude points using
-        linear interpolation. The function verifies that all ocean child points
-        (based on ``mask``) lie within the parent grid and gives ValueError if
-        child wet points exist outside of the parent domain.
-
-        Parameters
-        ----------
-        parent_grid_ds : xarray.Dataset
-            Grid information of parent grid.
-        child_grid_ds : xarray.Dataset
-            Grid information of child grid.
-        boundaries : dict[str, bool]
-            Indicator for whether each boundaries is open (True) or closed (False)
-            Used for generating informative warnings or errors.
-        """
-        bdry_coords_dict = get_boundary_coords()
-
[…]
