+++ roms_tools/setup/grid.py
@@ -5,7 +5,7 @@ import os
 import re
 from dataclasses import asdict, dataclass, field
 from pathlib import Path
-from typing import Any
+from typing import Any, TypeVar
 
 import numpy as np
 import xarray as xr
@@ -132,10 +132,18 @@ class Grid:
 
     def __post_init__(self):
         if self.filename is not None:
+            conflicting_params = {
+                "nx",
+                "ny",
+                "size_x",
+                "size_y",
+                "center_lon",
+                "center_lat",
+            }
             # Check to confirm other parameters are not provided with filename
-            min_param = {"rot", "hmin", "verbose", "filename"}
             param_provided = [k for k, v in vars(self).items() if v is not None]
-            extra = set(param_provided) - min_param
+
+            extra = set(param_provided).intersection(conflicting_params)
             if extra:
                 raise ValueError(
                     f"Unexpected parameter included with `filename`: {extra}.\n"
@@ -189,6 +197,8 @@ class Grid:
                 hc=self.hc,
                 verbose=self.verbose,
             )
+
+            self.parent_info = None
 
     def _input_checks(self):
         for var in ("nx", "ny", "size_x", "size_y", "center_lon", "center_lat"):
@@ -443,10 +453,6 @@ class Grid:
         - `angle` -> `angle_coarse`: Angle between the xi axis and true east.
         - `mask_rho` -> `mask_coarse`: Land/sea mask at rho points.
         """
-        # todo check for these keys
-        # NOTE: THERE IS ALREADY A CHECK FOR ALL OR NONE TO DO COARSEN UNDER FROM_FILE
-        # THERE'S NO CHECK IN THE post_init, but presumably, we need the coarsening to happen then.
-        # i.e. this check may be unecessary unless we want individual checks:
         d = {
             "angle": "angle_coarse",
             "mask_rho": "mask_coarse",
@@ -454,12 +460,9 @@ class Grid:
             "lon_rho": "lon_coarse",
         }
 
-        # Only coarsen variables that have not yet been coarsened
-        vars_coarsen = [(k, v) for k, v in d.items() if v not in self.ds]
-
         ds = self.ds
 
-        for fine_var, coarse_var in vars_coarsen:
+        for fine_var, coarse_var in d.items():
             fine_field = ds[fine_var]
             if self.straddle and fine_var == "lon_rho":
                 fine_field = xr.where(fine_field > 180, fine_field - 360, fine_field)
@@ -478,7 +481,7 @@ class Grid:
 
         ds["mask_coarse"] = xr.where(ds["mask_coarse"] > 0.5, 1, 0).astype(np.int32)
 
-        for fine_var, coarse_var in vars_coarsen:
+        for fine_var, coarse_var in d.items():
             long_name = ds[fine_var].attrs.get(
                 "long_name", ds[fine_var].attrs.get("Long_name", "")
             )
@@ -641,7 +644,6 @@ class Grid:
         theta_b: float | None = None,
         hc: float | None = None,
         N: int | None = None,
-        verbose: bool = False,
         **kwargs,
     ) -> "Grid":
         """Create a Grid instance from an existing file.
@@ -775,71 +777,38 @@ class Grid:
         # Manually set the remaining attributes by extracting parameters from dataset
         grid.nx = ds.sizes["xi_rho"] - 2
         grid.ny = ds.sizes["eta_rho"] - 2
-        if "center_lon" in ds.attrs:
-            center_lon = float(ds.attrs["center_lon"])
-        elif "tra_lon" in ds:
-            center_lon = float(extract_single_value(ds["tra_lon"]))
-        elif "title" in ds.attrs:
[…]

[…]
