### 2026-04-08 4b56527
add copy via prepare_grid_datasets
