### 2026-04-10 #10234 @jatkinson1000
Hi @esclapez, @manishvenu and @dcherian

Please could you give us an update on how your reviews are progressing?
I realise many places had holidays last week when you were all hoping to work on this, so please could you prioritise completing a first pass through your checklists and raising any issues for the authors to work on over the course of the next week. It has now been 3 weeks since the review began and we try to aim for a 4 week review process, including author updates.

If there are any particular issues with this please let me know.
