### 2026-04-10 #10234 @dcherian
@jatkinson1000 I have a COI here. I know many of the authors personally and collaborated with Bachman & Long closely until 2.5 years ago. I currently work with the second author (Nicholas).

I believe I can still review impartially, but leave it up to you since you have 2 other reviewers signed up.
