### 2026-04-17 #10234 @jatkinson1000
@dcherian thanks for raising this.
Based on that information I feel that's a little too close to multiple authors for JOSS so will suggest that you are excused from the review in this case. Drop me a 👍 if that's OK and I'll get the bot to remove you.
