### 2026-04-17 #10234 @jatkinson1000
@manishvenu and @esclapez please could you give us an update on when you might finish a first pass of your reviews and pass any feedback to the authors? It has now been 4 weeks since the review began which is the typical turnaround we aim for at JOSS.

If there are any issues that will prevent you doing this in the next week please do let us know.
