### 2026-04-24 CWorthy-ocean/roms-tools#595 @manishvenu
Statement of Need for Documentation
Hey @NoraLoose,

I am reviewing the JOSS submission: https://github.com/openjournals/joss-reviews/issues/10234

[The review criteria](https://joss.readthedocs.io/en/latest/review_criteria.html#a-statement-of-need) seems to want a statement of need in the documentation. I see most of the parts, but there doesn't seem much mention of "relation to other work.".  However, I do see pyroms, regional_mom6, and the matlab scripts mentioned in the paper. Is there somewhere (in the docs) this section is tackled?

Thanks,
Manish V.
