### 2026-04-29 CWorthy-ocean/roms-tools#595 @NoraLoose
Hi @manishvenu, thanks for reviewing the package! Good point on the statement of need for the docs. I am on vacation this week, but will get to this issue next week!
