### 2026-08-18 #10234 @danielskatz
As a non-conflicted AEiC, I'll finish the handling for this submission
