### 2026-08-18 #10234 @danielskatz
👋 @NoraLoose - I've proofread this submission.

Please add a country for the second affiliation in the paper.

Also, I'm a bit confused by the authors for Rutgers ROMS, as I don't see `Arango & contributors` suggested by the repo. I also only see three contributors to the repo. Can you explain why the authors are listed this way? I have the same question for Pyroms and the two UCLA-ROMS and C-Star. I would have expected all of these to have authors more like "Rutgers ROMS contributors", or "Rutgers ROMS development team", similar to the authors listed for Dask.

Additionally, please check the reference list titles, and only use {{}}s when needed to protect casing. JOSS defaults to Sentence case, but doesn't automatically pick up things that should be protected, such as Python in Hoyer.

Please ping me (@danielskatz) when you've made changes, as I'm not assigned to this so might otherwise miss a notification.
