### 2026-08-19 #10234 @NoraLoose
@danielskatz thanks for proofreading the paper!

I've opened [PR #666](https://github.com/CWorthy-ocean/roms-tools/pull/666) to address your comments. Maybe you could take a look to confirm I've interpreted your comments correctly? Thanks so much!
