### 2026-08-19 #10234 @danielskatz
I've added a couple of comments, but this basically looks good. We can merge it then check to make sure it worked ok in the next draft of the article.
