### 2026-08-20 #10234 @editorialbot
:wave: @openjournals/ese-eics, this paper is ready to be accepted and published.
