### 2026-08-20 #10234 @danielskatz
Congratulations to @NoraLoose (Nora Loose) and many co-authors on your JOSS publication!!

And thanks to @manishvenu & @esclapez for reviewing, and to @jatkinson1000 for editing
JOSS is completely dependent on volunteers; we couldn't do this without you
