### 2026-08-20 #10234 @editorialbot
:tada::tada::tada: Congratulations on your paper acceptance! :tada::tada::tada:
